% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct-conversion-qty-unit.R
\name{A2T}
\alias{A2T}
\alias{A2T.default}
\alias{A2T.numeric}
\alias{A2T.filter_spct}
\alias{A2T.filter_mspct}
\title{Convert absorbance into transmittance}
\usage{
A2T(x, action, byref, ...)

\method{A2T}{default}(x, action = NULL, byref = FALSE, ...)

\method{A2T}{numeric}(x, action = NULL, byref = FALSE, ...)

\method{A2T}{filter_spct}(x, action = "add", byref = FALSE, ...)

\method{A2T}{filter_mspct}(x, action = "add", byref = FALSE, ..., .parallel = FALSE, .paropts = NULL)
}
\arguments{
\item{x}{an R object.}

\item{action}{a character string "add" or "replace".}

\item{byref}{logical indicating if new object will be created by reference or
by copy of \code{x}.}

\item{...}{not used in current version.}

\item{.parallel}{if TRUE, apply function in parallel, using parallel backend
provided by foreach}

\item{.paropts}{a list of additional options passed into the foreach function
when parallel computation is enabled. This is important if (for example)
your code relies on external data or packages: use the .export and
.packages arguments to supply them so that all cluster nodes have the
correct environment set up for computing.}
}
\value{
A copy of \code{x} with a column \code{Tfr} added and \code{A} and
  \code{Afr} possibly deleted except for \code{w.length}. If \code{action =
  "replace"}, in all cases, the additional columns are removed, even if no
  column needs to be added.
}
\description{
Function that converts absorbance (a.u.) into transmittance (fraction).
}
\seealso{
Other quantity conversion functions:
\code{\link[=Afr2T]{Afr2T()}},
\code{\link[=T2A]{T2A()}},
\code{\link[=T2Afr]{T2Afr()}},
\code{\link[=any2T]{any2T()}},
\code{\link[=as_quantum]{as_quantum()}},
\code{\link[=e2q]{e2q()}},
\code{\link[=e2qmol_multipliers]{e2qmol_multipliers()}},
\code{\link[=e2quantum_multipliers]{e2quantum_multipliers()}},
\code{\link[=q2e]{q2e()}}
}
\concept{quantity conversion functions}
