% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.absorbance.R
\name{absorbance_spct}
\alias{absorbance_spct}
\title{Calculate absorbance from spectral absorbance.}
\usage{
absorbance_spct(spct, w.band, quantity, wb.trim, use.hinges, naming, ...)
}
\arguments{
\item{spct}{filter_spct}

\item{w.band}{waveband or list of waveband objects or a numeric vector of
length two. The waveband(s) determine the region(s) of the spectrum that
are summarized. If a numeric range is supplied a waveband object is
constructed on the fly from it.}

\item{quantity}{character string One of "average" or "mean", "total",
"contribution", "contribution.pc", "relative" or "relative.pc"}

\item{wb.trim}{logical Flag if wavebands crossing spectral data boundaries
are trimmed or ignored}

\item{use.hinges}{logical Flag indicating whether to use hinges to reduce
interpolation errors}

\item{naming}{character one of "long", "default", "short" or "none". Used to
select the type of names to assign to returned value.}

\item{...}{other arguments (possibly used by derived methods).}
}
\description{
This function returns the mean absorbance for a given
waveband of a absorbance spectrum.
}
\keyword{internal}
