% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.smooth.spct.r
\name{adaptive_smoothing}
\alias{adaptive_smoothing}
\title{Custom smoothing}
\usage{
adaptive_smoothing(
  x,
  y,
  zero.limit.cnst = max(y) * 1e-04,
  smooth.limit = 0.001,
  smooth.threshold = max(y) * 0.1,
  ...
)
}
\arguments{
\item{x, y}{numeric vectors of equal length.}

\item{smooth.limit}{numeric mad/median value above which no smoothing is
applied.}

\item{smooth.threshold}{numeric y value above which no smoothing is applied.}

\item{zero.limit}{numeric vector of length one or of the same leangth as
\code{x} and \code{y}. Smaller values in y are forced to zero.}
}
\description{
Generic implementation of the custom smoothing method
}
\keyword{internal}
