% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{altitude}
\alias{altitude}
\title{function to be numerically minimized}
\usage{
altitude(x, lon, lat, twlght_angl = 0)
}
\value{
an elevation angle delta
}
\description{
function to be numerically minimized
}
\keyword{internal}

