% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.new.r
\name{as.calibration_spct}
\alias{as.calibration_spct}
\alias{as.calibration_spct.default}
\title{Coerce to a spectrum}
\usage{
as.calibration_spct(x, ...)

\method{as.calibration_spct}{default}(x, ...)
}
\arguments{
\item{x}{an R object.}

\item{...}{other arguments passed to "set" functions.}
}
\value{
A copy of \code{x} converted into a \code{calibration_spct} object.
}
\description{
Return a copy of an R object with its class set to a given type of spectrum.
}
\seealso{
\code{\link{setGenericSpct}}

Other constructors of spectral objects:
\code{\link[=as.chroma_spct]{as.chroma_spct()}},
\code{\link[=as.cps_spct]{as.cps_spct()}},
\code{\link[=as.filter_spct]{as.filter_spct()}},
\code{\link[=as.generic_spct]{as.generic_spct()}},
\code{\link[=as.object_spct]{as.object_spct()}},
\code{\link[=as.raw_spct]{as.raw_spct()}},
\code{\link[=as.reflector_spct]{as.reflector_spct()}},
\code{\link[=as.response_spct]{as.response_spct()}},
\code{\link[=as.solute_spct]{as.solute_spct()}},
\code{\link[=as.source_spct]{as.source_spct()}},
\code{\link[=source_spct]{source_spct()}}
}
\concept{constructors of spectral objects}
