% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{as.solar_date}
\alias{as.solar_date}
\title{Convert a solar_time object into solar_date object}
\usage{
as.solar_date(x, time)
}
\arguments{
\item{x}{solar_time object.}

\item{time}{an R date time object}
}
\value{
For method \code{as.solar_date()} a date-time object with the class attr
  set to "solar.time". This is needed only for unambiguous formatting and
  printing.
}
\description{
Convert a solar_time object into solar_date object
}
