% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.hinges.r
\name{auto_hinges}
\alias{auto_hinges}
\title{Guess whether insertion of hinges is needed or not}
\usage{
auto_hinges(w.length, step.limit = 0.25)
}
\arguments{
\item{w.length}{numeric vector of wavelengths (nm).}

\item{step.limit}{numeric value (nm) for step, so that larger wavelength
steps trigger insertion of hinges before computations on spectra.}
}
\value{
A logical value, TRUE if insertion of hinges is deemed necessary.
}
\description{
Assess from wavelength vector whether insertion of hinges before further
calculations is needed.
}
\keyword{internal}
