% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-replace-combine.r
\name{c}
\alias{c}
\alias{c.generic_mspct}
\title{Combine collections of spectra}
\usage{
\method{c}{generic_mspct}(..., recursive = FALSE, ncol = 1, byrow = FALSE)
}
\arguments{
\item{...}{one or more generic_mspct objects to combine.}

\item{recursive}{logical ignored as nesting of collections of spectra is
not supported.}

\item{ncol}{numeric Virtual number of columns}

\item{byrow}{logical When object has two dimensions, how to map member
objects to columns and rows.}
}
\value{
A collection of spectra object belonging to the most derived class
shared among the combined objects.
}
\description{
Combine two or more generic_mspct objects into a single object.
}
