% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.multipliers.r
\name{calc_multipliers}
\alias{calc_multipliers}
\title{Spectral weights}
\usage{
calc_multipliers(
  w.length,
  w.band,
  unit.out = "energy",
  unit.in = "energy",
  use.cached.mult = FALSE,
  fill = 0
)
}
\arguments{
\item{w.length}{numeric vector of wavelengths (nm).}

\item{w.band}{waveband object.}

\item{unit.out}{character One of "photon" or "energy", default is "energy".}

\item{unit.in}{character One of "photon" or "energy", default is "energy".}

\item{use.cached.mult}{logical Flag indicating whether multiplier values
should be cached between calls.}

\item{fill}{numeric If \code{fill = NA} then values returned for wavelengths
outside the range of the waveband are set to \code{NA}.}
}
\value{
a numeric vector of multipliers of the same length as
  \code{w.length}.
}
\description{
Calculate multipliers for selecting a range of wavelengths and optionally
applying a biological spectral weighting function (BSWF) and wavelength
normalization. This function returns numeric multipliers that can be used to
select a waveband and apply a weight.
}
\examples{
with(sun.data,
     calc_multipliers(w.length = w.length,
                      w.band = new_waveband(400,700),
                      unit.out = "photon"))
with(sun.data,
     calc_multipliers(w.length = w.length,
                      w.band = new_waveband(400,700),
                      unit.out = "photon"),
                      use.cached.mult = TRUE)

}
\seealso{
Other low-level functions operating on numeric vectors.:
\code{\link[=as_energy]{as_energy()}},
\code{\link[=as_quantum_mol]{as_quantum_mol()}},
\code{\link[=div_spectra]{div_spectra()}},
\code{\link[=energy_irradiance]{energy_irradiance()}},
\code{\link[=energy_ratio]{energy_ratio()}},
\code{\link[=insert_hinges]{insert_hinges()}},
\code{\link[=integrate_xy]{integrate_xy()}},
\code{\link[=interpolate_spectrum]{interpolate_spectrum()}},
\code{\link[=irradiance]{irradiance()}},
\code{\link[=l_insert_hinges]{l_insert_hinges()}},
\code{\link[=oper_spectra]{oper_spectra()}},
\code{\link[=photon_irradiance]{photon_irradiance()}},
\code{\link[=photon_ratio]{photon_ratio()}},
\code{\link[=photons_energy_ratio]{photons_energy_ratio()}},
\code{\link[=prod_spectra]{prod_spectra()}},
\code{\link[=s_e_irrad2rgb]{s_e_irrad2rgb()}},
\code{\link[=split_energy_irradiance]{split_energy_irradiance()}},
\code{\link[=split_photon_irradiance]{split_photon_irradiance()}},
\code{\link[=subt_spectra]{subt_spectra()}},
\code{\link[=sum_spectra]{sum_spectra()}},
\code{\link[=trim_tails]{trim_tails()}},
\code{\link[=v_insert_hinges]{v_insert_hinges()}},
\code{\link[=v_replace_hinges]{v_replace_hinges()}}
}
\concept{low-level functions operating on numeric vectors.}
