% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{check_and_rename_vars}
\alias{check_and_rename_vars}
\title{check and rename variables}
\usage{
check_and_rename_vars(
  x,
  target.var,
  alternative.vars,
  multiplier,
  required = FALSE,
  fill = NULL
)
}
\arguments{
\item{x}{data frame or equivalent R object.}

\item{target.var}{character vector of length one, with the name of the
target variable.}

\item{alternative.vars}{character vector of any length, with the names of the
of alternative variables that could replace the target.}

\item{multiplier}{numeric vector with the multipliers to apply to the
alternative variables.}

\item{required}{logical Indicating whether an error should be triggered if
no variable is found.}

\item{fill}{if \code{required = TRUE} and \code{fill} different from
\code{NULL} the value is used to fill the target variable if it and
all alternative variables are missing.}
}
\description{
check and rename variables
}
\keyword{internal}
