% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.spectrum.r
\name{check_w.length}
\alias{check_w.length}
\title{Sanity check of wavelengths (internal function).}
\usage{
check_w.length(w.length)
}
\arguments{
\item{w.length}{numeric array of wavelength (nm)}
}
\value{
a single logical value indicating whether test was passed or not
}
\description{
This function checks a w.length vector for compliance with assumptions
expected for valid calculations.
}
\examples{

with(sun.data, photobiology:::check_w.length(w.length))

}
\seealso{
Other data validity check functions:
\code{\link[=check_spct]{check_spct()}},
\code{\link[=check_spectrum]{check_spectrum()}},
\code{\link[=enable_check_spct]{enable_check_spct()}}
}
\concept{data validity check functions}
