% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.peaks.r
\name{check_wl_stepsize}
\alias{check_wl_stepsize}
\title{Check wavelength stepsize}
\usage{
check_wl_stepsize(x, span)
}
\arguments{
\item{x}{numeric vector. Hint: to find valleys, change the sign of the
argument with the unary operator \code{-}.}

\item{span}{odd integer A peak is defined as an element in a sequence which
is greater than all other elements within a moving window of width
\code{span} centred at that element. The default value is 5, meaning that a
peak is taller than its four nearest neighbours. \code{span = NULL} extends
the span to the whole length of \code{x}.}
}
\value{
logical \code{TRUE} if check is passed and otherwise \code{FALSE}
with a warning.
}
\description{
Check wavelength stepsize
}
\details{
As the search for peaks uses a window based on a fixed number of
  observations at neighbouring wavelengths, if the wavelength step between
  observations varies drastically, the window expressed in nanometres of
  wavelength becomes very irregular. With the default \code{span = 5} the
  search for peaks in most cases still works.

  The typical case are spectra returned by \code{thin_wl()}, which retain
  the original local maxima, and a reasonably narrow wavelength maximum
  step size when using default arguments.
}
\keyword{internal}
