% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.methods.R
\name{checkMspctVersion}
\alias{checkMspctVersion}
\title{Check that the "mspct.version" attribute is set}
\usage{
checkMspctVersion(x)
}
\arguments{
\item{x}{a generic_mspct object}
}
\value{
numeric value
}
\description{
Function to check the "mspct.version" attribute of an existing generic_spct
object.
}
\note{
if x is not a \code{generic_mspct} object, \code{NA} is returned,
  and if it the attribute is missing, zero is returned with a warning.
}
\keyword{internal}
