% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{checkSpctVersion}
\alias{checkSpctVersion}
\title{Check that the "spct.version" attribute is set}
\usage{
checkSpctVersion(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
numeric value
}
\description{
Function to check the "spct.version" attribute of an existing generic_spct
object.
}
\note{
if x is not a \code{generic_spct} object, \code{NA} is returned,
  and if it the attribute is missing, zero is returned with a warning.
}
\keyword{internal}
