% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{checkTimeUnit}
\alias{checkTimeUnit}
\title{Check the "time.unit" attribute of an existing source_spct object}
\usage{
checkTimeUnit(x)
}
\arguments{
\item{x}{a source_spct object}
}
\value{
x possibly with the \code{time.unit} attribute modified
}
\description{
Function to read the "time.unit" attribute
}
\note{
if x is not a \code{source_spct} or a \code{response_spct} object, NA
is returned
}
\seealso{
Other time attribute functions:
\code{\link[=convertThickness]{convertThickness()}},
\code{\link[=convertTimeUnit]{convertTimeUnit()}},
\code{\link[=getTimeUnit]{getTimeUnit()}},
\code{\link[=setTimeUnit]{setTimeUnit()}}
}
\concept{time attribute functions}
