% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciexyz.data.r
\docType{data}
\name{ciev10.spct}
\alias{ciev10.spct}
\title{Linear energy CIE 2008 luminous efficiency function 10 deg data}
\format{
A chroma_spct object with 441 rows and 4 variables
}
\usage{
ciev10.spct
}
\description{
A dataset containing wavelengths at a 1 nm interval (390 nm to
  830 nm) and the corresponding response values for a 10 degrees target.
  Original data from \url{http://www.cvrl.org/} downloaded on 2014-04-29 The
  variables are as follows:

\itemize{
  \item w.length (nm)
  \item s.e.response}
}
\examples{
ciev10.spct

}
\seealso{
Other Visual response data examples:
\code{\link{beesxyzCMF.spct}},
\code{\link{ciev2.spct}},
\code{\link{ciexyzCC10.spct}},
\code{\link{ciexyzCC2.spct}},
\code{\link{ciexyzCMF10.spct}},
\code{\link{ciexyzCMF2.spct}},
\code{\link{cone_fundamentals10.spct}}
}
\author{
CIE
}
\concept{Visual response data examples}
\keyword{datasets}
