% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{class_spct}
\alias{class_spct}
\title{Query which is the class of a spectrum}
\usage{
class_spct(x)
}
\arguments{
\item{x}{any R object}
}
\value{
class_spct returns a vector containing all matching xxxx.spct
  classes.
}
\description{
Functions to check if an object is a generic spectrum, or coerce it if
possible.
}
\examples{
class_spct(sun.spct)
class(sun.spct)

}
