% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.clean.r
\name{clean_spct}
\alias{clean_spct}
\title{Clean a spectrum}
\usage{
clean_spct(x, range, range.s.data, fill, col.names, col.pattern = NULL, ...)
}
\arguments{
\item{x}{an R object}

\item{range}{numeric vector of wavelengths}

\item{range.s.data}{numeric vector of length two giving the allowable
range for the spectral data.}

\item{fill}{numeric vector of length 1 or 2, giving the replacement
values to use at each extreme of the range.}

\item{col.names}{character The name of the variable to clean.}

\item{...}{currently ignored}
}
\description{
These functions implement the equivalent of replace() but for spectral
objects instead of vectors.
}
\keyword{internal}
