% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.multipliers.r
\name{clear_photobio.cache}
\alias{clear_photobio.cache}
\title{Clear the spectral weights cache}
\usage{
clear_photobio.cache(pattern = "*")
}
\arguments{
\item{pattern}{character string passed to ls() for selecting within the
environment .photobio.cache the objects to be deleted}
}
\description{
Clear the cache objects stored in environment .photobio.cache
}
\keyword{internal}
