% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.methods.R, R/spct.summaries.r
\name{color.source_mspct}
\alias{color.source_mspct}
\alias{color.source_spct}
\title{Color}
\usage{
\method{color}{source_mspct}(x, ..., idx = !is.null(names(x)))

\method{color}{source_spct}(x, type = "CMF", ...)
}
\arguments{
\item{x}{an object of class "source_spct"}

\item{...}{not used in current version}

\item{idx}{logical whether to add a column with the names of the elements of
spct}

\item{type}{character telling whether "CMF", "CC", or "both" should be returned.}
}
\description{
A function that returns the equivalent RGB colour of an object of class
"source_spct".
}
\examples{
color(sun.spct)

}

