% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.methods.R
\name{convolve_each}
\alias{convolve_each}
\title{Convolve function for collections of spectra}
\usage{
convolve_each(e1, e2, oper = `*`, sep = "_", ...)
}
\arguments{
\item{e1}{an object of class \code{generic_mspct} or \code{generic_scpt} or
\code{numeric}}

\item{e2}{an object of class \code{generic_mspct} or \code{generic_scpt} or
\code{numeric}}

\item{oper}{function, usually but not necessarily an operator with two
arguments.}

\item{sep}{character Used when pasting the names of members of \code{e1} and
\code{e2} to form the names of members of the returned collection of
spectra.}

\item{...}{additional arguments passed to \code{oper} if present.}
}
\description{
Convolve function for collections of spectra which applies an operation on
all the individual members of the collection(s) of spectra.
}
\note{
At least one of e1 and e2 must be a \code{generic_mspct} object or
  derived.
}
\seealso{
Other math operators and functions:
\code{\link{MathFun}},
\code{\link[=^.generic_spct]{^.generic_spct()}},
\code{\link{div-.generic_spct}},
\code{\link{log}},
\code{\link{minus-.generic_spct}},
\code{\link{mod-.generic_spct}},
\code{\link{plus-.generic_spct}},
\code{\link{round}},
\code{\link{sign}},
\code{\link{slash-.generic_spct}},
\code{\link{times-.generic_spct}}
}
\concept{math operators and functions}
