% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2.fel.data.r
\docType{data}
\name{D2.UV654}
\alias{D2.UV654}
\title{Data for typical calibration lamps}
\format{
A \code{polynom::polynomial} object with 6 constants.
}
\description{
A dataset containing fitted constants to be used as input for
  function \code{D2_spectrum}.
}
\details{
An object of class \code{polynom::polynomial}.
}
\author{
Lasse Ylianttila (data)
}
\keyword{datasets}
