% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{date2seconds}
\alias{date2seconds}
\title{date argument check and conversion}
\usage{
date2seconds(t, tz)
}
\value{
numeric representtaion of the date
}
\description{
date argument check and conversion
}
\keyword{internal}

