% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{date2tod}
\alias{date2tod}
\title{Convert date to time-of-day in hours}
\usage{
date2tod(date, unit.out)
}
\arguments{
\item{date}{a date object accepted by lubridate functions}
}
\description{
Convert date to time-of-day in hours
}
\keyword{internal}

