% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.normalize.r
\name{denormalize_spct}
\alias{denormalize_spct}
\title{Undo previously applied normalization}
\usage{
denormalize_spct(spct, wipe.away = FALSE)
}
\arguments{
\item{spct}{A \code{generic_spct} object or a \code{generic_mspct} object,
or objects of derived classes.}

\item{wipe.away}{logical If \code{TRUE} the normalization metadata is removed
after undoing the effect of the normalization.}
}
\value{
A modified copy of \code{spct} if it was previously normalized or
\code{spct} unchanged, otherwise.
}
\description{
Undo previously applied normalization
}
\keyword{internal}
