% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.spikes.R
\name{despike}
\alias{despike}
\alias{despike.default}
\alias{despike.numeric}
\alias{despike.data.frame}
\alias{despike.generic_spct}
\alias{despike.source_spct}
\alias{despike.response_spct}
\alias{despike.filter_spct}
\alias{despike.reflector_spct}
\alias{despike.solute_spct}
\alias{despike.cps_spct}
\alias{despike.raw_spct}
\alias{despike.generic_mspct}
\alias{despike.source_mspct}
\alias{despike.response_mspct}
\alias{despike.filter_mspct}
\alias{despike.reflector_mspct}
\alias{despike.solute_mspct}
\alias{despike.cps_mspct}
\alias{despike.raw_mspct}
\title{Remove spikes from spectrum}
\usage{
despike(
  x,
  height.threshold,
  z.threshold,
  k,
  spike.direction,
  window.width,
  method,
  na.rm,
  max.spike.width,
  ...
)

\method{despike}{default}(
  x,
  height.threshold,
  z.threshold = NA,
  k = NA,
  spike.direction = NA,
  window.width = NA,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...
)

\method{despike}{numeric}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...
)

\method{despike}{data.frame}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...,
  y.var.name = NULL,
  var.name = y.var.name
)

\method{despike}{generic_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...,
  y.var.name = NULL,
  var.name = y.var.name
)

\method{despike}{source_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  ...
)

\method{despike}{response_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  ...
)

\method{despike}{filter_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  filter.qty = getOption("photobiology.filter.qty", default = "transmittance"),
  ...
)

\method{despike}{reflector_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...
)

\method{despike}{solute_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...
)

\method{despike}{cps_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...
)

\method{despike}{raw_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...
)

\method{despike}{generic_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...,
  y.var.name = NULL,
  var.name = y.var.name,
  .parallel = FALSE,
  .paropts = NULL
)

\method{despike}{source_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  ...,
  .parallel = FALSE,
  .paropts = NULL
)

\method{despike}{response_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  ...,
  .parallel = FALSE,
  .paropts = NULL
)

\method{despike}{filter_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  filter.qty = getOption("photobiology.filter.qty", default = "transmittance"),
  ...,
  .parallel = FALSE,
  .paropts = NULL
)

\method{despike}{reflector_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...,
  .parallel = FALSE,
  .paropts = NULL
)

\method{despike}{solute_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...,
  .parallel = FALSE,
  .paropts = NULL
)

\method{despike}{cps_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...,
  .parallel = FALSE,
  .paropts = NULL
)

\method{despike}{raw_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  window.width = 11,
  method = "run.mean",
  na.rm = FALSE,
  max.spike.width = NA,
  ...,
  .parallel = FALSE,
  .paropts = NULL
)
}
\arguments{
\item{x}{numeric vector containing the data.}

\item{height.threshold}{numeric The minimum height of spikes expressed
relative to the median amplitude of the baseline local variation of
\code{x}.}

\item{z.threshold}{numeric Modified local \eqn{Z} values larger than
\code{z.threshold} are detected as boundaries of spikes.}

\item{k}{integer width of median window used for smoothing; must be odd}

\item{spike.direction}{character Controls the direction of spikes to be
detected. Accepted arguments are \code{"up"}, \code{"down"},
\code{"both"}.}

\item{window.width}{integer. The full width of the window used for the
running mean.}

\item{method}{character The name of the method: \code{"run.mean"} is running
mean as described in Whitaker and Hayes (2018); \code{"adj.mean"} is mean
of adjacent neighbors (isolated bad pixels only).}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped
before searching for spikes.}

\item{max.spike.width}{integer The width of the widest spike to be detected,
\code{NA} puts no limit.}

\item{...}{passed in recursive calls.}

\item{var.name, y.var.name}{character Names of columns where to look
for spikes to remove.}

\item{unit.out}{character One of "energy" or "photon"}

\item{filter.qty}{character One of "transmittance" or "absorbance"}

\item{.parallel}{if TRUE, apply function in parallel, using parallel backend
provided by foreach}

\item{.paropts}{a list of additional options passed into the foreach function
when parallel computation is enabled. This is important if (for example)
your code relies on external data or packages: use the .export and
.packages arguments to supply them so that all cluster nodes have the
correct environment set up for computing.}
}
\value{
A copy of the object passed as argument to \code{x} with values
  detected as spikes replaced by a local average of neighbours
  outside the spike.
}
\description{
Function that returns an R object with observations corresponding to spikes
replaced by values computed from neighboring pixels. Spikes are values in
spectra that are unusually high compared to neighbors. They are usually
individual values or very short runs of similar "unusual" values.
}
\section{Spike detection}{

  Spikes are detected based on a modified \eqn{Z} score calculated
  from the differenced spectrum. The \eqn{Z} threshold used should be
  adjusted to the characteristics of the input and desired sensitivity. The
  lower the threshold the more stringent the test becomes, with shorter
  spikes being detected.

  \strong{The algorithms assume a consistent step size for the underlying
  independent variable, e.g., wavelength or time, and should not be applied
  if the data do not fulfil this assumption, at least approximately. As
  \code{find_spkikes()} operates on a single vector, checking this remains
  the responsibility of calling functions or methods such as
  \code{\link{spikes}()} and \code{\link{despike}()}.}

  The algorithm uses running differences to detect abrupt changes in value,
  compared to an estimate of the baseline variation of the differences,
  approximating a baseline \eqn{Z} from MAD and a baseline value from the
  median differences. Currently, a single estimate of MAD is used but running
  medians, when possible, as baseline. This comparison detects running
  differences that are unusually large, in most cases signalling a transition
  between values near the baseline and far from it, in both directions.

  Transitions into- and out of spikes are distinguished based on the median
  of the non-differenced values, as a descriptor of the data baseline. As for
  the median of the differences, a running median is used when possible.

  This function thus detects the start and end of each spike, and
  distinguishes upward and downward spikes.

  \code{k} is the width in number of observations of the window used for
  running median smoothing to extract the baseline. A value several times the
  width of the broader spike but narrow enough to track broader peaks needs
  to be manually set in most cases.

  With \code{na.rm = TRUE}, \code{NA} values are omitted before searching for
  spikes and set to \code{0L} in the returned vector.

  If all spikes are guaranteed to be one observation-wide and either going up
  or down from the baseline, it is possible to detect them based purely on
  the \code{z.threshold} by passing \code{height.threshold = NA} and either
  \code{spike.direction = "up"} or \code{spike.direction = "down"}, which
  ensures very fast computation.

  Parameters of the algorithm need to be adjusted depending on the data, so
  inspection of returned values is needed together with adjustment by trial
  and error of suitable values for \code{z.threshold},
  \code{height.threshold}, and \code{k}.

  Parameter \code{max.spike.width} searches for too wide spikes in the
  output of the algorithms described above and ignores them. This is
  possibly redundant, but maintained for partial backwards compatibility.
}

\section{Replacement values}{

Simple interpolation enabled by \code{method = "adj.mean"} replaces values of
isolated bad pixels by the mean of their two closest neighbours. The running
mean approach enabled by \code{method = "run.mean"} allows the replacement of
short runs of bad pixels by the running mean of neighboring pixels within a
window of user-specified width. The first approach works well for spectra
from array spectrometers to correct for hot and dead pixels in an instrument.
The second approach is most suitable for Raman spectra in which spikes
triggered by radiation are wider than a single pixel but usually not more
than five pixels wide.

Simple interpolation can replace spikes at any position in \code{x}, using
a single neighbour as replacement at the extremes of \code{x} instead of the
mean of two neighbours. The
running mean approach does not replace those pixels whose distance to the
first or last member of \code{x} is less than half the window used for the
running mean, issuing a warning.

When \code{na.rm = TRUE}, \code{NA} values are considered "bad pixels" and
replaced as such rather than discarded with no replacement. This is the
default behaviour.
}

\examples{

white_led.raw_spct[120:125, ]

# find and replace spike at 245.93 nm
despike(white_led.raw_spct,
        z.threshold = 5,
        window.width = 7)[120:125, ]

# A high z.threshold value detects more extreme spikes
despike(white_led.raw_spct,
        z.threshold = 50,
        window.width = 7)[120:125, ]

}
\seealso{
See \code{\link{find_spikes}()} for locating spikes in a vector,
  \code{\link{spikes}()} for extracting/detecting spikes in spectra and
  and \code{\link{replace_bad_pixs}()} for replacing by interpolation
  missing or bad values in a vector.
}
\concept{despike and valleys functions}
