% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmspct.classes.R
\name{dim.generic_mspct}
\alias{dim.generic_mspct}
\alias{dim<-.generic_mspct}
\title{Dimensions of an Object}
\usage{
\method{dim}{generic_mspct}(x)

\method{dim}{generic_mspct}(x) <- value
}
\arguments{
\item{x}{A \code{generic_mspct} object or of a derived class.}

\item{value}{Either NULL or a numeric vector, which is coerced to integer (by truncation).}
}
\value{
Either NULL or a numeric vector, which is coerced to integer (by
  truncation).
}
\description{
Retrieve or set the dimension of an object.
}
