% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e2quantum.multipliers.r
\name{e2quantum_multipliers}
\alias{e2quantum_multipliers}
\title{Calculate energy to quantum multipliers}
\usage{
e2quantum_multipliers(w.length, molar = FALSE)
}
\arguments{
\item{w.length}{numeric Vector of wavelengths (nm)}

\item{molar}{logical Flag indicating whether output should be in moles or
numbers}
}
\value{
A numeric vector of multipliers
}
\description{
Gives multipliers as a function of wavelength, for converting from energy to
photon (quantum) units (number of photons as default, or moles of photons).
}
\examples{
with(sun.data, e2quantum_multipliers(w.length))
with(sun.data, e2quantum_multipliers(w.length, molar = TRUE))

}
\seealso{
Other quantity conversion functions:
\code{\link[=A2T]{A2T()}},
\code{\link[=Afr2T]{Afr2T()}},
\code{\link[=T2A]{T2A()}},
\code{\link[=T2Afr]{T2Afr()}},
\code{\link[=any2T]{any2T()}},
\code{\link[=as_quantum]{as_quantum()}},
\code{\link[=e2q]{e2q()}},
\code{\link[=e2qmol_multipliers]{e2qmol_multipliers()}},
\code{\link[=q2e]{q2e()}}
}
\concept{quantity conversion functions}
