% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.operators.r
\name{energy_as_default}
\alias{energy_as_default}
\alias{photon_as_default}
\alias{quantum_as_default}
\alias{Tfr_as_default}
\alias{Afr_as_default}
\alias{A_as_default}
\alias{unset_radiation_unit_default}
\alias{unset_filter_qty_default}
\alias{unset_user_defaults}
\title{Set spectral-data options}
\usage{
energy_as_default()

photon_as_default()

quantum_as_default()

Tfr_as_default()

Afr_as_default()

A_as_default()

unset_radiation_unit_default()

unset_filter_qty_default()

unset_user_defaults()
}
\value{
Previous value of the modified option.
}
\description{
Set spectral-data related options easily.
}
