% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.et.r
\name{ET_PM}
\alias{ET_PM}
\title{Potential evapotranspiration}
\usage{
ET_PM(
  temperature,
  water.vp,
  wind.speed,
  net.irradiance,
  lw.emissivity = 0.98,
  Rs = 0,
  atmospheric.pressure = 10.13,
  method = "FMI.PM",
  check.range = TRUE
)
}
\arguments{
\item{temperature}{numeric vector of air temperatures (C) at 2 m height.}

\item{water.vp}{numeric vector of water vapour pressure in air (Pa).}

\item{wind.speed}{numeric Wind speed (m/s) at 2 m height.}

\item{net.irradiance}{numeric Net radiation balance (W/m2).}

\item{lw.emissivity}{numeric Emissivity for long wave radiation (/1)}

\item{Rs}{numeric Surface resistance (?).}

\item{atmospheric.pressure}{numeric Atmospheric pressure (Pa).}

\item{method}{character The name of an estimation method.}

\item{check.range}{logical Flag indicating whether to check or not that
arguments for temperature are within range of method. Passed to
function calls to \code{water_vp_sat()} and \code{water_vp_sat_slope()}.}
}
\value{
A numeric vector of reference evapotranspiration estimates expressed
  as mm / h.
}
\description{
Formulation of the original Penman-Monteith equation used in the computation
of the forest fire index (FFI) by the Finnish Meteorological Institute.
Compared to the simplified formulation used by FAO this formualtion requires
downwelling long wave radiation as input. The present version is based on the
FORTRAN code by Ari Venäläinen from 1996. However, we compute saturated water
vapour pressure and its slope vs. temperature using Tetens' formulation as
implemented in functions \code{water_vp_sat()} and
\code{water_vp_sat_slope()}.
}
\examples{
# instantaneous
ET_PM(temperature = 20,
      water.vp = water_RH2vp(0.7, 20),
      wind.speed = 0,
      net.irradiance = 20)

}
\seealso{
Other Evapotranspiration and energy balance related functions.: 
\code{\link{ET_ref}()},
\code{\link{net_irradiance}()}
}
\concept{Evapotranspiration and energy balance related functions.}
