% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evapotranspiration.r
\name{ET_ref}
\alias{ET_ref}
\alias{net_lw_radiation}
\title{Evapotranspiration}
\usage{
ET_ref(
  temperature,
  water.vp,
  wind.speed,
  radiation,
  nighttime = FALSE,
  atmospheric.pressure = 10.13,
  soil.heat.flux = 0,
  method = "FAO.PM",
  check.range = TRUE
)

net_lw_radiation(temperature, water.vp)
}
\arguments{
\item{temperature}{numeric vector of air temperatures (C) at 2 m height.}

\item{water.vp}{numeric vector of water vapour pressure in air (Pa).}

\item{wind.speed}{numeric Wind speed (m/s) at 2 m height.}

\item{radiation}{numeric Global radiation (W/m2).}

\item{nighttime}{logical Used only for methods that distinguish between
daytime- and nighttime canopy conductances.}

\item{atmospheric.pressure}{numeric Atmospheric pressure (Pa).}

\item{soil.heat.flux}{numeric Soil heat flux (W/m2), positive if soil
temperature is increasing.}

\item{method}{character The name of an estimation method.}

\item{check.range}{logical Flag indicating whether to check or not that
arguments for temperature are within range of method. Passed to
function calls to \code{water_vp_sat()} and \code{water_vp_sat_slope()}.}
}
\value{
A numeric vector of $ET_{ref}$ estimates expressed as mm/h.

A numeric vector of $R_{nl}$ estimates expressed as W / m-2.
}
\description{
Compute an estimate of actual or potential evapotranspiration from
meteorologial data. Evapotranspiration from vegetation includes
transpiraction by plants plus evaporation from the soil or other wet
surfaces. $ET_{ref}$ is the reference value assuming no limitation to
transpiration due to soil water, similar to potential evapotranspiration
(PET). An actual value $ET_{ref}$ can be estimated only if additional
information on the plants and soil is available.
}
\details{
Currently only one method, based on the Penman-Monteith equation
  formulated as recommended by FAO (Allen et al., 1998), is implemented. It
  relies on data meassured according WHO standards at 2 m above ground level
  to estimate reference evapotranspiration ($ET_{ref}$).
  Corrections are included to account for the assumption that the vegetation
  has a height of 0.12 m. These input data can be averages over different
  lengths of time. Allen et al. (1998) describes different approaches to
  estimate input data when not available. The auxiliary function
  \code{net_lw_radiation()} can be used to estimate the net lw radiation
  uncorrected for cloudiness.
}
\examples{
# instantaneous
ET_ref(temperature = 20,
       water.vp = water_RH2vp(0.7, 20),
       wind.speed = 0,
       radiation = 10)

ET_ref(temperature = 35,
       water.vp = water_RH2vp(0.1, 35),
       wind.speed = 5,
       radiation = 400)

ET_ref(temperature = 35,
       water.vp = water_RH2vp(0.1, 35),
       wind.speed = 5,
       radiation = 400,
       method = "ASCE.PM.short")

ET_ref(temperature = 5,
       water.vp = water_RH2vp(0.95, 5),
       wind.speed = 0.5,
       radiation = -10,
       nighttime = TRUE,
       method = "ASCE.PM.short")

ET_ref(temperature = 35,
       water.vp = water_RH2vp(0.10, 35),
       wind.speed = 5,
       radiation = 400,
       method = "ASCE.PM.tall")

}
\references{
Allen R G, Pereira L S, Raes D, Smith M. 1998. Crop evapotranspiration:
  Guidelines for computing crop water requirements. Rome: FAO.
}
