% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.operators.r
\name{f_dispatcher_spct}
\alias{f_dispatcher_spct}
\title{Math function dispatcher for spectra}
\usage{
f_dispatcher_spct(x, .fun, ...)
}
\arguments{
\item{x}{an object of class "generic_spct"}

\item{.fun}{an R function with signature function(x, ...)}

\item{...}{additional arguments passed to f}
}
\description{
Function that dispatches the function supplied as argument using different variables depending
on the class of the spectrum argument.
}
\keyword{internal}
