% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2.fel.data.r
\docType{data}
\name{FEL.BN.9101.165}
\alias{FEL.BN.9101.165}
\title{Data for typical calibration lamps}
\format{
A \code{numeric vector}.
}
\description{
A dataset containing fitted constants to be used as input for
  function \code{FEL_spectrum}.
}
\author{
Lasse Ylianttila (data)
}
\keyword{datasets}
