% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.data.r
\docType{data}
\name{filter_cps.mspct}
\alias{filter_cps.mspct}
\title{Counts per second from a measurement of a plastic film}
\format{
A \code{filter_cps.mspct} object with 611 rows and 2 variables
}
\description{
A dataset containing the wavelengths at a varying interval and
  counts per second data. Two spectra are included a reference measurement
  of the light source plus a measurement of a plastic film.
}
\details{
A collection of two spectra as a \code{cps_mspct} object, each
 each spectrum being a \code{cps_spct} object with two columns:
 \itemize{ \item w.length (nm). \item cps (counts/s)  }
}
\seealso{
Other Spectral data examples:
\code{\link{A.illuminant.spct}},
\code{\link{D50.illuminant.spct}},
\code{\link{D65.illuminant.spct}},
\code{\link{Ler_leaf.spct}},
\code{\link{black_body.spct}},
\code{\link{ccd.spct}},
\code{\link{clear.spct}},
\code{\link{green_leaf.spct}},
\code{\link{phenylalanine.spct}},
\code{\link{photodiode.spct}},
\code{\link{sun.spct}},
\code{\link{sun_daily.spct}},
\code{\link{sun_evening.spct}},
\code{\link{two_filters.spct}},
\code{\link{two_sensors.mspct}},
\code{\link{water.spct}},
\code{\link{white_led.source_spct}}
}
\concept{Spectral data examples}
\keyword{datasets}
\keyword{internal}
