% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{findMultipleWl}
\alias{findMultipleWl}
\title{Find repeated w.length values}
\usage{
findMultipleWl(x, same.wls = TRUE)
}
\arguments{
\item{x}{a generic_spct object}

\item{same.wls}{logical If TRUE all spectra spected to share same w.length
values.}
}
\value{
integer Number of spectra, guessed from the number of copies of each
  individual w.length value.
}
\description{
Find repeated w.length values
}
