% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{format.tod_time}
\alias{format.tod_time}
\title{Encode in a Common Format}
\usage{
\method{format}{tod_time}(x, ..., sep = ":")
}
\arguments{
\item{x}{an R object}

\item{...}{ignored}

\item{sep}{character used as separator}
}
\description{
Format a \code{tod_time} object for pretty printing
}
\seealso{
Other astronomy related functions: 
\code{\link{day_night}()},
\code{\link{format.solar_time}()},
\code{\link{is.solar_time}()},
\code{\link{print.solar_time}()},
\code{\link{print.tod_time}()},
\code{\link{solar_time}()},
\code{\link{sun_angles}()}
}
\concept{astronomy related functions}
