% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photobio-utils.R
\name{formatted_range}
\alias{formatted_range}
\title{Compute range and format it}
\usage{
formatted_range(x, na.rm = TRUE, digits = 3, nsmall = 2, collapse = "..")
}
\arguments{
\item{x}{an R object}

\item{na.rm}{logical, indicating if \code{\link[base]{NA}}'s should be
omitted.}

\item{digits, nsmall}{numeric, passed to same name parameters of
\code{format()}.}

\item{collapse}{character, passed to same name parameter of \code{paste()}.}
}
\description{
Compute the range of an R object, and format it as string suitable for
printing.
}
\examples{
formatted_range(c(1, 3.5, -0.01))

}
\seealso{
\code{\link[base]{range}}, \code{\link[base]{format}} and
  \code{\link[base]{paste}}.
}
