% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getAfrType}
\alias{getAfrType}
\title{Get the "Afr.type" attribute}
\usage{
getAfrType(x)
}
\arguments{
\item{x}{a filter_spct or object_spct object}
}
\value{
character string
}
\description{
Function to read the "Afr.type" attribute of an existing filter_spct or
object_spct object.
}
\note{
If x is not a \code{filter_spct} or an \code{object_spct} object,
  \code{NA} is returned.
}
\examples{
getAfrType(polyester.spct)

}
\seealso{
Other Afr attribute functions: \code{\link{setAfrType}}
}
\concept{Afr attribute functions}
