% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getBSWFUsed}
\alias{getBSWFUsed}
\title{Get the "bswf.used" attribute}
\usage{
getBSWFUsed(x)
}
\arguments{
\item{x}{a source_spct object}
}
\value{
character string
}
\description{
Function to read the "time.unit" attribute of an existing source_spct object
}
\note{
if x is not a \code{source_spct} object, NA is returned
}
\examples{
getBSWFUsed(sun.spct)

}
\seealso{
Other BSWF attribute functions: 
\code{\link{setBSWFUsed}()}
}
\concept{BSWF attribute functions}
