% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getFilterProperties}
\alias{getFilterProperties}
\alias{filter_properties}
\alias{getFilterProperties.default}
\alias{getFilterProperties.filter_spct}
\alias{getFilterProperties.summary_filter_spct}
\alias{getFilterProperties.generic_mspct}
\title{Get the "filter.properties" attribute}
\usage{
getFilterProperties(x, return.null, ...)

filter_properties(x, return.null, ...)

\method{getFilterProperties}{default}(x, return.null = FALSE, ...)

\method{getFilterProperties}{filter_spct}(x, return.null = FALSE, ...)

\method{getFilterProperties}{summary_filter_spct}(x, return.null = FALSE, ...)

\method{getFilterProperties}{generic_mspct}(x, return.null = FALSE, ..., idx = getIdFactor(x))
}
\arguments{
\item{x}{a filter_spct object}

\item{return.null}{logical If true, \code{NULL} is returned if the attribute
is not set, otherwise the expected list is returned with all fields set to
\code{NA}.}

\item{...}{Allows use of additional arguments in methods for other classes.}

\item{idx}{character Name of the column with the names of the members of the
collection of spectra.}
}
\value{
a list with fields named \code{"Rfr.constant"} [\eqn{/1}],
  \code{"thickness"} [\eqn{m}] and \code{"attenuation.mode"}. If the
  attribute is not set, and \code{return.null} is FALSE, a list with fields
  set to \code{NA} is returned, otherwise, \code{NULL}.
}
\description{
Function to read the "filter.properties" attribute of an existing filter_spct
or a filter_mspct.
}
\note{
The method for collections of spectra returns the a tibble with a
  column of lists.
}
\examples{
filter_properties(polyester.spct)

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
