% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.metadata.r
\name{getHowMeasured}
\alias{getHowMeasured}
\alias{how_measured}
\alias{getHowMeasured.default}
\alias{getHowMeasured.generic_spct}
\alias{getHowMeasured.summary_generic_spct}
\alias{getHowMeasured.data.frame}
\alias{getHowMeasured.generic_mspct}
\title{Get the "how.measured" attribute}
\usage{
getHowMeasured(x, ...)

how_measured(x, ...)

\method{getHowMeasured}{default}(x, ...)

\method{getHowMeasured}{generic_spct}(x, ..., simplify = FALSE)

\method{getHowMeasured}{summary_generic_spct}(x, ..., simplify = FALSE)

\method{getHowMeasured}{data.frame}(x, ..., simplify = FALSE)

\method{getHowMeasured}{generic_mspct}(x, ..., idx = "spct.idx", simplify = FALSE)
}
\arguments{
\item{x}{an R object.}

\item{...}{Allows use of additional arguments in methods for other classes.}

\item{simplify}{logical If all members share the same attribute value return
one copy instead of a data.frame.}

\item{idx}{character Name of the column with the names of the members of the
collection of spectra.}
}
\value{
character vector An object containing a verbal description of the
  data.
}
\description{
Method to read the \code{"how.measured"} attribute of an R object.
}
\note{
The method for collections of spectra returns the
  a data frame with a column of character strings.
}
\examples{
my.spct <- sun.spct
how_measured(my.spct)
how_measured(my.spct) <- "Simulated with a radiation transfer model"
how_measured(my.spct)
how_measured(my.spct) <- NULL
how_measured(my.spct)

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
