% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getIdFactor}
\alias{getIdFactor}
\alias{id_factor}
\title{Get the "idfactor" attribute}
\usage{
getIdFactor(x)

id_factor(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
character
}
\description{
Function to read the \code{idfactor} attribute of an existing
\code{generic_spct}.
}
\note{
If x is not a \code{generic_spct} or an object of a derived class
  \code{NA} is returned.
}
\examples{
id_factor(sun_evening.spct)

}
\seealso{
Other idfactor attribute functions:
\code{\link[=setIdFactor]{setIdFactor()}}
}
\concept{idfactor attribute functions}
