% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getKType}
\alias{getKType}
\title{Get the "K.type" attribute}
\usage{
getKType(x)
}
\arguments{
\item{x}{a solute_spct object}
}
\value{
character string
}
\description{
Function to read the "K.type" attribute of an existing solute_spct object.
}
\note{
If x is not a \code{solute_spct} or a \code{summary_solute_spct}
  object, \code{NA} is returned.
}
\examples{
print("missing example")

}
\seealso{
Other K attribute functions:
\code{\link[=setKType]{setKType()}}
}
\concept{K attribute functions}
