% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.methods.R
\name{getMspctVersion}
\alias{getMspctVersion}
\title{Get the "mspct.version" attribute}
\usage{
getMspctVersion(x)
}
\arguments{
\item{x}{a generic_mspct object}
}
\value{
numeric value
}
\description{
Function to read the "mspct.version" attribute of an existing generic_mspct
object.
}
\note{
if x is not a \code{generic_mspct} object, \code{NA} is returned,
  and if it the attribute is missing, zero is returned with a warning.
}
