% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getRfrType}
\alias{getRfrType}
\title{Get the "Rfr.type" attribute}
\usage{
getRfrType(x)
}
\arguments{
\item{x}{a source_spct object}
}
\value{
character string
}
\description{
Function to read the "Rfr.type" attribute of an existing reflector_spct
object or object_spct object.
}
\note{
if x is not a \code{filter_spct} object, \code{NA} is returned
}
\seealso{
Other Rfr attribute functions: 
\code{\link{setRfrType}()}
}
\concept{Rfr attribute functions}
