% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.fscale.r
\name{getScaled}
\alias{getScaled}
\alias{getScaling}
\title{Get the "scaled" attribute}
\usage{
getScaled(x, .force.list = FALSE)

getScaling(x)
}
\arguments{
\item{x}{a generic_spct object}

\item{.force.list}{logical If \code{TRUE} always silently return a
list, with \code{FALSE} encoded field \code{multiplier = 1}.}
}
\value{
logical
}
\description{
Function to read the "scaled" attribute of an existing generic_spct
object.
}
\note{
if x is not a \code{filter_spct} object, \code{NA} is returned
}
\examples{

scaled.spct <- fscale(sun.spct)
getScaled(scaled.spct)

}
\seealso{
Other rescaling functions:
\code{\link[=fscale]{fscale()}},
\code{\link[=fshift]{fshift()}},
\code{\link[=getNormalized]{getNormalized()}},
\code{\link[=is_normalized]{is_normalized()}},
\code{\link[=is_scaled]{is_scaled()}},
\code{\link[=normalize]{normalize()}},
\code{\link[=setNormalized]{setNormalized()}},
\code{\link[=setScaled]{setScaled()}}
}
\concept{rescaling functions}
