% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getSensorProperties}
\alias{getSensorProperties}
\alias{sensor_properties}
\alias{getSensorProperties.default}
\alias{getSensorProperties.response_spct}
\alias{getSensorProperties.summary_response_spct}
\alias{getSensorProperties.generic_mspct}
\title{Get the "sensor.properties" attribute}
\usage{
getSensorProperties(x, ...)

sensor_properties(x, ...)

\method{getSensorProperties}{default}(x, return.null = FALSE, ...)

\method{getSensorProperties}{response_spct}(x, return.null = FALSE, ...)

\method{getSensorProperties}{summary_response_spct}(x, return.null = FALSE, ...)

\method{getSensorProperties}{generic_mspct}(x, ..., idx = getIdFactor(x))
}
\arguments{
\item{x}{a source_spct object}

\item{...}{Allows use of additional arguments in methods for other classes.}

\item{return.null}{logical If true, \code{NULL} is returned if the attribute
is not set, otherwise the expected list is returned with all fields set to
\code{NA}.}

\item{idx}{character Name of the column with the names of the members of the
collection of spectra.}
}
\value{
an object of class \code{sensor_properties} derived from \code{list},
  guaranteed to contain fields named \code{"sensor.name"},
  \code{"sensor.type"} and \code{"num.channels"}. If the
  attribute is not set, these fields are set to \code{NA} in the returned
  object.
}
\description{
Function to read the "sensor.properties" attribute of an existing source_spct
or a source_mspct.
}
\note{
The method for collections of spectra returns a tibble with a
  column of lists.
}
\examples{
sensor_properties(ccd.spct)

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
