% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getSoluteProperties}
\alias{getSoluteProperties}
\alias{solute_properties}
\alias{getSoluteProperties.default}
\alias{getSoluteProperties.solute_spct}
\alias{getSoluteProperties.summary_solute_spct}
\alias{getSoluteProperties.solute_mspct}
\title{Get the "solute.properties" attribute}
\usage{
getSoluteProperties(x, return.null, ...)

solute_properties(x, return.null, ...)

\method{getSoluteProperties}{default}(x, return.null = FALSE, ...)

\method{getSoluteProperties}{solute_spct}(x, return.null = FALSE, ...)

\method{getSoluteProperties}{summary_solute_spct}(x, return.null = FALSE, ...)

\method{getSoluteProperties}{solute_mspct}(x, return.null = FALSE, ..., idx = getIdFactor(x))
}
\arguments{
\item{x}{solute_spct A spectrum of coefficients of attenuation.}

\item{return.null}{logical If true, \code{NULL} is returned if the attribute
is not set, otherwise the expected list is returned with all fields set to
\code{NA}.}

\item{...}{Allows use of additional arguments in methods for other classes.}

\item{idx}{character Name of the column with the names of the members of the
collection of spectra.}
}
\value{
a \code{list} with fields named \code{"mass"}, \code{"formula"},
  \code{"structure"}, \code{"name"} and \code{"ID"}. If the attribute is not
  set, and \code{return.null} is \code{FALSE}, a list with fields set to
  \code{NA} is returned, otherwise, \code{NULL}.
}
\description{
Function to read the \code{"solute.properties"} attribute of an existing
\code{solute_spct} or a \code{solute_mspct} objects.
}
\note{
The method for collections of spectra returns the a tibble with a
  column of lists.
}
\examples{
solute_properties(water.spct)

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
