% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getSpctVersion}
\alias{getSpctVersion}
\title{Get the "spct.version" attribute}
\usage{
getSpctVersion(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
integer value
}
\description{
Function to read the "spct.version" attribute of an existing generic_spct
object.
}
\note{
if x is not a \code{generic_spct} object, \code{NA} is returned,
  and if it the attribute is missing, zero is returned with a warning.
}
