% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getTimeUnit}
\alias{getTimeUnit}
\title{Get the "time.unit" attribute of an existing source_spct object}
\usage{
getTimeUnit(x, force.duration = FALSE)
}
\arguments{
\item{x}{a source_spct object}

\item{force.duration}{logical If TRUE a lubridate::duration is returned even
if the object attribute is a character string, if no conversion is possible
NA is returned.}
}
\value{
character string or a lubridate::duration
}
\description{
Function to read the "time.unit" attribute
}
\note{
if x is not a \code{source_spct} or a \code{response_spct} object, NA
is returned
}
\examples{
getTimeUnit(sun.spct)

}
\seealso{
Other time attribute functions:
\code{\link[=checkTimeUnit]{checkTimeUnit()}},
\code{\link[=convertThickness]{convertThickness()}},
\code{\link[=convertTimeUnit]{convertTimeUnit()}},
\code{\link[=setTimeUnit]{setTimeUnit()}}
}
\concept{time attribute functions}
