% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.metadata.r
\name{getWhenMeasured}
\alias{getWhenMeasured}
\alias{when_measured}
\alias{getWhenMeasured.default}
\alias{getWhenMeasured.generic_spct}
\alias{getWhenMeasured.summary_generic_spct}
\alias{getWhenMeasured.data.frame}
\alias{getWhenMeasured.generic_mspct}
\title{Get the "when.measured" attribute}
\usage{
getWhenMeasured(x, ...)

when_measured(x, ...)

\method{getWhenMeasured}{default}(x, ...)

\method{getWhenMeasured}{generic_spct}(x, as.df = FALSE, ..., simplify = FALSE)

\method{getWhenMeasured}{summary_generic_spct}(x, as.df = FALSE, ..., simplify = FALSE)

\method{getWhenMeasured}{data.frame}(x, as.df = FALSE, ..., simplify = FALSE)

\method{getWhenMeasured}{generic_mspct}(x, ..., idx = "spct.idx", simplify = FALSE)
}
\arguments{
\item{x}{an R object}

\item{...}{Allows use of additional arguments in methods for other classes.}

\item{as.df}{logical If \code{TRUE} return a data frame instead of a list,
when the value stored in the attribute is a list.}

\item{simplify}{logical If all members share the same attribute value return
one copy instead of a data.frame.}

\item{idx}{character Name of the column with the names of the members of the
collection of spectra.}
}
\value{
a \code{POSIXct} object with date and time, or named list of such
  objects, or, on user request, a data frame.
}
\description{
Method to read the \code{"when.measured"} attribute of an R object.
}
\note{
If \code{x} is not an object of one of the supported classes,
  \code{NA} is returned.

The method for collections of spectra returns a tibble with the
  times expressed in TZ = "UTC".
}
\examples{
my.spct <- sun.spct
when_measured(my.spct)
when_measured(my.spct) <- lubridate::ymd_hms("2020-01-01 08:00:00")
when_measured(my.spct)
when_measured(my.spct) <- NULL
when_measured(my.spct)

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
