% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.metadata.r
\name{getWhereMeasured}
\alias{getWhereMeasured}
\alias{where_measured}
\alias{getWhereMeasured.default}
\alias{getWhereMeasured.generic_spct}
\alias{getWhereMeasured.summary_generic_spct}
\alias{getWhereMeasured.generic_mspct}
\alias{getWhereMeasured.data.frame}
\title{Get the "where.measured" attribute}
\usage{
getWhereMeasured(x, ...)

where_measured(x, ...)

\method{getWhereMeasured}{default}(x, ...)

\method{getWhereMeasured}{generic_spct}(
  x,
  ...,
  idx = getIdFactor(x),
  simplify = TRUE,
  .bind.geocodes = TRUE
)

\method{getWhereMeasured}{summary_generic_spct}(
  x,
  ...,
  idx = getIdFactor(x),
  simplify = TRUE,
  .bind.geocodes = TRUE
)

\method{getWhereMeasured}{generic_mspct}(
  x,
  ...,
  idx = "spct.idx",
  .bind.geocodes = TRUE,
  simplify = FALSE
)

\method{getWhereMeasured}{data.frame}(x, ...)
}
\arguments{
\item{x}{a generic_spct object}

\item{...}{Allows use of additional arguments in methods for other classes.}

\item{idx}{character Name of the column with the names of the members of the
collection of spectra.}

\item{simplify}{logical If all members share the same attribute value return
one copy instead of a data.frame.}

\item{.bind.geocodes}{logical In the case of collections of spectra if
\code{.bind.geocodes = TRUE}, the default, the returned value is a single
geocode with one row for each member spectrum. Otherwise the individual
geocode data frames are returned in a list column within a tibble.}
}
\value{
a data.frame with a single row and at least columns "lon" and "lat",
   unless expand is set to \code{FALSE}.
}
\description{
Method to read the "where.measured" attribute of
generic_spct, generic_mspct, summary_generic_spct, data.frame or a
derived-class object.
}
\note{
If x is not a \code{generic_spct} or an object of a derived class
  \code{NA} is returned.
}
\examples{
getWhereMeasured(sun.spct)
getWhereMeasured(sun_evening.spct)
getWhereMeasured(sun_evening.spct, simplify = TRUE)
getWhereMeasured(sun_evening.mspct)
getWhereMeasured(sun_evening.mspct, .bind.geocodes = FALSE)
getWhereMeasured(sun_evening.mspct, simplify = TRUE)
getWhereMeasured(sun_evening.mspct, simplify = FALSE)

getWhereMeasured(summary(sun.spct))
getWhereMeasured(summary(sun_evening.spct))
getWhereMeasured(summary(sun_evening.mspct, expand = "each")[[1]])

getWhereMeasured(polyester.spct)

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
