% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.data.R
\docType{data}
\name{green_leaf.spct}
\alias{green_leaf.spct}
\title{Green birch leaf reflectance.}
\format{
A \code{reflector_spct} object with 226 rows and 2 variables
}
\usage{
green_leaf.spct
}
\description{
A dataset of spectral reflectance expressed as a fraction of one.
}
\details{
\itemize{ \item w.length (nm) \item Rfr (0..1) }
}
\examples{
green_leaf.spct

}
\references{
Aphalo, P. J. & Lehto, T. Effects of light quality on growth and N
accumulation in birch seedlings Tree Physiology, 1997, 17, 125-132
}
\seealso{
Other Spectral data examples:
\code{\link{A.illuminant.spct}},
\code{\link{D50.illuminant.spct}},
\code{\link{D65.illuminant.spct}},
\code{\link{Ler_leaf.spct}},
\code{\link{black_body.spct}},
\code{\link{ccd.spct}},
\code{\link{clear.spct}},
\code{\link{filter_cps.mspct}},
\code{\link{phenylalanine.spct}},
\code{\link{photodiode.spct}},
\code{\link{sun.spct}},
\code{\link{sun_daily.spct}},
\code{\link{sun_evening.spct}},
\code{\link{two_filters.spct}},
\code{\link{two_sensors.mspct}},
\code{\link{water.spct}},
\code{\link{white_led.source_spct}}
}
\concept{Spectral data examples}
\keyword{datasets}
