% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.integrate.r
\name{integrate_spct}
\alias{integrate_spct}
\title{Integrate spectral data.}
\usage{
integrate_spct(spct)
}
\arguments{
\item{spct}{generic_spct}
}
\value{
One or more numeric values with no change in scale factor: e.g. [W
  m-2 nm-1] -> [W m-2]. Each value in the returned vector corresponds to a
  variable in the spectral object, except for wavelength.
}
\description{
This function gives the result of integrating spectral data over wavelengths.
}
\examples{

integrate_spct(sun.spct)

}
