% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.upgrade.r
\name{is.old_spct}
\alias{is.old_spct}
\title{Query if an object has old class names}
\usage{
is.old_spct(object)
}
\arguments{
\item{object}{an R object}
}
\value{
logical
}
\description{
Query if an object has old class names Query if an object has old class names
as used in photobiology (>= 0.6.0).
}
\seealso{
Other upgrade from earlier versions:
\code{\link[=upgrade_spct]{upgrade_spct()}},
\code{\link[=upgrade_spectra]{upgrade_spectra()}}
}
\concept{upgrade from earlier versions}
