% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{is.solar_time}
\alias{is.solar_time}
\alias{is.solar_date}
\title{Query class}
\usage{
is.solar_time(x)

is.solar_date(x)
}
\arguments{
\item{x}{an R object.}
}
\description{
Query class
}
\seealso{
Other Local solar time functions: 
\code{\link{as.solar_date}()},
\code{\link{print.solar_time}()},
\code{\link{solar_time}()}
}
\concept{Local solar time functions}
