% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.summaries.r
\name{is.summary_generic_spct}
\alias{is.summary_generic_spct}
\alias{is.summary_raw_spct}
\alias{is.summary_cps_spct}
\alias{is.summary_source_spct}
\alias{is.summary_response_spct}
\alias{is.summary_filter_spct}
\alias{is.summary_reflector_spct}
\alias{is.summary_object_spct}
\alias{is.summary_chroma_spct}
\alias{is.any_summary_spct}
\title{Query class of spectrum summary objects}
\usage{
is.summary_generic_spct(x)

is.summary_raw_spct(x)

is.summary_cps_spct(x)

is.summary_source_spct(x)

is.summary_response_spct(x)

is.summary_filter_spct(x)

is.summary_reflector_spct(x)

is.summary_object_spct(x)

is.summary_chroma_spct(x)

is.any_summary_spct(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
These functions return \code{TRUE} if its argument is a of the queried type
  of spectrum and \code{FALSE} otherwise.
}
\description{
Functions to check if an object is of a given type of spectrum, or coerce it if
possible.
}
\note{
Derived types also return TRUE for a query for a base type such as
\code{generic_spct}.
}
\examples{
sm <- summary(sun.spct)
is.summary_source_spct(sm)

}
