% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.new.r
\name{is.waveband}
\alias{is.waveband}
\title{Query if it is a waveband}
\usage{
is.waveband(x)
}
\arguments{
\item{x}{any R object}
}
\value{
is.waveband returns TRUE if its argument is a waveband and FALSE
  otherwise.
}
\description{
Functions to check if an object is waveband.
}
