% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{is_effective}
\alias{is_effective}
\alias{is_effective.default}
\alias{is_effective.waveband}
\alias{is_effective.generic_spct}
\alias{is_effective.source_spct}
\alias{is_effective.summary_generic_spct}
\alias{is_effective.summary_source_spct}
\title{Is an R object "effective"}
\usage{
is_effective(x)

\method{is_effective}{default}(x)

\method{is_effective}{waveband}(x)

\method{is_effective}{generic_spct}(x)

\method{is_effective}{source_spct}(x)

\method{is_effective}{summary_generic_spct}(x)

\method{is_effective}{summary_source_spct}(x)
}
\arguments{
\item{x}{an R object}
}
\value{
A \code{logical}.
}
\description{
A generic function for querying if a biological spectral weighting function
(BSWF) has been applied to an object or is included in its definition.
}
\examples{
is_effective(summary(sun.spct))

}
\seealso{
Other waveband attributes:
\code{\link{labels}},
\code{\link[=normalization]{normalization()}}
}
\concept{waveband attributes}
