% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.normalize.r
\name{is_normalized}
\alias{is_normalized}
\alias{is_normalised}
\title{Query whether a generic spectrum has been normalized.}
\usage{
is_normalized(x)

is_normalised(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A \code{logical} value indicating if \code{x} is normalized or not,
  for collections of spectra, a named list with \code{logicals} as members.
  If \code{x} is not a \code{generic_spct} or \code{generic_mspct} object the
  value returned is \code{NA}.
}
\description{
This function tests a \code{generic_spct} object for an attribute that
signals whether the spectral data has been normalized or not after the object
was created.
}
\note{
While method \code{getNormalized()} returns in some cases numeric
  values or possibly even character values if stored in attribute
  \code{normalized}, \code{is_normalized()} always returns a logical value
  and can be safely used in conditional code clauses.

\code{is_normalised()} is another name for \code{is_normalized()}.
}
\examples{
norm_sun.spct <- normalize(sun.spct)
is_normalized(norm_sun.spct)

norm_sun_evening.mspct <- normalize(sun_evening.mspct[1:3])
str(is_normalized(norm_sun_evening.mspct))

}
\seealso{
Other rescaling functions:
\code{\link[=fscale]{fscale()}},
\code{\link[=fshift]{fshift()}},
\code{\link[=getNormalized]{getNormalized()}},
\code{\link[=getScaled]{getScaled()}},
\code{\link[=is_scaled]{is_scaled()}},
\code{\link[=normalize]{normalize()}},
\code{\link[=setNormalized]{setNormalized()}},
\code{\link[=setScaled]{setScaled()}}
}
\concept{rescaling functions}
