% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.fscale.r
\name{is_scaled}
\alias{is_scaled}
\title{Query whether a generic spectrum has been scaled}
\usage{
is_scaled(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A \code{logical} value. If \code{x} is not scaled or \code{x} is
  not a \code{generic_spct} object the value returned is \code{FALSE}.
}
\description{
This function tests a \code{generic_spct} object for an attribute that
signals whether the spectral data has been rescaled or not after the object
was created.
}
\examples{

scaled.spct <- fscale(sun.spct)
is_scaled(sun.spct)
is_scaled(scaled.spct)

}
\seealso{
Other rescaling functions:
\code{\link[=fscale]{fscale()}},
\code{\link[=fshift]{fshift()}},
\code{\link[=getNormalized]{getNormalized()}},
\code{\link[=getScaled]{getScaled()}},
\code{\link[=is_normalized]{is_normalized()}},
\code{\link[=normalize]{normalize()}},
\code{\link[=setNormalized]{setNormalized()}},
\code{\link[=setScaled]{setScaled()}}
}
\concept{rescaling functions}
