% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{is_tagged}
\alias{is_tagged}
\title{Query if a spectrum is tagged}
\usage{
is_tagged(x)
}
\arguments{
\item{x}{any R object}
}
\value{
\code{is_tagged} returns a logical value, \code{TRUE} if its
  argument is a a spectrum that contains tags and \code{FALSE} if it
  is an untagged spectrum, but returns \code{NA} for any other R object.
}
\description{
Functions to check if an spct object contains tags.
}
\examples{
is_tagged(sun.spct)

}
\seealso{
Other tagging and related functions:
\code{\link[=tag]{tag()}},
\code{\link[=untag]{untag()}},
\code{\link[=wb2rect_spct]{wb2rect_spct()}},
\code{\link[=wb2spct]{wb2spct()}},
\code{\link[=wb2tagged_spct]{wb2tagged_spct()}}
}
\concept{tagging and related functions}
