% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.metadata.r
\name{isValidInstrDesc}
\alias{isValidInstrDesc}
\title{Check the "instr.desc" attribute}
\usage{
isValidInstrDesc(x)
}
\arguments{
\item{x}{a \code{generic_spct} object or a \code{summary_generic_spct}
object.}
}
\value{
A \code{logical} vector of length one.
}
\description{
Function to validate the \code{"instr.settings"} attribute of an existing
\code{generic_spct} object or \code{summary_generic_spct} object.
}
\details{
Test if at least one of instrument name (field
 \code{spectrometer.name}) or serial number (field \code{spectrometer.sn})
 is found in the value of the R attribute \code{"instr.desc"} of \code{x}.
 \code{FALSE} is silently returned if \code{x} does not belong to a class
 derived from class \code{generic_spct} or from class
 \code{summary_generic_spct}, or if it is derived from these classes but the
 attribute is not set.
}
\examples{
isValidInstrDesc(white_led.cps_spct)
isValidInstrDesc(white_body.spct)

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
