% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.summaries.r
\name{labels.generic_spct}
\alias{labels.generic_spct}
\title{Find labels from spectral object}
\usage{
\method{labels}{generic_spct}(object, ...)
}
\arguments{
\item{object}{an object of generic_spct}

\item{...}{not used in current version}
}
\description{
A function to obtain the labels of a spectrum. Currently returns 'names'.
}
\examples{
labels(sun.spct)

}

