% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{labels.waveband}
\alias{labels.waveband}
\title{Find labels from "waveband" object}
\usage{
\method{labels}{waveband}(object, ...)
}
\arguments{
\item{object}{an object of class "waveband"}

\item{...}{not used in current version}
}
\description{
A function to obtain the name and label of objects of class "waveband".
}
\seealso{
Other waveband attributes: \code{\link{is_effective.summary_generic_spct}},
  \code{\link{normalization}}
}

