% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.data.R
\docType{data}
\name{Ler_leaf_trns_i.spct}
\alias{Ler_leaf_trns_i.spct}
\title{Green Arabidopsis leaf spectral transmittance.}
\format{
An \code{filter_spct} object with 2401 rows and 2 variables
}
\usage{
Ler_leaf_trns_i.spct
}
\description{
A dataset of internal spectral transmittance expressed as a
  fraction of one from the upper surface of a leaf of an Arabidopsis thaliana
  'Ler' rosette.
}
\details{
\itemize{ \item w.length (nm) \item Tfr (0..1) }
}
\note{
Measured with a Jaz spectrometer from Ocean Optics (USA) configured
  with a PX Xenon lamp module and Spectroclip double integrating spheres.
}
\examples{
Ler_leaf_trns_i.spct

}
\seealso{
Other Spectral data examples: 
\code{\link{A.illuminant.spct}},
\code{\link{D65.illuminant.spct}},
\code{\link{Ler_leaf.spct}},
\code{\link{Ler_leaf_rflt.spct}},
\code{\link{Ler_leaf_trns.spct}},
\code{\link{black_body.spct}},
\code{\link{ccd.spct}},
\code{\link{clear.spct}},
\code{\link{clear_body.spct}},
\code{\link{filter_cps.mspct}},
\code{\link{green_leaf.spct}},
\code{\link{opaque.spct}},
\code{\link{photodiode.spct}},
\code{\link{polyester.spct}},
\code{\link{sun.daily.data}},
\code{\link{sun.daily.spct}},
\code{\link{sun.data}},
\code{\link{sun.spct}},
\code{\link{white_body.spct}},
\code{\link{white_led.cps_spct}},
\code{\link{white_led.raw_spct}},
\code{\link{white_led.source_spct}},
\code{\link{yellow_gel.spct}}
}
\author{
Aphalo, P. J. & Wang, F (unpublished data)
}
\concept{Spectral data examples}
\keyword{datasets}
