% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.row.funs.r
\name{make_wl_consistent}
\alias{make_wl_consistent}
\title{Make wavelengths consistent}
\usage{
make_wl_consistent(x, w.length.out = NULL)
}
\arguments{
\item{x}{a generic_spct object}

\item{w.length.out}{numeric vector of wavelengths (nanometres)}
}
\description{
Ensure wavelengths in all spectra in a \code{mspct} object are consistent,
filling head and tail of spectral quantities when needed with NAs, and using
interpolation if needed. If individual \code{spct} objects contain multiple
spectra in long form these are split into multiple \code{spct} objects, each
containing data for one spectrum.
}
\keyword{internal}
