% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.methods.R, R/spct.summaries.r
\name{max.generic_mspct}
\alias{max.generic_mspct}
\alias{max.generic_spct}
\title{Maximun wavelength}
\usage{
\method{max}{generic_mspct}(..., na.rm = FALSE, idx = NULL)

\method{max}{generic_spct}(..., na.rm = FALSE)
}
\arguments{
\item{...}{not used in current version}

\item{na.rm}{a logical indicating whether missing values should be removed.}

\item{idx}{logical whether to add a column with the names of the elements of
spct}
}
\description{
Maximun function for spectra, returning wavelength maximum.
}
\examples{
max(sun.spct)

}
\seealso{
Other wavelength summaries: \code{\link{midpoint.generic_mspct}},
  \code{\link{min.generic_mspct}},
  \code{\link{min.waveband}},
  \code{\link{range.generic_mspct}},
  \code{\link{range.waveband}},
  \code{\link{stepsize.generic_mspct}}
}

