% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{max.waveband}
\alias{max.waveband}
\title{Wavelength maximum}
\usage{
\method{max}{waveband}(..., na.rm = FALSE)
}
\arguments{
\item{...}{not used in current version}

\item{na.rm}{ignored}
}
\description{
A function that returns the wavelength maximum from objects of class "waveband".
}

