% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.attr.R
\name{merge_attributes}
\alias{merge_attributes}
\alias{merge_attributes.default}
\alias{merge_attributes.generic_spct}
\title{Merge and copy attributes}
\usage{
merge_attributes(x, y, z, which, which.not, ...)

\method{merge_attributes}{default}(x, y, z, which = NULL, which.not = NULL, ...)

\method{merge_attributes}{generic_spct}(
  x,
  y,
  z,
  which = NULL,
  which.not = NULL,
  copy.class = FALSE,
  ...
)
}
\arguments{
\item{x, y, z}{R objects. Objects \code{x} and \code{y} must be of the same class,
\code{z} must be an object with a structure valid for this same class.}

\item{which}{character Names of attributes to copy, if NULL all those
relevant according to the class of \code{x} are used as default,}

\item{which.not}{character Names of attributes not to be copied. The
names passed here are removed from the list for \code{which}, which
is most useful when we want to modify the default.}

\item{...}{not used}

\item{copy.class}{logical If TRUE class attributes are also copied.}
}
\value{
A copy of \code{z} with additional attributes set.
}
\description{
Merge attributes from \code{x} and \code{y} and copy them to \code{z}.
Methods defined for spectral objects of classes from package 'photobiology'.
}
