% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{wl_min}
\alias{wl_min}
\alias{min}
\alias{min.waveband}
\alias{min.generic_spct}
\alias{min.generic_mspct}
\title{Wavelength minimum}
\usage{
wl_min(x, na.rm = FALSE)

\method{min}{waveband}(..., na.rm = FALSE)

\method{min}{generic_spct}(..., na.rm = FALSE)

\method{min}{generic_mspct}(..., na.rm = FALSE, idx = "spct.idx")
}
\arguments{
\item{x}{generic_spct, generic_mspct or waveband object.}

\item{na.rm}{ignored}

\item{...}{not used in current version}

\item{idx}{character Name of the column with the names of the members of the
collection of spectra.}
}
\value{
a length-one vector for individual objects or numeric vectors
  or a data frame for collections of spectra.
}
\description{
A method specialization that returns the wavelength minimum [\eqn{nm}] from
objects of classes \code{waveband} or of class \code{generic_spct} or
derived.
}
\examples{
min(sun.spct)
wl_min(sun.spct)

}
\seealso{
Other wavelength summaries:
\code{\link[=wl_midpoint]{wl_midpoint()}},
\code{\link[=wl_range]{wl_range()}},
\code{\link[=wl_stepsize]{wl_stepsize()}}
}
\concept{wavelength summaries}
