% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{min.waveband}
\alias{min.waveband}
\title{Wavelength minimum}
\usage{
\method{min}{waveband}(..., na.rm = FALSE)
}
\arguments{
\item{...}{not used in current version}

\item{na.rm}{ignored}
}
\description{
A function that returns the wavelength minimum from objects of class "waveband".
}
\seealso{
Other wavelength summaries: \code{\link{max.generic_mspct}},
  \code{\link{midpoint.generic_mspct}},
  \code{\link{min.generic_mspct}},
  \code{\link{range.generic_mspct}},
  \code{\link{range.waveband}},
  \code{\link{stepsize.generic_mspct}}
}

