% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.methods.R
\name{msmsply}
\alias{msaply}
\alias{msdply}
\alias{mslply}
\alias{msmsply}
\title{Multi-spct transform methods}
\usage{
msmsply(mspct, .fun, ...)

msdply(mspct, .fun, ..., idx = NULL, col.names = NULL)

mslply(mspct, .fun, ...)

msaply(mspct, .fun, ..., .drop = TRUE)
}
\arguments{
\item{mspct}{an object of class generic_mspct or a derived class}

\item{.fun}{a function}

\item{...}{other arguments passed to .fun}

\item{idx}{logical whether to add a column with the names of the elements of
mspct, if \code{NULL}, the default, a column is added only if all members
of \code{mscpt} are named.}

\item{col.names}{character Names to be used for data columns.}

\item{.drop}{should extra dimensions of length 1 in the output be dropped,
simplifying the output. Defaults to TRUE}
}
\value{
a collection of spectra in the case of \code{msmsply}

a data frame in the case of \code{msdply}

a list in the case of \code{mslply}

an array in the case of \code{msaply}
}
\description{
Apply a function or operator to a collection of spectra.
}

