% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmspct.classes.R
\name{mspct_classes}
\alias{mspct_classes}
\title{Names of multi-spectra classes}
\usage{
mspct_classes()
}
\value{
A \code{character} vector of class names.
}
\description{
Function that returns a vector containing the names of multi-spectra classes
using for collections of spectra.
}
\examples{
mspct_classes()

}
