% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.et.r
\name{net_lw_radiation}
\alias{net_lw_radiation}
\title{Net long wave radiation}
\usage{
net_lw_radiation(
  temperature,
  lw.down.irradiance = NULL,
  water.vp = 0,
  R_rel = 0.75,
  emissivity = 0.98
)
}
\arguments{
\item{temperature}{numeric vector of air temperatures (C) at 2 m height.}

\item{lw.down.irradiance}{numeric Long wave downwelling radiation (W/m2)}

\item{water.vp}{numeric vector of water vapour pressure in air (Pa),
ignored if \code{lw.down.irradiance} is available.}

\item{R_rel}{numeric The ratio of short wave irradiance at ground level and
at the top of the atmosphere (/1).}

\item{emissivity}{numeric Emissivity of the surface (ground or vegetation)
for long wave radiation.}
}
\value{
A numeric vector of $R_{nl}$ estimates expressed as W / m-2.
}
\description{
Estimate net long-wave ration balance. If
\code{lw.down.irradiance} is passed a value in W / m2 the difference is
computed directly and if not an approximate value is estimated, using
\code{R_rel = 0.75} the vales are for clear sky, i.e., uncorrected for
cloudiness. This is the approach to estimation is that recommended by FAO
for hourly estimates while here we use it for instantaneous or mean flux
rates.
}
\seealso{
Other Evapotranspiration and energy balance related functions.: 
\code{\link{ET_ref}()},
\code{\link{ET_zero}()}
}
\concept{Evapotranspiration and energy balance related functions.}
