% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.data.R
\docType{data}
\name{opaque.spct}
\alias{opaque.spct}
\title{Theoretical spectrum of an opaque material}
\format{A \code{filter_spct} object with 4 rows and 2 variables}
\description{
A dataset for a hypothetical object with transmittance 0/1
  (0\%)
}
\details{
\itemize{ \item w.length (nm). \item Tfr
(0..1)  }
}
\keyword{datasets}
