% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.methods.R
\name{print.generic_mspct}
\alias{print.generic_mspct}
\title{Print a spectral collection object}
\usage{
\method{print}{generic_mspct}(x, ..., n = NULL, width = NULL)
}
\arguments{
\item{x}{generic_mspct A collection of spectra}

\item{...}{not used in current version}

\item{n}{Number of rows to show for each member spectrum.}

\item{width}{Width of text output to generate.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for objects of the collection of spectra classes. Any object
of a class derived from \code{geenric_mspct} is printed with this  method.
}
\seealso{
\code{\link{print.generic_spct}} for details on \code{n}
and \code{width}.
}

