% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.waveband.r
\name{print.waveband}
\alias{print.waveband}
\title{Print a "waveband" object}
\usage{
\method{print}{waveband}(x, ...)
}
\arguments{
\item{x}{an object of class "waveband"}

\item{...}{not used in current version}
}
\description{
A function to more nicely print objects of class "waveband".
}
