% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{put_hinges}
\alias{put_hinges}
\title{Insert interpolated spectral irradiance values.}
\arguments{
\item{x}{numeric array of wavelength (nm)}

\item{y}{numeric array of spectral irradiance values}

\item{hinges}{a numeric array giving the wavelengths at which the s.irrad
should be inserted by interpolation, no interpolation is indicated by an
empty array (numeric(0))}
}
\value{
a numeric vector
}
\description{
Inserting wavelengths values immediately before and after a discontinuity in
the SWF, greatly reduces the errors caused by interpolating the weighted
irradiance during integration of the effective spectral irradiance. This is
specially true when data has a large wavelength step size.
}
\keyword{manip}
\keyword{misc}

