% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-spct-ratios.R
\name{q_ratio}
\alias{q_ratio}
\alias{q_ratio.default}
\alias{q_ratio.source_spct}
\alias{q_ratio.source_mspct}
\title{Photon:photon ratio}
\usage{
q_ratio(
  spct,
  w.band.num,
  w.band.denom,
  scale.factor,
  wb.trim,
  use.cached.mult,
  use.hinges,
  ...
)

\method{q_ratio}{default}(
  spct,
  w.band.num,
  w.band.denom,
  scale.factor,
  wb.trim,
  use.cached.mult,
  use.hinges,
  ...
)

\method{q_ratio}{source_spct}(
  spct,
  w.band.num = NULL,
  w.band.denom = NULL,
  scale.factor = 1,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.cached.mult = FALSE,
  use.hinges = NULL,
  quantity = "total",
  naming = "short",
  name.tag = NULL,
  ...
)

\method{q_ratio}{source_mspct}(
  spct,
  w.band.num = NULL,
  w.band.denom = NULL,
  scale.factor = 1,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.cached.mult = FALSE,
  use.hinges = NULL,
  quantity = "total",
  naming = "short",
  name.tag = "[q:q]",
  ...,
  attr2tb = NULL,
  idx = "spct.idx",
  .parallel = FALSE,
  .paropts = NULL
)
}
\arguments{
\item{spct}{an object of class "source_spct".}

\item{w.band.num}{waveband object or a list of waveband objects used to
compute the numerator(s) of the ratio(s).}

\item{w.band.denom}{waveband object or a list of waveband objects used to
compute the denominator(s) of the ratio(s).}

\item{scale.factor}{numeric vector of length 1, or length equal to that of
\code{w.band}. Numeric multiplier applied to returned values.}

\item{wb.trim}{logical if TRUE wavebands crossing spectral data boundaries
are trimmed, if FALSE, they are discarded}

\item{use.cached.mult}{logical indicating whether multiplier values should be
cached between calls}

\item{use.hinges}{logical Flag indicating whether to insert "hinges" into the
spectral data before integration so as to reduce interpolation errors at
the boundaries of the wavebands.}

\item{...}{other arguments (possibly ignored)}

\item{quantity}{character One of "total", "average" or "mean".}

\item{naming}{character one of "long", "default", "short" or "none". Used to
select the type of names to assign to returned value.}

\item{name.tag}{character Used to tag the name of the returned values.}

\item{attr2tb}{character vector, see \code{\link{add_attr2tb}} for the syntax
for \code{attr2tb} passed as is to formal parameter \code{col.names}.}

\item{idx}{character Name of the column with the names of the members of the
collection of spectra.}

\item{.parallel}{if TRUE, apply function in parallel, using parallel backend
provided by foreach}

\item{.paropts}{a list of additional options passed into the foreach function
when parallel computation is enabled. This is important if (for example)
your code relies on external data or packages: use the .export and
.packages arguments to supply them so that all cluster nodes have the
correct environment set up for computing.}
}
\value{
In the case of methods for individual spectra, a \code{numeric}
  vector with name attribute set. The name is based on the name of the
  wavebands unless a named list of wavebands is supplied in which case the
  names of the list elements are used. "[q:q]" is appended if \code{quantity
  = "total"} and "[q(wl):q(wl)]" if \code{quantity = "mean"} or
  \code{quantity = "average"}.

  A \code{data.frame} is returned in the case of collections of spectra,
  containing one column for each fraction definition, an index column with
  the names of the spectra, and optionally additional columns with metadata
  values retrieved from the attributes of the member spectra.
}
\description{
This function returns the photon ratio for a given pair of wavebands of a
light source spectrum.
}
\details{
With the default \code{quantity = "total"} the ratio is based on
  two photon irradiances, one computed for each waveband.

  \deqn{\frac{Q(s, wb_\mathrm{num})}{Q(s, wb_\mathrm{denom})}}

If the argument is set to \code{quantity = "mean"} or
 \code{quantity = "average"} the ratio is based on
  two mean spectral photon irradiances, one computed for each waveband.

  \deqn{\frac{\overline{Q_\lambda}(s, wb_\mathrm{num})}{\overline{Q_\lambda}(s, wb_\mathrm{denom})}}

  Ratios based on totals and means are numerically identical only if the
  wavelength expanse of the two wavebands is the same.

  Fraction definitions are "assembled" from the arguments passed to
  \code{w.band.num} and \code{w.band.denom}. If both arguments are lists of
  waveband definitions, with an equal number of members, then the wavebands
  are paired to obtain as many fractions as the number of wavebands in each
  list. Recycling for wavebands takes place when the number of denominator
  and numerator wavebands differ.

  The last two parameters control speed optimizations. The defaults
  should be suitable in most cases. If you will use repeatedly the same SWFs
  on many spectra measured at exactly the same wavelengths you may obtain
  some speed up by setting \code{use.cached.mult=TRUE}. However, be aware
  that you are responsible for ensuring that the wavelengths are the same in
  each call, as the only test done is for the length of the \code{w.length}
  vector.
}
\section{Performance}{
 As this method accepts spectra as its input, it
  computes irradiances before computing the ratios. If you need to compute
  both ratios and irradiances from several hundreds or thousands of spectra,
  computing the ratios from previously computed irradiances avoids their
  repeated computation. A less dramatic, but still important, increase in
  performance is available when computing in the same function call ratios
  that share the same denominator.
}

\examples{
q_ratio(sun.spct,
        waveband(c(400,500), wb.name = "Blue"),
        waveband(c(400,700), wb.name = "White"))

}
\seealso{
Other photon and energy ratio functions:
\code{\link[=e_fraction]{e_fraction()}},
\code{\link[=e_ratio]{e_ratio()}},
\code{\link[=eq_ratio]{eq_ratio()}},
\code{\link[=q_fraction]{q_fraction()}},
\code{\link[=qe_ratio]{qe_ratio()}}
}
\concept{photon and energy ratio functions}
