% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4p.pkgs.data.r
\docType{data}
\name{r4p_pkgs}
\alias{r4p_pkgs}
\title{Packages in R for Photobiology suite}
\format{
A \code{character vector}.
}
\usage{
r4p_pkgs
}
\description{
A dataset containing the names of all the packages in this
  suite.
}
\details{
A character vector.
}
\examples{
r4p_pkgs

}
\keyword{datasets}
