% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{range.waveband}
\alias{range.waveband}
\title{Wavelength range}
\usage{
\method{range}{waveband}(..., na.rm = FALSE)
}
\arguments{
\item{...}{not used in current version}

\item{na.rm}{ignored}
}
\description{
A function that returns the wavelength range from objects of class "waveband".
}
\seealso{
Other wavelength summaries: \code{\link{max.generic_mspct}},
  \code{\link{midpoint.generic_mspct}},
  \code{\link{min.generic_mspct}},
  \code{\link{min.waveband}},
  \code{\link{range.generic_mspct}},
  \code{\link{stepsize.generic_mspct}}
}

