% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.reflectance.r
\name{reflectance_spct}
\alias{reflectance_spct}
\title{Calculate reflectance from spectral reflectance}
\usage{
reflectance_spct(spct, w.band, quantity, wb.trim, use.hinges, naming)
}
\arguments{
\item{spct}{an object of class generic_spct"}

\item{w.band}{waveband or list of waveband objects or a numeric vector of
length two. The waveband(s) determine the region(s) of the spectrum that
are summarized. If a numeric range is supplied a waveband object is
constructed on the fly from it.}

\item{quantity}{character string One of "total", "average" or "mean",
"contribution", "contribution.pc", "relative" or "relative.pc"}

\item{wb.trim}{logical if TRUE wavebands crossing spectral data boundaries
are trimmed, if FALSE, they are discarded}

\item{use.hinges}{logical Flag indicating whether to insert "hinges" into the
spectral data before integration so as to reduce interpolation errors at
the boundaries of the wavebands.}

\item{naming}{character one of "long", "default", "short" or "none". Used to
select the type of names to assign to returned value.}
}
\value{
A single numeric value expressed as a fraction of one
}
\description{
This function returns the mean reflectance for a given waveband and a
reflectance spectrum.
}
\keyword{internal}
