% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.normalize.r
\name{restore_normalization}
\alias{restore_normalization}
\title{Restore normalization}
\usage{
restore_normalization(x, old.normalization.ls)
}
\arguments{
\item{x}{generic_spct One or more spectra stored in long form, with no
normalization applied.}

\item{old.normalization.ls}{list A list describing the normalization criteria
in the format returned by \code{getNormalization()}.}
}
\description{
After altering the columns present in a spectrum restore normalization
with the previously used criteria.
}
\details{
The normalization criteria are extracted from
\code{old.normalization.ls} and applied. In the case of spectra in
long form, the normalization can be stored as a named list with
values for each spectrum as members.

If \code{x} contains multiple spectra in long form, the same normalization
criterion is applied to all of them based on \code{old.normalization.ls}.
If \code{getNormalization()} is a named list and stored values of
\code{norm.type} are equal, it is used. If \code{norm.type == "wavelength"},
\code{norm.wl} is used if consistent. \code{norm.range} is used as is if
consistent, and otherwise expanded to the joint range.
If metadata are no consistent across multiple spectra stored in long form
re-normalization is skipped with a warning.
}
\keyword{internal}
