% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.sample.R
\name{sample_spct}
\alias{sample_spct}
\alias{sample_mspct}
\title{Random sample from collection of spectra}
\usage{
sample_spct(x, size = 1)

sample_mspct(x, size = 1, recursive = FALSE)
}
\arguments{
\item{x}{generic_spct An object possibly containing multiple spectra.}

\item{size}{integer The number of spectra to extract, if available.}

\item{recursive}{logical Sample members of x directly, or first split
members containing multiple spectra.}
}
\value{
If \code{x} is an object, such as a
  "filter_spct" object, the returned object is of same class but possibly
  containing fewer spectra in long form than \code{x}.
}
\description{
A method to extract a random sample of spectra from a collection of spectra
or a spectrum object containing multiple spectra in long form.
}
\seealso{
See \code{\link[base]{sample}} for the method used for
  the sampling.
}
