% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.mean.se.band.r
\name{se.p}
\alias{se.p}
\title{Standard error of the mean}
\usage{
se.p(x, na.rm = FALSE, mult = 1, ...)
}
\arguments{
\item{x}{numeric vector}

\item{mult}{numeric number of multiples of standard error}
}
\description{
Standard error of the mean
}
\note{
mult can be used to construct confidence intervals
}
\keyword{internal}
