% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setFilterProperties}
\alias{setFilterProperties}
\alias{filter_properties<-}
\title{Set the "filter.properties" attribute}
\usage{
setFilterProperties(
  x,
  filter.properties = NULL,
  pass.null = FALSE,
  Rfr.constant = NA_real_,
  thickness = NA_real_,
  attenuation.mode = NA_character_,
  verbose = TRUE
)

filter_properties(x) <- value
}
\arguments{
\item{x}{a filter_spct object}

\item{filter.properties, value}{a list with fields named
\code{"Rfr.constant"}, \code{"thickness"} and
\code{"attenuation.mode"}.}

\item{pass.null}{logical If TRUE, the parameters to the next three
parameters will be always ignored, otherwise they will be used to
build an object of class \code{"filter.properties"} when the argument
passed to parameter \code{filter.properties} is \code{NULL}.}

\item{Rfr.constant}{numeric The value of the reflection factor [/1].}

\item{thickness}{numeric The thickness of the material [\eqn{m}].}

\item{attenuation.mode}{character One of \code{"reflection"},
\code{"absorption"}, \code{"absorption.layer"}, \code{"scattering"},
\code{"mixed"} or \code{"stack"}.}

\item{verbose}{logical Flag to enable warning when applied to object of
unsuported class.}
}
\value{
\code{x}
}
\description{
Function to set by reference the "filter.properties" attribute  of an
existing filter_spct object.
}
\details{
Storing filter properties allows inter-conversion between internal
  and total transmittance, as well as computation of transmittance for
  arbitrary thickness of the material. Whether computations are valid depend
  on the homogeneity of the material. The parameter \code{pass.null} makes
  it possible to remove the attribute.
}
\note{
This function alters \code{x} itself by reference and in addition
  returns \code{x} invisibly. If \code{x} is not a \code{filter_spct} object,
  \code{x} is not modified.

  The values of \code{attenuation.mode} \code{"reflection"},
  \code{"absorption"}, \code{"absorption.layer"} or \code{"scattering"}
  should be used when one of these processes is clearly the main one;
  \code{"mixed"} is for when multiple modes play a significanr role, i.e.,
  when a simple correction using a single value of \code{Rfr} across
  wavelengths is not possible; \code{"absorption.layer"} is for cases when a
  thin absorbing layer is deposited on the surface of a transparent support
  or enclosed between two sheets of glass or other transparent material.
  Finally \code{"stack"} is for multiple individual filters piled. If in
  doubt, set this argument to \code{NA} to ensure that computation of spectra
  for a different thickness remains disabled.
}
\examples{

my.spct <- polyester.spct
filter_properties(my.spct)
filter_properties(my.spct) <- NULL
filter_properties(my.spct)
filter_properties(my.spct, return.null = TRUE)
filter_properties(my.spct) <- list(Rfr.constant = 0.01,
                                   thickness = 125e-6,
                                   attenuation.mode = "absorption")
filter_properties(my.spct)

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
