% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.metadata.r
\name{setHowMeasured}
\alias{setHowMeasured}
\alias{how_measured<-}
\alias{setHowMeasured.default}
\alias{setHowMeasured.generic_spct}
\alias{setHowMeasured.summary_generic_spct}
\alias{setHowMeasured.data.frame}
\alias{setHowMeasured.generic_mspct}
\title{Set the "how.measured" attribute}
\usage{
setHowMeasured(x, ...)

how_measured(x) <- value

\method{setHowMeasured}{default}(x, how.measured, ...)

\method{setHowMeasured}{generic_spct}(x, how.measured, ...)

\method{setHowMeasured}{summary_generic_spct}(x, how.measured, ...)

\method{setHowMeasured}{data.frame}(x, how.measured, ...)

\method{setHowMeasured}{generic_mspct}(x, how.measured, ...)
}
\arguments{
\item{x}{a R object.}

\item{...}{Allows use of additional arguments in methods for other classes.}

\item{how.measured, value}{a list or a character string.}
}
\value{
x modified by reference.
}
\description{
Method to set the \code{"how.measured"} attribute of an R object.
}
\note{
This function alters \code{x} itself by reference and in addition
  returns \code{x} invisibly. If \code{x} is not an object of a supported
  class, \code{x} is silently returned unchanged.
}
\examples{
my.spct <- sun.spct
how_measured(my.spct)
how_measured(my.spct) <- "Simulated with a radiation transfer model"
how_measured(my.spct)
how_measured(my.spct) <- NULL
how_measured(my.spct)

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
