% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.metadata.r
\name{setInstrSettings}
\alias{setInstrSettings}
\alias{instr_settings<-}
\title{Set the "instr.settings" attribute}
\usage{
setInstrSettings(x, instr.settings)

instr_settings(x) <- value
}
\arguments{
\item{x}{a \code{generic_spct} object or a \code{summary_generic_spct}
object.}

\item{instr.settings, value}{a \code{list} or a \code{instr_settings} object.}
}
\value{
x
}
\description{
Function to set by reference the \code{"what.measured"} attribute  of a
\code{generic_spct}, or of a \code{summary_generic_spct} object.
}
\details{
This function alters \code{x} itself by reference and in addition
  returns \code{x} invisibly. If \code{x} is not a \code{generic_spct} object
  or a \code{summary_generic_spct} object, \code{x} is not modified,
  silently. If \code{inst.desc = NULL} is passed in the call, the attribute
  \code{instr.settings} is removed. \emph{This function is very rarely called
  from user code.}
}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
