% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setSensorProperties}
\alias{setSensorProperties}
\alias{sensor_properties<-}
\title{Set the "sensor.properties" attribute}
\usage{
setSensorProperties(x, sensor.properties = NULL, verbose = FALSE)

sensor_properties(x) <- value
}
\arguments{
\item{x}{a response_spct object}

\item{sensor.properties, value}{a list with fields named as described in
\strong{Details}.}

\item{verbose}{logical Flag to enable warning when applied to object of
unsupported class.}
}
\value{
\code{x}
}
\description{
Function to set by reference the "sensor.properties" attribute  of an
existing response_spct object.
}
\details{
Storing sensor properties in an attribute with a well defined
format and minimum of information is intended to help with reproducibility.
Unlike the \code{filter.properties} metadata, the \code{sensor.properties}
metadata are currently not used in any computations. The attribute is
intended to be used with photoelectric sensors, broadband and image, i.e.,
not for biological photoreceptors or photobiological responses.

\describe{
\item{model}{\code{character}, \emph{identification code or name used by the supplier}, required}
\item{type}{\code{character}, \code{"broadband"}, \code{"spectral"}, \code{"image"} required}
\item{supplier}{\code{character}, required}
\item{entrance.optics}{\code{character}, \code{"cosine"}, \code{"dome"}, \code{"sphere"}, \code{"narrow"} required}
\item{channels}{\code{data.frame}, description of the channels, one channel per row, optional}
\item{signal.interface}{\code{character}, \code{"analog"}, \code{"digital"}, and possibly others.}
\item{module.model}{\code{character}, optional}
\item{module.supplier}{\code{character}, optional}
\item{module.interface}{\code{character}, optional}
\item{note}{\code{character}, optional}
}
}
\note{
This function alters \code{x} itself by reference and in addition
  returns \code{x} invisibly. If \code{x} is not a \code{source_spct} object,
  \code{x} is not modified.
}
\examples{

my.spct <- ccd.spct
sensor_properties(my.spct)
sensor_properties(my.spct) <- NULL
sensor_properties(my.spct)
sensor_properties(my.spct) <- list(model = "ccd",
                                   type = "ABC",
                                   supplier = "unknown")
sensor_properties(my.spct)

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
