% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.metadata.r
\name{setWhatMeasured}
\alias{setWhatMeasured}
\alias{what_measured<-}
\alias{setWhatMeasured.default}
\alias{setWhatMeasured.generic_spct}
\alias{setWhatMeasured.summary_generic_spct}
\alias{setWhatMeasured.data.frame}
\alias{setWhatMeasured.generic_mspct}
\title{Set the "what.measured" attribute}
\usage{
setWhatMeasured(x, ...)

what_measured(x) <- value

\method{setWhatMeasured}{default}(x, what.measured, ...)

\method{setWhatMeasured}{generic_spct}(x, what.measured, ...)

\method{setWhatMeasured}{summary_generic_spct}(x, what.measured, ...)

\method{setWhatMeasured}{data.frame}(x, what.measured, ...)

\method{setWhatMeasured}{generic_mspct}(x, what.measured, ...)
}
\arguments{
\item{x}{an R object.}

\item{...}{Allows use of additional arguments in methods for other classes.}

\item{what.measured, value}{a list}
}
\value{
x
}
\description{
Method to set by reference the \code{"what.measured"} attribute of an R
object.
}
\details{
This function alters \code{x} itself by reference and in addition
  returns \code{x} invisibly. If \code{x} does not belong to one of the
  supported classes, \code{x} is not modified.
}
\examples{
my.spct <- sun.spct
what_measured(my.spct)
what_measured(my.spct) <- "Sun"
what_measured(my.spct)
what_measured(my.spct) <- NULL
what_measured(my.spct)

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
