% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul-response.R
\name{simul_sensor_response}
\alias{simul_sensor_response}
\title{Simulate light sensor response}
\usage{
simul_sensor_response(
  source.spct,
  sensor.mspct,
  norm = "skip",
  range = NULL,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  time.unit = NULL,
  scale.factor = 1,
  attr2tb = NULL
)
}
\arguments{
\item{source.spct}{source_spct Light source spectral irradiance.}

\item{sensor.mspct}{response_spct or response_mspct Light sensor
spectral responsiveness.}

\item{norm}{numeric Normalization wavelength (nm) or character string
\code{"max"}, or \code{"min"} for normalization at the corresponding
wavelength, \code{"update"} to update the normalization with previous
criterion, \code{"undo"} to revert an existing normalization or
\code{"skip"} to force return of \code{x} unchanged. See \strong{Details}.}

\item{range}{numeric vector of length two Range of wavelengths to use
(nanometres, nm)}

\item{unit.out}{character Allowed values \code{"energy"}, and
\code{"photon"}, or its alias \code{"quantum"}.}

\item{time.unit}{character or lubridate::duration object.}

\item{scale.factor}{numeric vector of length 1, or length equal to that of
\code{w.band}. Numeric multiplier applied to returned values.}

\item{attr2tb}{character vector, see \code{\link{add_attr2tb}} for the syntax for \code{attr2tb} passed as is to formal parameter \code{col.names}.}
}
\value{
A data frame
}
\description{
Simulate light sensor response
}
\details{
Compute sensor response spectrum by convolution of light source
  emission spectrum or illumination spectrum and the responsiveness spectrum
  of a sensor with one or more channels. Return the integral over
  wavelengths' \code{range} for each sensor channel, or sensor in
  \code{sensor.mspct}.
}
\examples{
simul_sensor_response(sun.spct, ccd.spct)
simul_sensor_response(sun.spct, ccd.spct, range = c(400, 700))
simul_sensor_response(sun.spct, ccd.spct, unit.out = "photon")

}
