% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct-math.r
\name{slash-.generic_spct}
\alias{slash-.generic_spct}
\alias{/.generic_spct}
\title{Arithmetic Operators}
\usage{
\method{/}{generic_spct}(e1, e2)
}
\arguments{
\item{e1}{an object of class "generic_spct"}

\item{e2}{an object of class "generic_spct"}
}
\description{
Division operator for generic spectra.
}
\seealso{
Other math operators and functions:
\code{\link{MathFun}},
\code{\link[=^.generic_spct]{^.generic_spct()}},
\code{\link[=convolve_each]{convolve_each()}},
\code{\link{div-.generic_spct}},
\code{\link{log}},
\code{\link{minus-.generic_spct}},
\code{\link{mod-.generic_spct}},
\code{\link{plus-.generic_spct}},
\code{\link{round}},
\code{\link{sign}},
\code{\link{times-.generic_spct}}
}
\concept{math operators and functions}
