% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.attr.R
\name{spct_attr2tb}
\alias{spct_attr2tb}
\title{Copy attributes into a tibble}
\usage{
spct_attr2tb(
  x,
  which = c("-", "names", "row.names", "spct.tags", "spct.version", "comment"),
  ...
)
}
\arguments{
\item{x}{a generic_spct object.}

\item{which}{character vector Names of attributes to retrieve.}

\item{...}{currently ignored}
}
\value{
A tibble with the values stored in the attributes whose names were
  selected through the argument to \code{which} if present in \code{x}.
}
\description{
Method returning attributes of an object of class generic_spct or derived,
or of class waveband. Only attributes defined and/or set by package
'photobiology' for objects of the corresponding class are returned.
}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
