% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{spct_classes}
\alias{spct_classes}
\title{Function that returns a vector containing the names of spectra classes.}
\usage{
spct_classes()
}
\value{
A \code{character} vector of class names.
}
\description{
Function that returns a vector containing the names of spectra classes.
}
\examples{
spct_classes()

}
