% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.attr.R
\name{subset_attributes}
\alias{subset_attributes}
\alias{subset_attributes.default}
\alias{subset_attributes.generic_spct}
\title{Subset the metadata attributes}
\usage{
subset_attributes(x, to.keep, ...)

\method{subset_attributes}{default}(x, to.keep, ...)

\method{subset_attributes}{generic_spct}(
  x,
  to.keep = character(),
  target.attributes = spct_attributes(class(x)[1]),
  ...
)
}
\arguments{
\item{x}{a generic_spct object.}

\item{to.keep}{character vector Indices to the spectra for
which attributes are to be extracted and retained.}

\item{...}{currently ignored}

\item{target.attributes}{character vector Names of attributes to be
subset using \code{which}.}
}
\value{
Named \code{list} of attribute values.
}
\description{
Method returning attributes of an object of class generic_spct or derived,
or of class waveband. Only attributes defined and/or set by package
'photobiology' for objects of the corresponding class are returned. Parameter
\code{which} can be used to subset each attribute when a generic_spct object
with multiple spectra in long form is subset.
}
\details{
The values in \code{which} are passed as argument to the extract
  operator only if the value of the attribute is a list of the same length
  as the logical vector passed as argument to \code{which}.
}
\seealso{
\code{\link{select_spct_attributes}}

Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=trimInstrDesc]{trimInstrDesc()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
