% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.summary.r, R/spct.summaries.r
\name{summary}
\alias{summary}
\alias{summary.generic_mspct}
\alias{summary.generic_spct}
\title{Summary of a collection of spectra}
\usage{
\method{summary}{generic_mspct}(
  object,
  maxsum = 7,
  digits = max(3, getOption("digits") - 3),
  idx = "spct.idx",
  which.metadata = NULL,
  ...
)

\method{summary}{generic_spct}(
  object,
  maxsum = 7,
  digits = max(3, getOption("digits") - 3),
  ...,
  expand = FALSE
)
}
\arguments{
\item{object}{An object of one of the spectral classes for which a summary is
desired}

\item{maxsum}{integer Indicates how many levels should be shown for factors.}

\item{digits}{integer Used for number formatting with \code{\link{format}()}.}

\item{idx}{character Name of the column with the names of the members of the
collection of spectra.}

\item{which.metadata}{character vector Names of attributes to retrieve, or
"none" or "all".}

\item{...}{additional arguments affecting the summary produced, ignored in
current version}

\item{expand}{logical or character One of "none", "collection" or "each"
indicating that multiple spectra in long form should be converted to a
collection of spectra in advance of summary, summarized as a collection or
individually.}
}
\value{
A summary object matching the class of \code{object}.

A summary object matching the class of \code{object}, a list of
  such objects or a summary object for a matching collection of spectra.
}
\description{
Method of generic function summary for objects of spectral collection
classes.

Methods of generic function summary for objects of spectral classes.
}
\examples{
summary(sun_evening.mspct)
summary(sun_evening.mspct, which.metadata = "when.measured")
summary(two_filters.mspct, which.metadata = "what.measured")

summary(sun.spct)
summary(two_filters.spct)
summary(two_filters.spct, expand = TRUE)
summary(two_filters.spct, expand = "each")
summary(two_filters.spct, expand = "collection")

}
