% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.summaries.r
\name{summary_spct_classes}
\alias{summary_spct_classes}
\title{Function that returns a vector containing the names of spectral summary
classes.}
\usage{
summary_spct_classes()
}
\value{
A \code{character} vector of class names.
}
\description{
Function that returns a vector containing the names of spectral summary
classes.
}
