% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.data.r
\docType{data}
\name{sun.daily.data}
\alias{sun.daily.data}
\title{Daily solar spectral irradiance (simulated)}
\format{A \code{data.frame} object with 511 rows and 3 variables}
\description{
A dataset containing the wavelengths at a 1 nm interval and the corresponding
spectral (energy) irradiance. Values simulated for 2 June 2012, at Helsinki,
under clear sky conditions. The variables are as follows:
}
\details{
\itemize{ \item w.length (nm), range 290 to 800 nm. \item s.e.irrad (J d-1
m-2 nm-1) \item s.q.irrad (mol d-1 m-2 nm-1) }
}
\author{
Anders K. Lindfors (data)
}
\references{
Lindfors, A.; Heikkilä, A.; Kaurola, J.; Koskela, T. & Lakkala,
K. (2009) Reconstruction of Solar Spectral Surface UV Irradiances Using
Radiative Transfer Simulations. Photochemistry and Photobiology, 85:
1233–1239
}
\keyword{datasets}

