% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{time2seconds}
\alias{time2seconds}
\title{time argument check and conversion}
\usage{
time2seconds(t, tz)
}
\value{
numeric representtaion of the date
}
\description{
time argument check and conversion
}
\keyword{internal}

