% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.operators.r
\name{Trig}
\alias{Trig}
\alias{cos.generic_spct}
\alias{sin.generic_spct}
\alias{tan.generic_spct}
\alias{acos.generic_spct}
\alias{asin.generic_spct}
\alias{atan.generic_spct}
\title{Trigonometric Functions}
\usage{
\method{cos}{generic_spct}(x)

\method{sin}{generic_spct}(x)

\method{tan}{generic_spct}(x)

\method{acos}{generic_spct}(x)

\method{asin}{generic_spct}(x)

\method{atan}{generic_spct}(x)
}
\arguments{
\item{x}{an object of class "generic_spct" or a derived class.}
}
\description{
Trigonometric functions for object of \code{generic_spct} and derived
classes.  \\
The functions are applied to the spectral data, not the wavelengths. The
quantity in the spectrum to which the function is applied depends on the class
of \code{x} and the current value of output options.
}
