% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.metadata.r
\name{trimInstrDesc}
\alias{trimInstrDesc}
\title{Trim the "instr.desc" attribute}
\usage{
trimInstrDesc(
  x,
  fields = c("time", "spectrometer.name", "spectrometer.sn", "bench.grating",
    "bench.slit", "entrance.optics")
)
}
\arguments{
\item{x}{a \code{generic_spct} object or a \code{summary_generic_spct}
object.}

\item{fields}{a character vector with the names of the fields to keep,
or if first member is \code{"-"}, the names of fields to delete; \code{"*"}
as the first member of the vector makes the function a no-op, leaving the
spectrum object unaltered.}
}
\value{
\code{x}, possibly with the \code{"instr.desc"} attribute
  modified.
}
\description{
Function to trim the \code{"instr.desc"} attribute of a \code{generic_spct}
or a \code{summary_generic_spct} object, by default discarding all fields
except for \code{spectrometer.name}, \code{spectrometer.sn},
\code{bench.grating}, \code{bench.slit}, and \code{entrance.optics}.
}
\details{
This function alters \code{x} itself by reference and in addition
  returns \code{x} invisibly. If \code{x} is not a \code{generic_spct} object
  or a \code{summary_generic_spct} object, or if the \code{"instr.desc"}
  attribute is not present in a \code{generic_spct} object, \code{x} is not
  modified.

  Attempts to remove or keep fields that are not present in the attribute are
  ignored silently. The value of fields in the attribute is never modified,
  fields are either kept unchanged or removed.
}
\note{
Some of the spectrometer-specific metadata can be large, as they can
  include calibration coefficients. In the case of R package 'ooacquire' also
  pointers to Java objects may need to be deleted.
}
\examples{
my.spct <- white_led.cps_spct
names(instr_descriptor(my.spct))
trimInstrDesc(my.spct) # modified by reference!
names(instr_descriptor(my.spct))

}
\seealso{
Other measurement metadata functions:
\code{\link[=add_attr2tb]{add_attr2tb()}},
\code{\link[=getFilterProperties]{getFilterProperties()}},
\code{\link[=getHowMeasured]{getHowMeasured()}},
\code{\link[=getInstrDesc]{getInstrDesc()}},
\code{\link[=getInstrSettings]{getInstrSettings()}},
\code{\link[=getSensorProperties]{getSensorProperties()}},
\code{\link[=getSoluteProperties]{getSoluteProperties()}},
\code{\link[=getWhatMeasured]{getWhatMeasured()}},
\code{\link[=getWhenMeasured]{getWhenMeasured()}},
\code{\link[=getWhereMeasured]{getWhereMeasured()}},
\code{\link[=get_attributes]{get_attributes()}},
\code{\link[=isValidInstrDesc]{isValidInstrDesc()}},
\code{\link[=isValidInstrSettings]{isValidInstrSettings()}},
\code{\link[=select_spct_attributes]{select_spct_attributes()}},
\code{\link[=setFilterProperties]{setFilterProperties()}},
\code{\link[=setHowMeasured]{setHowMeasured()}},
\code{\link[=setInstrDesc]{setInstrDesc()}},
\code{\link[=setInstrSettings]{setInstrSettings()}},
\code{\link[=setSensorProperties]{setSensorProperties()}},
\code{\link[=setSoluteProperties]{setSoluteProperties()}},
\code{\link[=setWhatMeasured]{setWhatMeasured()}},
\code{\link[=setWhenMeasured]{setWhenMeasured()}},
\code{\link[=setWhereMeasured]{setWhereMeasured()}},
\code{\link[=spct_attr2tb]{spct_attr2tb()}},
\code{\link[=spct_metadata]{spct_metadata()}},
\code{\link[=subset_attributes]{subset_attributes()}},
\code{\link[=trimInstrSettings]{trimInstrSettings()}}
}
\concept{measurement metadata functions}
