% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{twilight2angle}
\alias{twilight2angle}
\title{twilight argument check and conversion}
\usage{
twilight2angle(twilight)
}
\value{
numeric  Solar elevation angle at sunrise or sunset
}
\description{
twilight argument check and conversion
}
\keyword{internal}
