% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.data.r
\docType{data}
\name{two_sensors.mspct}
\alias{two_sensors.mspct}
\alias{two_sensors.spct}
\title{Spectral response of two light sensors.}
\format{
A \code{response_spct} object with 186 rows and 2 variables

An object of class \code{response_spct} (inherits from \code{generic_spct}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 280 rows and 4 columns.
}
\usage{
two_sensors.mspct

two_sensors.spct
}
\description{
A dataset containing a collection of two spectra.
}
\details{
The spectra in \code{\link{photodiode.spct}} and
  \code{\link{ccd.spct}} stored as a collection in a
  \code{\link{response_mspct}} object named \code{response.mspct} with
  members \code{photodiode} and \code{ccd}, and and in long form in a
  \code{link{response_spct}} object named \code{response.mspct} identified
  bit the levels of factor \code{spct.idx}.
}
\examples{
two_sensors.mspct
two_sensors.spct

}
\seealso{
\code{\link{photodiode.spct}} and \code{\link{ccd.spct}}.

Other Spectral data examples:
\code{\link{A.illuminant.spct}},
\code{\link{D50.illuminant.spct}},
\code{\link{D65.illuminant.spct}},
\code{\link{Ler_leaf.spct}},
\code{\link{black_body.spct}},
\code{\link{ccd.spct}},
\code{\link{clear.spct}},
\code{\link{filter_cps.mspct}},
\code{\link{green_leaf.spct}},
\code{\link{phenylalanine.spct}},
\code{\link{photodiode.spct}},
\code{\link{sun.spct}},
\code{\link{sun_daily.spct}},
\code{\link{sun_evening.spct}},
\code{\link{two_filters.spct}},
\code{\link{water.spct}},
\code{\link{white_led.source_spct}}
}
\concept{Spectral data examples}
\keyword{datasets}
