% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.upgrade.r
\name{upgrade_spct}
\alias{upgrade_spct}
\title{Upgrade one spectral object}
\usage{
upgrade_spct(object)
}
\arguments{
\item{object}{generic.spct A single object to upgrade}
}
\value{
The modified object (invisibly).
}
\description{
Update the spectral class names of objects to those used in photobiology (>=
0.6.0) and add 'version' attribute as used in photobiology (>= 0.70).
}
\note{
The object is modified by reference. The class names with ending
  ".spct" replaced by their new equivalents ending in "_spct".
}
\seealso{
Other upgrade from earlier versions:
\code{\link[=is.old_spct]{is.old_spct()}},
\code{\link[=upgrade_spectra]{upgrade_spectra()}}
}
\concept{upgrade from earlier versions}
