% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.operators.r
\name{using_Tfr}
\alias{using_Tfr}
\alias{using_Afr}
\alias{using_A}
\alias{using_energy}
\alias{using_photon}
\alias{using_quantum}
\title{Use photobiology options}
\usage{
using_Tfr(expr)

using_Afr(expr)

using_A(expr)

using_energy(expr)

using_photon(expr)

using_quantum(expr)
}
\arguments{
\item{expr}{an R expression to execute.}
}
\value{
The value returned by the execution of \code{expression}.
}
\description{
Execute an R expression, possibly compound, using a certain setting for
spectral data related options.
}
\references{
Based on \code{withOptions()} as offered by Thomas Lumley, and
  listed in \url{https://www.burns-stat.com/the-options-mechanism-in-r/},
  section Deep End, of "The Options mechanism in R" by Patrick Burns.
}
