% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{validate_geocode}
\alias{validate_geocode}
\title{Validate a geocode}
\usage{
validate_geocode(geocode)
}
\description{
Convert to tibble, check data bounds, convert address to character if
it is not, or add character NAs if the address column is missing.
}
\keyword{internal}
