% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.operators.r
\name{verbose_as_default}
\alias{verbose_as_default}
\alias{strict_range_as_default}
\title{Set error reporting options}
\usage{
verbose_as_default(flag = TRUE)

strict_range_as_default(flag = TRUE)
}
\arguments{
\item{flag}{logical.}
}
\value{
Previous value of the modified option.
}
\description{
Set error reporting related options easily.
}
