% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.tag.r
\name{wb2spct}
\alias{wb2spct}
\title{Create spectrum from wavebands}
\usage{
wb2spct(w.band)
}
\arguments{
\item{w.band}{waveband or list of waveband objects The waveband(s) determine
the wavelengths in variable \code{w.length} of the returned spectrum}
}
\value{
A generic.spectrum object, with columns w.length set to the
  \emph{union} of all  boundaries and hinges defined in the waveband(s).
  Different spectral data variables are set to zero and added making the
  returned value compatible with classes derived from \code{generic_spct}.
}
\description{
Create a generic_spct object with wavelengths from wavebands in a list.
}
\seealso{
Other tagging and related functions:
\code{\link[=is_tagged]{is_tagged()}},
\code{\link[=tag]{tag()}},
\code{\link[=untag]{untag()}},
\code{\link[=wb2rect_spct]{wb2rect_spct()}},
\code{\link[=wb2tagged_spct]{wb2tagged_spct()}}
}
\concept{tagging and related functions}
