% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.operators.r
\name{wb_trim_as_default}
\alias{wb_trim_as_default}
\alias{use_cached_mult_as_default}
\title{Set computation options}
\usage{
wb_trim_as_default(flag = TRUE)

use_cached_mult_as_default(flag = TRUE)
}
\arguments{
\item{flag}{logical.}
}
\value{
Previous value of the modified option.
}
\description{
Set computation related options easily.
}
