% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.data.r
\docType{data}
\name{white_body.spct}
\alias{white_body.spct}
\title{Theoretical white body}
\format{A \code{object_spct} object with 4 rows and 3 variables}
\description{
A dataset for a hypothetical object with transmittance 0/1
  (0\%), reflectance 1/1 (100\%)
}
\details{
\itemize{ \item w.length (nm) \item Tfr (0..1)  \item Rfr (0..1) }
}
\keyword{datasets}

