% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/led.data.R
\docType{data}
\name{white_led.cps_spct}
\alias{white_led.cps_spct}
\title{White led bulb spectrum}
\format{
A \code{data.frame} object with 2068 rows and 2 variables
}
\usage{
white_led.cps_spct
}
\description{
A dataset containing wavelengths and the
  corresponding spectral data as counts per second for an Osram warm white led lamp:
}
\details{
\itemize{ \item w.length (nm), range 188 to 1117 nm. \item cps }
}
\examples{
white_led.cps_spct

}
\seealso{
Other Spectral data examples: 
\code{\link{A.illuminant.spct}},
\code{\link{D65.illuminant.spct}},
\code{\link{Ler_leaf.spct}},
\code{\link{Ler_leaf_rflt.spct}},
\code{\link{Ler_leaf_trns.spct}},
\code{\link{Ler_leaf_trns_i.spct}},
\code{\link{black_body.spct}},
\code{\link{ccd.spct}},
\code{\link{clear.spct}},
\code{\link{clear_body.spct}},
\code{\link{filter_cps.mspct}},
\code{\link{green_leaf.spct}},
\code{\link{opaque.spct}},
\code{\link{phenylalanine.spct}},
\code{\link{photodiode.spct}},
\code{\link{polyester.spct}},
\code{\link{sun.daily.data}},
\code{\link{sun.daily.spct}},
\code{\link{sun.data}},
\code{\link{sun.spct}},
\code{\link{water.spct}},
\code{\link{white_body.spct}},
\code{\link{white_led.raw_spct}},
\code{\link{white_led.source_spct}},
\code{\link{yellow_gel.spct}}
}
\concept{Spectral data examples}
\keyword{datasets}
