% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/led.data.R
\docType{data}
\name{white_led.raw_spct}
\alias{white_led.raw_spct}
\title{White led bulb spectrum}
\format{
An object of class \code{raw_spct} (inherits from \code{generic_spct}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2068 rows and 4 columns.
}
\usage{
white_led.raw_spct
}
\description{
A dataset containing wavelengths and the
  corresponding spectral data as raw instrument counts for an Osram warm white led lamp,
  for three different integration times:
}
\details{
\itemize{ \item w.length (nm), range 188 to 1117 nm. \item counts_1 \item counts_2 \item counts_3 }

\itemize{ \item w.length (nm), range 188 to 1117 nm. \item cps }
}
\examples{
white_led.raw_spct

}
\seealso{
Other Spectral data examples: 
\code{\link{A.illuminant.spct}},
\code{\link{D65.illuminant.spct}},
\code{\link{Ler_leaf.spct}},
\code{\link{Ler_leaf_rflt.spct}},
\code{\link{Ler_leaf_trns.spct}},
\code{\link{Ler_leaf_trns_i.spct}},
\code{\link{black_body.spct}},
\code{\link{ccd.spct}},
\code{\link{clear.spct}},
\code{\link{clear_body.spct}},
\code{\link{filter_cps.mspct}},
\code{\link{green_leaf.spct}},
\code{\link{opaque.spct}},
\code{\link{photodiode.spct}},
\code{\link{polyester.spct}},
\code{\link{sun.daily.data}},
\code{\link{sun.daily.spct}},
\code{\link{sun.data}},
\code{\link{sun.spct}},
\code{\link{white_body.spct}},
\code{\link{white_led.cps_spct}},
\code{\link{white_led.source_spct}},
\code{\link{yellow_gel.spct}}
}
\concept{Spectral data examples}
\keyword{datasets}
