% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.of.R
\docType{data}
\name{wl_colors.spct}
\alias{wl_colors.spct}
\title{Precomputed rgb colors}
\format{
A \code{generic_spct} object with 9001 rows and 3 variables.
}
\usage{
wl_colors.spct
}
\description{
A dataset containing wavelengths at a 0.1 nm interval (100 nm to
  1000 nm) and the corresponding CMF and CC colors for human vision.
}
\details{
\itemize{
\item w.length (nm), numeric vector
\item CMF, named character vector
\item CC, named character vector}
}
\note{
Data computed with function \code{color_of()} and used by function
  \code{fast_color_of_wl()}

  A chroma_spct object with variables as follows:
}
\examples{
wl_colors.spct

}
\keyword{datasets}
