% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.data.r
\docType{data}
\name{yellow_gel.spct}
\alias{yellow_gel.spct}
\title{Transmittance spectrum of yellow theatrical gel.}
\format{A \code{filter_spct} object with 611 rows and 2 variables}
\description{
A dataset containing the wavelengths at a 1 nm interval and
  fractional total transmittance for polyester film.
}
\details{
\itemize{ \item w.length (nm). \item Tfr
(0..1)  }
}
\keyword{datasets}

