% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.operators.R
\name{A2T}
\alias{A2T}
\alias{A2T.default}
\alias{A2T.filter_spct}
\alias{A2T.filter_mspct}
\title{Convert absorbance into transmittance}
\usage{
A2T(x, action, byref, ...)

\method{A2T}{default}(x, action = NULL, byref = FALSE, ...)

\method{A2T}{filter_spct}(x, action = "add", byref = FALSE, ...)

\method{A2T}{filter_mspct}(x, action = "add", byref = FALSE, ...)
}
\arguments{
\item{x}{an R object}

\item{action}{a character string}

\item{byref}{logical indicating if new object will be created by reference or by copy of x}

\item{...}{not used in current version}
}
\description{
Function that coverts absorbance into transmittance (fraction).
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for generic function

\item \code{filter_spct}: Method for filter spectra

\item \code{filter_mspct}: Method for collections of filter spectra
}}

\seealso{
Other quantity conversion functions: \code{\link{T2A}},
  \code{\link{as_energy}}, \code{\link{as_quantum_mol}},
  \code{\link{as_quantum}},
  \code{\link{e2qmol_multipliers}},
  \code{\link{e2quantum_multipliers}}, \code{\link{e2q}},
  \code{\link{q2e}}
}
