% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.new.r
\name{as.generic_spct}
\alias{as.generic_spct}
\alias{as.generic_spct.default}
\title{Coerce to a spectrum}
\usage{
as.generic_spct(x, ...)

\method{as.generic_spct}{default}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{other arguments passed to "set" functions}
}
\value{
A copy of \code{x} converted into a \code{generic_spct} object.
}
\description{
Return a copy of an R object with its class set to a given type of spectrum.
}
\section{Methods (by class)}{
\itemize{
\item \code{as.generic_spct(default)}: 

}}
\seealso{
\code{\link{setGenericSpct}}

Other constructors of spectral objects: 
\code{\link{as.calibration_spct}()},
\code{\link{as.chroma_spct}()},
\code{\link{as.cps_spct}()},
\code{\link{as.filter_spct}()},
\code{\link{as.object_spct}()},
\code{\link{as.raw_spct}()},
\code{\link{as.reflector_spct}()},
\code{\link{as.response_spct}()},
\code{\link{as.solute_spct}()},
\code{\link{as.source_spct}()},
\code{\link{source_spct}()}
}
\concept{constructors of spectral objects}
