% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmspct.classes.R
\name{as.object_mspct}
\alias{as.object_mspct}
\alias{as.object_mspct.default}
\alias{as.object_mspct.data.frame}
\alias{as.object_mspct.object_spct}
\alias{as.object_mspct.list}
\title{Coerce to a collection-of-spectra}
\usage{
as.object_mspct(x, ...)

\method{as.object_mspct}{default}(x, ...)

\method{as.object_mspct}{data.frame}(
  x,
  Tfr.type = c("total", "internal"),
  Rfr.type = c("total", "specular"),
  strict.range = TRUE,
  ...
)

\method{as.object_mspct}{object_spct}(x, ...)

\method{as.object_mspct}{list}(
  x,
  Tfr.type = c("total", "internal"),
  Rfr.type = c("total", "specular"),
  strict.range = TRUE,
  ...,
  ncol = 1,
  byrow = FALSE
)
}
\arguments{
\item{x}{a list of spectral objects or a list of objects such as data frames
that can be converted into spectral objects.}

\item{...}{passed to individual spectrum object constructor}

\item{Tfr.type}{a character string, either "total" or "internal"}

\item{Rfr.type}{a character string, either "total" or "specular"}

\item{strict.range}{logical Flag indicating how off-range values are handled}

\item{ncol}{integer Number of 'virtual' columns in data}

\item{byrow}{logical If \code{ncol > 1} how to read in the data}
}
\value{
A copy of \code{x} converted into a \code{object_mspct} object.
}
\description{
Return a copy of an R object with its class set to a given type
  of spectrum.
}
\section{Methods (by class)}{
\itemize{
\item \code{as.object_mspct(default)}: 

\item \code{as.object_mspct(data.frame)}: 

\item \code{as.object_mspct(object_spct)}: 

\item \code{as.object_mspct(list)}: 

}}
\seealso{
Other Coercion methods for collections of spectra: 
\code{\link{as.calibration_mspct}()},
\code{\link{as.chroma_mspct}()},
\code{\link{as.cps_mspct}()},
\code{\link{as.filter_mspct}()},
\code{\link{as.generic_mspct}()},
\code{\link{as.raw_mspct}()},
\code{\link{as.reflector_mspct}()},
\code{\link{as.response_mspct}()},
\code{\link{as.solute_mspct}()},
\code{\link{as.source_mspct}()},
\code{\link{split2mspct}()},
\code{\link{subset2mspct}()}
}
\concept{Coercion methods for collections of spectra}
