% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmspct.classes.R
\name{as.response_mspct}
\alias{as.response_mspct}
\alias{as.response_mspct.default}
\alias{as.response_mspct.data.frame}
\alias{as.response_mspct.response_spct}
\alias{as.response_mspct.list}
\alias{as.response_mspct.matrix}
\title{Coerce to a collection-of-spectra}
\usage{
as.response_mspct(x, ...)

\method{as.response_mspct}{default}(x, ...)

\method{as.response_mspct}{data.frame}(x, time.unit = "second", ...)

\method{as.response_mspct}{response_spct}(x, ...)

\method{as.response_mspct}{list}(x, time.unit = "second", ..., ncol = 1, byrow = FALSE)

\method{as.response_mspct}{matrix}(
  x,
  w.length,
  spct.data.var = "s.e.response",
  multiplier = 1,
  byrow = NULL,
  spct.names = "spct_",
  ...
)
}
\arguments{
\item{x}{a list of spectral objects or a list of objects such as data frames
that can be converted into spectral objects.}

\item{...}{passed to individual spectrum object constructor}

\item{time.unit}{character A string, "second", "day" or "exposure"}

\item{ncol}{integer Number of 'virtual' columns in data}

\item{byrow}{logical If \code{ncol > 1} how to read in the data}

\item{w.length}{numeric A vector of wavelengthvalues sorted in strictly
ascending order (nm).}

\item{spct.data.var}{character The name of the variable that will contain the
spectral data. This indicates what physical quantity is stored in the
matrix and the units of expression used.}

\item{multiplier}{numeric A multiplier to be applied to the values in
\code{x} to do unit or scale conversion.}

\item{spct.names}{character Vector of names to be assigned to collection
members, either of length 1, or with length equal to the number of spectra.}
}
\value{
A copy of \code{x} converted into a \code{response_mspct} object.
}
\description{
Return a copy of an R object with its class set to a given type
  of spectrum.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{data.frame}: 

\item \code{response_spct}: 

\item \code{list}: 

\item \code{matrix}: 
}}

\note{
When \code{x} is a square matrix an explicit argument is needed for
  \code{byrow} to indicate how data in \code{x} should be read. In every case
  the length of the \code{w.length} vector must match one of the dimensions
  of \code{x}.
}
\seealso{
Other Coercion methods for collections of spectra: 
\code{\link{as.calibration_mspct}()},
\code{\link{as.chroma_mspct}()},
\code{\link{as.cps_mspct}()},
\code{\link{as.filter_mspct}()},
\code{\link{as.generic_mspct}()},
\code{\link{as.object_mspct}()},
\code{\link{as.raw_mspct}()},
\code{\link{as.reflector_mspct}()},
\code{\link{as.solute_mspct}()},
\code{\link{as.source_mspct}()},
\code{\link{split2mspct}()},
\code{\link{subset2mspct}()}
}
\concept{Coercion methods for collections of spectra}
