% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.new.r
\name{as.response_spct}
\alias{as.response_spct}
\alias{as.response_spct.default}
\title{Coerce to a spectrum}
\usage{
as.response_spct(x, ...)

\method{as.response_spct}{default}(x, time.unit = "second", ...)
}
\arguments{
\item{x}{an R object.}

\item{...}{other arguments passed to "set" functions.}

\item{time.unit}{character string indicating the time unit used for spectral
irradiance or exposure (\code{"second"}, \code{"day"} or \code{"exposure"})
or an object of class duration as defined in package lubridate.}
}
\value{
A copy of \code{x} converted into a \code{response_spct} object.
}
\description{
Return a copy of an R object with its class set to a given type of spectrum.
}
\section{Methods (by class)}{
\itemize{
\item \code{as.response_spct(default)}: 

}}
\seealso{
\code{\link{setGenericSpct}}

Other constructors of spectral objects: 
\code{\link{as.calibration_spct}()},
\code{\link{as.chroma_spct}()},
\code{\link{as.cps_spct}()},
\code{\link{as.filter_spct}()},
\code{\link{as.generic_spct}()},
\code{\link{as.object_spct}()},
\code{\link{as.raw_spct}()},
\code{\link{as.reflector_spct}()},
\code{\link{as.solute_spct}()},
\code{\link{as.source_spct}()},
\code{\link{source_spct}()}
}
\concept{constructors of spectral objects}
