% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{as_tod}
\alias{as_tod}
\title{Convert date to time-of-day in hours, minutes or seconds}
\usage{
as_tod(x, unit.out = "hours", tz = NULL)
}
\arguments{
\item{x}{a datetime object accepted by lubridate functions}

\item{unit.out}{charater string, One of "datetime", "hour", "minute", or "second".}

\item{tz}{character string indicating time zone to be used in output.}
}
\description{
Convert date to time-of-day in hours, minutes or seconds
}
