% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beesxyz.data.r
\docType{data}
\name{beesxyzCMF.spct}
\alias{beesxyzCMF.spct}
\title{Honeybee xyz chromaticity colour matching function data}
\format{
A data frame with 81 rows and 4 variables
}
\usage{
beesxyzCMF.spct
}
\description{
A dataset containing wavelengths at a 5 nm interval (300 nm to
  700 nm) and the corresponding x, y, and z chromaticity coordinates.
  Original data from XXX.

  A chroma_spct object with variables as follows:
}
\details{
\itemize{
\item w.length (nm)
\item x
\item y
\item z }
}
\seealso{
Other Visual response data examples: 
\code{\link{ciev10.spct}},
\code{\link{ciev2.spct}},
\code{\link{ciexyzCC10.spct}},
\code{\link{ciexyzCC2.spct}},
\code{\link{ciexyzCMF10.spct}},
\code{\link{ciexyzCMF2.spct}},
\code{\link{cone_fundamentals10.spct}}
}
\concept{Visual response data examples}
\keyword{datasets}
