% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beesxyz.data.r
\docType{data}
\name{beesxyzCMF.spct}
\alias{beesxyzCMF.spct}
\title{Honeybee xyz chromaticity colour matching function data}
\format{A data frame with 81 rows and 4 variables}
\description{
A dataset containing wavelengths at a 5 nm interval (300 nm to
  700 nm) and the corresponding x, y, and z chromaticity coordinates.
  Original data from A chroma_spct object with variables as follows:
}
\details{
\itemize{
\item w.length (nm)
\item x
\item y
\item z }
}
\keyword{datasets}

