% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.data.r
\docType{data}
\name{black_body.spct}
\alias{black_body.spct}
\alias{white_body.spct}
\alias{clear_body.spct}
\title{Theoretical optical bodies}
\format{
A \code{object_spct} object with 4 rows and 3 variables
}
\description{
Datasets for a hypothetical objects with transmittance 0/1
  (0\%), reflectance 0/1 (0\%), with transmittance 0/1
  (0\%), reflectance 1/1 (100\%), and with with transmittance 1/1
  (100\%), reflectance 0/1 (0\%).
}
\details{
\itemize{ \item w.length (nm) \item Tfr (0..1) \item Rfr (0..1) }
}
\seealso{
Other Spectral data examples: 
\code{\link{A.illuminant.spct}},
\code{\link{D65.illuminant.spct}},
\code{\link{Ler_leaf.spct}},
\code{\link{ccd.spct}},
\code{\link{clear.spct}},
\code{\link{filter_cps.mspct}},
\code{\link{green_leaf.spct}},
\code{\link{phenylalanine.spct}},
\code{\link{photodiode.spct}},
\code{\link{sun.spct}},
\code{\link{sun_daily.spct}},
\code{\link{sun_evening.spct}},
\code{\link{two_filters.spct}},
\code{\link{water.spct}},
\code{\link{white_led.source_spct}}
}
\concept{Spectral data examples}
\keyword{datasets}
