% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.data.r
\docType{data}
\name{black_body.spct}
\alias{black_body.spct}
\title{Theoretical black body}
\format{
A \code{object_spct} object with 4 rows and 3 variables
}
\description{
A dataset for a hypothetical object with transmittance 0/1
  (0\%), reflectance 0/1 (0\%)
}
\details{
\itemize{ \item w.length (nm) \item Tfr (0..1) \item Rfr (0..1) }
}
\seealso{
Other Spectral data examples: 
\code{\link{A.illuminant.spct}},
\code{\link{D65.illuminant.spct}},
\code{\link{Ler_leaf.spct}},
\code{\link{Ler_leaf_rflt.spct}},
\code{\link{Ler_leaf_trns.spct}},
\code{\link{Ler_leaf_trns_i.spct}},
\code{\link{ccd.spct}},
\code{\link{clear.spct}},
\code{\link{clear_body.spct}},
\code{\link{filter_cps.mspct}},
\code{\link{green_leaf.spct}},
\code{\link{opaque.spct}},
\code{\link{phenylalanine.spct}},
\code{\link{photodiode.spct}},
\code{\link{polyester.spct}},
\code{\link{sun.daily.data}},
\code{\link{sun.daily.spct}},
\code{\link{sun.data}},
\code{\link{sun.spct}},
\code{\link{water.spct}},
\code{\link{white_body.spct}},
\code{\link{white_led.cps_spct}},
\code{\link{white_led.raw_spct}},
\code{\link{white_led.source_spct}},
\code{\link{yellow_gel.spct}}
}
\concept{Spectral data examples}
\keyword{datasets}
