% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.data.r
\docType{data}
\name{black_body.spct}
\alias{black_body.spct}
\title{Theoretical black body}
\format{A \code{object_spct} object with 4 rows and 3 variables}
\description{
A dataset for a hypothetical object with transmittance 0/1
  (0\%), reflectance 0/1 (0\%)
}
\details{
\itemize{ \item w.length (nm) \item Tfr (0..1) \item Rfr (0..1) }
}
\keyword{datasets}

