% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.data.R
\docType{data}
\name{ccd.spct}
\alias{ccd.spct}
\title{Spectral response of a back-thinned CCD image sensor.}
\format{A \code{response_spct} object with 201 rows and 2 variables}
\description{
A dataset containing wavelengths at a 1 nm interval and
  spectral response as $A W^{-1} nm^{-1}$ for CCD sensor type
  S11071/S10420 from Hamamatsu (measured without a quartz window). These
  arrays are frequently used as sensors in high-UV-sensitivity array
  spectrometers. Data digitized from manufacturer's data sheet.
}
\details{
\itemize{ \item w.length (nm). \item s.e.response
(A/W)  }
}
\references{
Hamamatsu (2014) Datasheet: CCD Image Sensors S11071/S10420-01 Series.
Hamamatsu Photonics KK, Hamamatsu, City.
http://www.hamamatsu.com/resources/pdf/ssd/s11071-1004_etc_kmpd1120e.pdf.
Visited 2016-02-01.
}
\seealso{
Other Spectral data examples: \code{\link{A.illuminant.spct}},
  \code{\link{D65.illuminant.spct}},
  \code{\link{black_body.spct}}, \code{\link{clear.spct}},
  \code{\link{clear_body.spct}},
  \code{\link{filter_cps.mspct}},
  \code{\link{green_leaf.spct}},
  \code{\link{photodiode.spct}},
  \code{\link{polyester.spct}},
  \code{\link{sun.daily.data}},
  \code{\link{sun.daily.spct}}, \code{\link{sun.data}},
  \code{\link{sun.spct}}, \code{\link{white_body.spct}},
  \code{\link{white_led.cps_spct}},
  \code{\link{white_led.raw_spct}},
  \code{\link{white_led.source_spct}},
  \code{\link{yellow_gel.spct}}
}
\keyword{datasets}
