% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciexyz.data.r
\docType{data}
\name{ciev2.spct}
\alias{ciev2.spct}
\title{Linear energy CIE 2008 luminous efficiency function 2 deg data}
\format{
A chroma_spct object with 441 rows and 4 variables
}
\usage{
ciev2.spct
}
\description{
A dataset containing wavelengths at a 1 nm interval (390 nm to
  830 nm) and the corresponding response values for a 2 degrees target.
  Original data from \url{http://www.cvrl.org/} downloaded on 2014-04-29 The
  variables are as follows:
}
\details{
\itemize{
  \item w.length (nm)
  \item x
  \item y
  \item z }
}
\examples{
ciev2.spct

}
\seealso{
Other Visual response data examples: 
\code{\link{beesxyzCMF.spct}},
\code{\link{ciev10.spct}},
\code{\link{ciexyzCC10.spct}},
\code{\link{ciexyzCC2.spct}},
\code{\link{ciexyzCMF10.spct}},
\code{\link{ciexyzCMF2.spct}},
\code{\link{cone_fundamentals10.spct}}
}
\author{
CIE
}
\concept{Visual response data examples}
\keyword{datasets}
