% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciexyz.data.r
\docType{data}
\name{ciexyzCC10.spct}
\alias{ciexyzCC10.spct}
\title{CIE xyz chromaticity coordinates (CC) 10 deg data}
\format{A chroma_spct object with 441 rows and 4 variables}
\usage{
ciexyzCC10.spct
}
\description{
A dataset containing wavelengths at a 1 nm interval (390 nm to
  830 nm) and the corresponding x, y, and z chromaticity coordinates. Derived
  from proposed CIE 2006 standard. Original data from
  \url{http://www.cvrl.org/} downloaded on 2014-04-29 The variables are as
  follows:

\itemize{
  \item w.length (nm)
  \item x
  \item y
  \item z }
}
\author{
CIE
}
\keyword{datasets}

