% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciexyz.data.r
\docType{data}
\name{ciexyzCC2.spct}
\alias{ciexyzCC2.spct}
\title{CIE xyz chromaticity coordinates 2 deg data}
\format{
A chroma_spct object with 441 rows and 4 variables
}
\usage{
ciexyzCC2.spct
}
\description{
A dataset containing wavelengths at a 1 nm interval (390 nm to
  830 nm) and the corresponding x, y, and z chromaticity coordinates.
  According to proposed CIE 2006 standard. Original data from
  \url{http://www.cvrl.org/} downloaded on 2014-04-28 The variables are as
  follows:

\itemize{
  \item w.length (nm)
  \item x
  \item y
  \item z }
}
\examples{
ciexyzCC2.spct

}
\seealso{
Other Visual response data examples: 
\code{\link{beesxyzCMF.spct}},
\code{\link{ciev10.spct}},
\code{\link{ciev2.spct}},
\code{\link{ciexyzCC10.spct}},
\code{\link{ciexyzCMF10.spct}},
\code{\link{ciexyzCMF2.spct}},
\code{\link{cone_fundamentals10.spct}}
}
\author{
CIE
}
\concept{Visual response data examples}
\keyword{datasets}
