% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{class_spct}
\alias{class_spct}
\title{Query which is the class of a spectrum}
\usage{
class_spct(x)
}
\arguments{
\item{x}{any R object}
}
\value{
A character vector containing all matching xxxx.spct S3 classes.
}
\description{
Extract class information from a generic spectrum.
}
\details{
The value returned is equivalent to the set intersection of the
  value returned by \code{class(x)} and the value returned by
  \code{\link{spct_classes}}, but preserving the order of the members of
  the character vector.
}
\examples{
class_spct(sun.spct)
class(sun.spct)

}
