% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.data.r
\docType{data}
\name{clear.spct}
\alias{clear.spct}
\title{Theoretical spectrum of a clear clear material}
\format{A \code{filter_spct} object with 4 rows and 2 variables}
\description{
A dataset for a hypothetical object with transmittance 1/1
  (100\%)
}
\details{
\itemize{ \item w.length (nm). \item Tfr
(0..1)  }
}
\keyword{datasets}

