% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.attr.R
\name{copy_attributes}
\alias{copy_attributes}
\alias{copy_attributes.default}
\alias{copy_attributes.generic_spct}
\alias{copy_attributes.waveband}
\title{Copy attributes from one R object to another}
\usage{
copy_attributes(x, y, which, ...)

\method{copy_attributes}{default}(x, y, which = NULL, ...)

\method{copy_attributes}{generic_spct}(x, y, which = NULL,
  copy.class = FALSE, ...)

\method{copy_attributes}{waveband}(x, y, which = NULL, ...)
}
\arguments{
\item{x, y}{R objects}

\item{which}{character}

\item{...}{not used}

\item{copy.class}{logical If TRUE class attributes are also copied.}
}
\value{
A copy of \code{y} with additional attributes set.
}
\description{
Copy attributes from \code{x} to \code{y}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default for generic function

\item \code{generic_spct}: 

\item \code{waveband}: 
}}

