% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{f_mspct}
\alias{mutate_mspct}
\alias{calc_filter_multipliers}
\alias{T2T}
\alias{getAfrType}
\alias{setAfrType}
\title{Defunct functions and methods}
\usage{
f_mspct(...)

mutate_mspct(...)

calc_filter_multipliers(...)

T2T(...)

getAfrType(...)

setAfrType(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Functions listed here have been removed or deleted, and temporarily replaced
by stubs that report this when they are called.
}
\note{
Function \code{f_mspct()} has been renamed \code{msdply()}.

Function \code{mutate_mspct()} has been renamed \code{msmsply()}.

Function \code{calc_filter_multipliers()} has been removed.

Function \code{calc_filter_multipliers()} has been removed.

Method \code{getAfrType()} has been removed.

Method \code{setAfrType()} has been removed.
}
