% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{enable_check_spct}
\alias{enable_check_spct}
\alias{disable_check_spct}
\title{Enable or disable checks}
\usage{
enable_check_spct()

disable_check_spct()
}
\description{
Choose between protection against errors or faster performance by enabling
(the default) or disabling data-consistency and sanity checks.
}
\seealso{
Other data validity check functions: 
\code{\link{check_spct}()},
\code{\link{check_spectrum}()},
\code{\link{check_w.length}()}
}
\concept{data validity check functions}
