% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.methods.R, R/spct.ratios.R
\name{eq_ratio.source_mspct}
\alias{eq_ratio}
\alias{eq_ratio.default}
\alias{eq_ratio.source_mspct}
\alias{eq_ratio.source_spct}
\title{Energy:photon ratio}
\usage{
\method{eq_ratio}{source_mspct}(spct, w.band = NULL,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.cached.mult = getOption("photobiology.use.cached.mult", default =
  FALSE), use.hinges = getOption("photobiology.use.hinges", default = NULL),
  ..., idx = !is.null(names(spct)))

eq_ratio(spct, w.band, wb.trim, use.cached.mult, use.hinges, ...)

\method{eq_ratio}{default}(spct, w.band, wb.trim, use.cached.mult, use.hinges,
  ...)

\method{eq_ratio}{source_spct}(spct, w.band = NULL,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.cached.mult = getOption("photobiology.use.cached.mult", default =
  FALSE), use.hinges = getOption("photobiology.use.hinges", default = NULL),
  ...)
}
\arguments{
\item{spct}{source_spct}

\item{w.band}{waveband or list of waveband objects}

\item{wb.trim}{logical if TRUE wavebands crossing spectral data boundaries
are trimmed, if FALSE, they are discarded}

\item{use.cached.mult}{logical Flag telling whether multiplier values should
be cached between calls}

\item{use.hinges}{logical Flag telling whether to use hinges to reduce
interpolation errors}

\item{...}{other arguments (possibly ignored)}

\item{idx}{logical whether to add a column with the names of the elements of spct}
}
\value{
a numeric value giving number of Joule per mol of photons for each
  waveband, with name attribute set to the name of each waveband unless a
  named list of wavebands is supplied in which case the names of the list
  elements are used, with "e:q" prepended..
}
\description{
This function returns the energy to mole of photons ratio for each waveband and a
light source spectrum.
}
\section{Methods (by class)}{
\itemize{
\item \code{source_mspct}: Calculates energy:photon from a \code{source_mspct}
object.

\item \code{default}: Default for generic function

\item \code{source_spct}: Method for \code{source_spct} objects
}}
\note{
The last two parameters control speed optimizations. The defaults
  should be suitable in mosts cases. If you will use repeatedly the same SWFs
  on many spectra measured at exactly the same wavelengths you may obtain
  some speed up by setting \code{use.cached.mult=TRUE}. However, be aware
  that you are responsible for ensuring that the wavelengths are the same in
  each call, as the only test done is for the length of the \code{w.length}
  vector.
}
\examples{
eq_ratio(sun.spct, new_waveband(400,700))

}
\seealso{
Other photon and energy ratio functions: \code{\link{e_ratio.source_mspct}},
  \code{\link{energy_ratio}}, \code{\link{photon_ratio}},
  \code{\link{photons_energy_ratio}},
  \code{\link{q_ratio.source_mspct}},
  \code{\link{qe_ratio.source_mspct}}
}

