% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.data.r
\docType{data}
\name{filter_cps.mspct}
\alias{filter_cps.mspct}
\title{Counts per second from a measurement of a plastic film}
\format{A \code{filter_cps.mspct} object with 611 rows and 2 variables}
\description{
A dataset containing the wavelengths at a varying interval and
  counts per second data. Two spectra are included a reference measurement
  of the light source plus a measurement of a plastic film.
}
\details{
A collection of two spectra as a \code{cps_mspct} object, each
 each spectrum being a \code{cps_spct} object with two columns:
 \itemize{ \item w.length (nm). \item cps (counts/s)  }
}
\keyword{datasets}
\keyword{internal}

