% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.spikes.R
\name{find_spikes}
\alias{find_spikes}
\title{Find spikes in vector}
\usage{
find_spikes(
  x,
  x.is.delta = FALSE,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  return.numeric = FALSE,
  na.rm = FALSE,
  max.spike.width = NA
)
}
\arguments{
\item{x}{numeric vector containing the data.}

\item{x.is.delta}{logical Flag indicating whether \code{x} contains
differences or original values.}

\item{height.threshold}{numeric The minimum height of spikes expressed
relative to the median amplitude of the baseline local variation of
\code{x}.}

\item{z.threshold}{numeric Modified local \eqn{Z} values larger than
\code{z.threshold} are detected as boundaries of spikes.}

\item{k}{integer width of median window used for smoothing; must be odd}

\item{spike.direction}{character Controls the direction of spikes to be
detected. Accepted arguments are \code{"up"}, \code{"down"},
\code{"both"}.}

\item{return.numeric}{logical If \code{TRUE} a numeric vector is returned
and otherwise a logical one.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped
before searching for spikes.}

\item{max.spike.width}{integer Sets an upper limit to the width in "pixels"
of the features detected as spikes. \code{NA}, the default, sets no limit.}
}
\value{
An integer vector of the same length as \code{x}. Values that are
  \code{0}, \code{+1} or \code{-1} corresponding to no-spike, upwards-spike,
  and downwards-spike in the data. Conversion to logical with
  \code{as.logical()} results in a vector with \code{TRUE} for spikes and
  \code{FALSE} otherwise.
}
\description{
Find spikes in a numeric vector using the algorithm of
Whitaker and Hayes (2018). Spikes are values in spectra that are unusually
high or low compared to neighbours. They are usually individual values or very
short runs of similar "unusual" values. Spikes caused by cosmic radiation are
a frequent problem in Raman spectra. Another source of spikes are "hot
pixels" in CCD and diode arrays. Other kinds of accidental "outliers" can
be also detected.
}
\section{Spike detection}{

  Spikes are detected based on a modified \eqn{Z} score calculated
  from the differenced spectrum. The \eqn{Z} threshold used should be
  adjusted to the characteristics of the input and desired sensitivity. The
  lower the threshold the more stringent the test becomes, with shorter
  spikes being detected.

  \strong{The algorithms assume a consistent step size for the underlying
  independent variable, e.g., wavelength or time, and should not be applied
  if the data do not fulfil this assumption, at least approximately. As
  \code{find_spkikes()} operates on a single vector, checking this remains
  the responsibility of calling functions or methods such as
  \code{\link{spikes}()} and \code{\link{despike}()}.}

  The algorithm uses running differences to detect abrupt changes in value,
  compared to an estimate of the baseline variation of the differences,
  approximating a baseline \eqn{Z} from MAD and a baseline value from the
  median differences. Currently, a single estimate of MAD is used but running
  medians, when possible, as baseline. This comparison detects running
  differences that are unusually large, in most cases signalling a transition
  between values near the baseline and far from it, in both directions.

  Transitions into- and out of spikes are distinguished based on the median
  of the non-differenced values, as a descriptor of the data baseline. As for
  the median of the differences, a running median is used when possible.

  This function thus detects the start and end of each spike, and
  distinguishes upward and downward spikes.

  \code{k} is the width in number of observations of the window used for
  running median smoothing to extract the baseline. A value several times the
  width of the broader spike but narrow enough to track broader peaks needs
  to be manually set in most cases.

  With \code{na.rm = TRUE}, \code{NA} values are omitted before searching for
  spikes and set to \code{0L} in the returned vector.

  If all spikes are guaranteed to be one observation-wide and either going up
  or down from the baseline, it is possible to detect them based purely on
  the \code{z.threshold} by passing \code{height.threshold = NA} and either
  \code{spike.direction = "up"} or \code{spike.direction = "down"}, which
  ensures very fast computation.

  Parameters of the algorithm need to be adjusted depending on the data, so
  inspection of returned values is needed together with adjustment by trial
  and error of suitable values for \code{z.threshold},
  \code{height.threshold}, and \code{k}.

  The argument to parameter \code{max.spike.width} is used in a final stage
  to discard spikes detected by the algorithms described above but considered
  to be too wide (as a run of successive observations, i.e., detector pixels
  in the case of array spectrometers).
}

\references{
Whitaker, D. A.; Hayes, K. (2018) A simple algorithm for despiking Raman
spectra. Chemometrics and Intelligent Laboratory Systems, 179, 82-84.
\doi{10.1016/j.chemolab.2018.06.009}.
}
\seealso{
Other peaks and valleys functions:
\code{\link[=find_peaks]{find_peaks()}},
\code{\link[=get_peaks]{get_peaks()}},
\code{\link[=peaks]{peaks()}},
\code{\link[=replace_bad_pixs]{replace_bad_pixs()}},
\code{\link[=spikes]{spikes()}},
\code{\link[=valleys]{valleys()}},
\code{\link[=wls_at_target]{wls_at_target()}}
}
\concept{peaks and valleys functions}
