% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{format.tod_time}
\alias{format.tod_time}
\title{Encode in a Common Format}
\usage{
\method{format}{tod_time}(x, ..., sep = ":")
}
\arguments{
\item{x}{an R object}

\item{...}{ignored}

\item{sep}{character used as separator}
}
\description{
Format a \code{tod_time} object for pretty printing
}
\seealso{
Other Time of day functions: 
\code{\link{as_tod}()},
\code{\link{print.tod_time}()}
}
\concept{Time of day functions}
