% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.fshift.r
\name{fshift_spct}
\alias{fshift_spct}
\title{fshift a spectrum}
\usage{
fshift_spct(spct, range, col.names, f, ...)
}
\arguments{
\item{spct}{generic_spct The spectrum to be normalized}

\item{range}{an R object on which range() returns a vector of length 2, with
min and max wavelengths (nm)}

\item{col.names}{character The name of the variable to shift with respect to
zero.}

\item{f}{function A summary function to be applied to \code{spct}}

\item{...}{other arguments passed to f()}
}
\value{
a new object of the same class as \code{spct}.
}
\description{
This function returns a spectral object of the same class as the one
supplied as argument but with the spectral data expressed on a zero-shifted
scale.
}
\details{
This private function is used internally to implement the
  \code{hshift()} methods that are exported.
}
\keyword{internal}
