% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.peaks.r
\name{get_peaks}
\alias{get_peaks}
\alias{get_valleys}
\title{Get peaks and valleys in a spectrum}
\usage{
get_peaks(
  x,
  y,
  global.threshold = 0,
  local.threshold = 0,
  local.reference = "minimum",
  threshold.range = NULL,
  span = 5,
  strict = TRUE,
  x_unit = "",
  x_digits = 3,
  na.rm = FALSE
)

get_valleys(
  x,
  y,
  global.threshold = 0,
  local.threshold = 0,
  local.reference = "minimum",
  threshold.range = NULL,
  span = 5,
  strict = TRUE,
  x_unit = "",
  x_digits = 3,
  na.rm = FALSE
)
}
\arguments{
\item{x, y}{numeric}

\item{global.threshold}{numeric A value between 0.0 and 1.0,
relative to \code{threshold.range} indicating the \emph{global} height
(depth) threshold below which peaks (valleys) will be ignored, or a
negative value, between 0.0 and -1.0 indicating the \emph{global} height
(depth) threshold above which peaks (valleys) will be ignored. If
\code{threshold.range = 0} or the value passed as argument belongs to class
\code{"AsIs"} the value is interpreted as an absolute value expressed in
data units.}

\item{local.threshold}{numeric A value between 0.0 and 1.0, relative to
\code{threshold.range}, indicating the
\emph{within-window} height (depth) threshold below which peaks (valleys)
will be ignored.  If \code{threshold.range = 0} or the value passed
as argument belongs to class \code{"AsIs"} the value is interpreted as an
absolute value expressed in data units.}

\item{local.reference}{character One of \code{"minimum"} (eqv.
\code{"maximum"}) or \code{"median"}. The reference used to assess the
height of the peak, either the minimum (maximum) value within the window or
the median of all values in the window.}

\item{threshold.range}{numeric vector of length 2 or a longer vector or list
on which a call to \code{range()} returns a numeric vector of length 2. If
\code{NULL}, the default, \code{range(x)} is used.}

\item{span}{odd integer A peak is defined as an element in a sequence which
is greater than all other elements within a moving window of width
\code{span} centred at that element. The default value is 5, meaning that a
peak is taller than its four nearest neighbours. \code{span = NULL} extends
the span to the whole length of \code{x}.}

\item{strict}{logical flag: if TRUE, an element must be strictly greater than
all other values in its window to be considered a peak. Default: TRUE.}

\item{x_unit}{character Vector of texts to be pasted at end of labels built
from x value at peaks.}

\item{x_digits}{numeric Number of significant digits in wavelength label.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped
before searching for peaks.}
}
\value{
A data frame with variables w.length and s.irrad with their values at
  the peaks or valleys plus a character variable of labels.
}
\description{
These functions find peaks (local maxima) or valleys (local minima) in a
spectrum, using a user selectable size threshold relative to the tallest peak
(global maximum). This a wrapper built on top of function peaks from package
splus2R.
}
\examples{
with(sun.spct, get_peaks(w.length, s.e.irrad))
with(sun.spct, get_valleys(w.length, s.e.irrad))

}
\seealso{
Other peaks and valleys functions: 
\code{\link{find_peaks}()},
\code{\link{find_spikes}()},
\code{\link{peaks}()},
\code{\link{replace_bad_pixs}()},
\code{\link{spikes}()},
\code{\link{valleys}()},
\code{\link{wls_at_target}()}
}
\concept{peaks and valleys functions}
