% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getIdFactor}
\alias{getIdFactor}
\title{Get the "idfactor" attribute}
\usage{
getIdFactor(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
character
}
\description{
Function to read the "idfactor" attribute of an existing generic_spct.
}
\note{
If x is not a \code{generic_spct} or an object of a derived class
  \code{NA} is returned.
}
\examples{
getMultipleWl(sun.spct)

}
\seealso{
Other idfactor attribute functions: 
\code{\link{setIdFactor}()}
}
\concept{idfactor attribute functions}
