% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getInstrDesc}
\alias{getInstrDesc}
\title{Get the "instr.desc" attribute}
\usage{
getInstrDesc(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
list (depends on instrument type)
}
\description{
Function to read the "instr.desc" attribute of an existing generic_spct
object.
}
\seealso{
Other measurement metadata functions: \code{\link{getInstrSettings}},
  \code{\link{getWhatMeasured}},
  \code{\link{getWhenMeasured}},
  \code{\link{getWhereMeasured}},
  \code{\link{get_attributes}},
  \code{\link{isValidInstrDesc}},
  \code{\link{isValidInstrSettings}},
  \code{\link{setInstrDesc}},
  \code{\link{setInstrSettings}},
  \code{\link{setWhatMeasured}},
  \code{\link{setWhenMeasured}},
  \code{\link{setWhereMeasured}},
  \code{\link{trimInstrDesc}},
  \code{\link{trimInstrSettings}}
}
