% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.metadata.r
\name{getInstrSettings}
\alias{getInstrSettings}
\alias{instr_settings}
\title{Get the "instr.settings" attribute}
\usage{
getInstrSettings(x)

instr_settings(x)
}
\arguments{
\item{x}{a \code{generic_spct} object or a \code{summary_generic_spct}
object.}
}
\value{
an object of class \code{"instr_settings"} derived from \code{"list"}.
}
\description{
Function to extract the \code{"instr.settings"} attribute from
\code{generic_spct} object or from a \code{summary_generic_spct}.
}
\details{
If \code{x} is derived from \code{generic_spct} or from
\code{summary_generic_spct}, the value of attribute \code{"instr.settings"}
is returned (\code{NULL}, if missing). Otherwise \code{list()} is returned.
}
\examples{
settings <- getInstrSettings(white_led.cps_spct)
class(settings)
print(settings)
print(str(settings))

}
\seealso{
Other measurement metadata functions: 
\code{\link{add_attr2tb}()},
\code{\link{getFilterProperties}()},
\code{\link{getHowMeasured}()},
\code{\link{getInstrDesc}()},
\code{\link{getSensorProperties}()},
\code{\link{getSoluteProperties}()},
\code{\link{getWhatMeasured}()},
\code{\link{getWhenMeasured}()},
\code{\link{getWhereMeasured}()},
\code{\link{get_attributes}()},
\code{\link{isValidInstrDesc}()},
\code{\link{isValidInstrSettings}()},
\code{\link{select_spct_attributes}()},
\code{\link{setFilterProperties}()},
\code{\link{setHowMeasured}()},
\code{\link{setInstrDesc}()},
\code{\link{setInstrSettings}()},
\code{\link{setSensorProperties}()},
\code{\link{setSoluteProperties}()},
\code{\link{setWhatMeasured}()},
\code{\link{setWhenMeasured}()},
\code{\link{setWhereMeasured}()},
\code{\link{spct_attr2tb}()},
\code{\link{spct_metadata}()},
\code{\link{subset_attributes}()},
\code{\link{trimInstrDesc}()},
\code{\link{trimInstrSettings}()}
}
\concept{measurement metadata functions}
