% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getInstrSettings}
\alias{getInstrSettings}
\title{Get the "instr.settings" attribute}
\usage{
getInstrSettings(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
list
}
\description{
Function to read the "instr.settings" attribute of an existing generic_spct
object.
}
\seealso{
Other measurement metadata functions: 
\code{\link{getHowMeasured}()},
\code{\link{getInstrDesc}()},
\code{\link{getWhatMeasured}()},
\code{\link{getWhenMeasured}()},
\code{\link{getWhereMeasured}()},
\code{\link{get_attributes}()},
\code{\link{isValidInstrDesc}()},
\code{\link{isValidInstrSettings}()},
\code{\link{select_spct_attributes}()},
\code{\link{setHowMeasured}()},
\code{\link{setInstrDesc}()},
\code{\link{setInstrSettings}()},
\code{\link{setWhatMeasured}()},
\code{\link{setWhenMeasured}()},
\code{\link{setWhereMeasured}()},
\code{\link{spct_attr2tb}()},
\code{\link{trimInstrDesc}()},
\code{\link{trimInstrSettings}()}
}
\concept{measurement metadata functions}
