% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getInstrSettings}
\alias{getInstrSettings}
\title{Get the "instr.settings" attribute}
\usage{
getInstrSettings(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
list (depends on instrument type)
}
\description{
Function to read the "instr.desc" attribute of an existing generic_spct
object.
}
\seealso{
Other measurement metadata functions: \code{\link{getInstrDesc}},
  \code{\link{getWhenMeasured.generic_mspct}},
  \code{\link{getWhereMeasured.generic_mspct}},
  \code{\link{setInstrDesc}},
  \code{\link{setInstrSettings}},
  \code{\link{setWhenMeasured.generic_mspct}},
  \code{\link{setWhereMeasured.generic_mspct}}
}

