% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getMultipleWl}
\alias{getMultipleWl}
\title{Get the "multiple.wl" attribute}
\usage{
getMultipleWl(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
integer
}
\description{
Function to read the "multiple.wl" attribute of an existing generic_spct.
}
\note{
If x is not a \code{generic_spct} or an object of a derived class
  \code{NA} is returned.
}
\examples{
getMultipleWl(sun.spct)

}
\seealso{
Other multiple.wl attribute functions: 
\code{\link{setMultipleWl}()}
}
\concept{multiple.wl attribute functions}
