% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.normalize.R
\name{getNormalized}
\alias{getNormalized}
\title{Get the "normalized" attribute}
\usage{
getNormalized(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
character or numeric or logical
}
\description{
Funtion to read the "normalized" attribute of an existing generic_spct
object.
}
\note{
if x is not a \code{generic_spct} object, \code{NA} is returned
}
\seealso{
Other rescaling functions: \code{\link{fscale}},
  \code{\link{fshift}}, \code{\link{is_normalized}},
  \code{\link{is_scaled}}, \code{\link{normalize}},
  \code{\link{setNormalized}}, \code{\link{setScaled}}
}
