% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.fscale.r
\name{getScaled}
\alias{getScaled}
\title{Get the "scaled" attribute}
\usage{
getScaled(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
logical
}
\description{
Function to read the "scaled" attribute of an existing generic_spct
object.
}
\note{
if x is not a \code{filter_spct} object, \code{NA} is returned
}
\seealso{
Other Rfr attribute functions: \code{\link{getRfrType}},
  \code{\link{setRfrType}}
}
\concept{Rfr attribute functions}
