% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.fscale.r
\name{getScaled}
\alias{getScaled}
\title{Get the "scaled" attribute}
\usage{
getScaled(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
logical
}
\description{
Function to read the "scaled" attribute of an existing generic_spct
object.
}
\note{
if x is not a \code{filter_spct} object, \code{NA} is returned
}
\examples{

scaled.spct <- fscale(sun.spct)
getScaled(scaled.spct)

}
\seealso{
Other rescaling functions: 
\code{\link{fscale}()},
\code{\link{fshift}()},
\code{\link{getNormalized}()},
\code{\link{is_normalized}()},
\code{\link{is_scaled}()},
\code{\link{normalize}()},
\code{\link{setNormalized}()},
\code{\link{setScaled}()}
}
\concept{rescaling functions}
