% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{getSensorProperties}
\alias{getSensorProperties}
\alias{sensor_properties}
\alias{getSensorProperties.default}
\alias{getSensorProperties.response_spct}
\alias{getSensorProperties.summary_response_spct}
\alias{getSensorProperties.generic_mspct}
\title{Get the "sensor.properties" attribute}
\usage{
getSensorProperties(x, ...)

sensor_properties(x, ...)

\method{getSensorProperties}{default}(x, return.null = FALSE, ...)

\method{getSensorProperties}{response_spct}(x, return.null = FALSE, ...)

\method{getSensorProperties}{summary_response_spct}(x, return.null = FALSE, ...)

\method{getSensorProperties}{generic_mspct}(x, ..., idx = getIdFactor(x))
}
\arguments{
\item{x}{a source_spct object}

\item{...}{Allows use of additional arguments in methods for other classes.}

\item{return.null}{logical If true, \code{NULL} is returned if the attribute
is not set, otherwise the expected list is returned with all fields set to
\code{NA}.}

\item{idx}{character Name of the column with the names of the members of the
collection of spectra.}
}
\value{
an object of class \code{sensor_properties} derived from \code{list},
  guaranteed to contain fields named \code{"sensor.name"},
  \code{"sensor.type"} and \code{"num.channels"}. If the
  attribute is not set, these fields are set to \code{NA} in the returned
  object.
}
\description{
Function to read the "sensor.properties" attribute of an existing source_spct
or a source_mspct.
}
\note{
The method for collections of spectra returns a tibble with a
  column of lists.
}
\examples{
sensor_properties(ccd.spct)

}
\seealso{
Other measurement metadata functions: 
\code{\link{add_attr2tb}()},
\code{\link{getFilterProperties}()},
\code{\link{getHowMeasured}()},
\code{\link{getInstrDesc}()},
\code{\link{getInstrSettings}()},
\code{\link{getSoluteProperties}()},
\code{\link{getWhatMeasured}()},
\code{\link{getWhenMeasured}()},
\code{\link{getWhereMeasured}()},
\code{\link{get_attributes}()},
\code{\link{isValidInstrDesc}()},
\code{\link{isValidInstrSettings}()},
\code{\link{select_spct_attributes}()},
\code{\link{setFilterProperties}()},
\code{\link{setHowMeasured}()},
\code{\link{setInstrDesc}()},
\code{\link{setInstrSettings}()},
\code{\link{setSensorProperties}()},
\code{\link{setSoluteProperties}()},
\code{\link{setWhatMeasured}()},
\code{\link{setWhenMeasured}()},
\code{\link{setWhereMeasured}()},
\code{\link{spct_attr2tb}()},
\code{\link{spct_metadata}()},
\code{\link{subset_attributes}()},
\code{\link{trimInstrDesc}()},
\code{\link{trimInstrSettings}()}
}
\concept{measurement metadata functions}
