% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.xy.r
\name{integrate_xy}
\alias{integrate_xy}
\title{Gives irradiance from spectral irradiance.}
\usage{
integrate_xy(x, y)
}
\arguments{
\item{x}{numeric array}

\item{y}{numeric array}
}
\value{
a single numeric value with no change in scale factor: e.g. [W m-2
  nm-1] -> [W m-2]
}
\description{
This function gives the result of integrating spectral irradiance over
wavelengths.
}
\examples{
with(sun.data, integrate_xy(w.length, s.e.irrad))

}
