% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irradiance.r
\name{irradiance}
\alias{irradiance}
\title{Photon (quantum) irradiance from spectral (energy) irradiance.}
\usage{
irradiance(w.length, s.irrad, w.band = NULL, unit.out = NULL,
  unit.in = "energy", check.spectrum = TRUE,
  use.cached.mult = getOption("photobiology.use.cached.mult", default =
  FALSE), use.hinges = getOption("photobiology.use.hinges", default = NULL))
}
\arguments{
\item{w.length}{numeric Vector of wavelength (nm)}

\item{s.irrad}{numeric Corresponding vector of spectral (energy) irradiances
(W m-2 nm-1)}

\item{w.band}{waveband or list of waveband objects The waveband(s) determine
the region(s) of the spectrum that are summarized.}

\item{unit.out}{character Allowed values "energy", and "photon", or its alias
"quantum"}

\item{unit.in}{character Allowed values "energy", and "photon", or its alias
"quantum"}

\item{check.spectrum}{logical Flag indicating whether to sanity check input
data, default is TRUE}

\item{use.cached.mult}{logical Flag indicating whether multiplier values
should be cached between calls}

\item{use.hinges}{logical Flag indicating whether to use hinges to reduce
interpolation errors}
}
\value{
a single numeric value with no change in scale factor: [W m-2 nm-1]
  -> [mol s-1 m-2]
}
\description{
This function returns the energy irradiance for a given waveband of a
radiation spectrum.
}
\note{
The last three parameters control speed optimizations. The defaults
  should be suitable in mosts cases. If you set \code{check.spectrum=FALSE}
  then you should call \code{check_spectrum()} at least once for your
  spectrum before using any of the other functions. If you will use
  repeatedly the same SWFs on many spectra measured at exactly the same
  wavelengths you may obtain some speed up by setting
  \code{use.cached.mult=TRUE}. However, be aware that you are responsible for
  ensuring that the wavelengths are the same in each call, as the only test
  done is for the length of the \code{w.length} vector. The is no reason for
  setting \code{use.cpp.code=FALSE} other than for testing the improvement in
  speed, or in cases where there is no suitable C++ compiler for building the
  package.
}
\examples{
data(sun.data)
with(sun.data, irradiance(w.length, s.e.irrad, new_waveband(400,700), "photon"))
}
\seealso{
Other irradiance functions: \code{\link{e_fluence.source_mspct}},
  \code{\link{e_irrad.source_mspct}},
  \code{\link{energy_irradiance}},
  \code{\link{fluence.source_mspct}},
  \code{\link{irrad.source_mspct}},
  \code{\link{photon_irradiance}},
  \code{\link{q_fluence.source_mspct}},
  \code{\link{q_irrad.source_mspct}}
}

