% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{is.solar_time}
\alias{is.solar_time}
\alias{is.solar_date}
\title{Query class}
\usage{
is.solar_time(x)

is.solar_date(x)
}
\arguments{
\item{x}{an R object.}
}
\description{
Query class
}
\seealso{
Other astronomy related functions: 
\code{\link{day_night}()},
\code{\link{format.solar_time}()},
\code{\link{format.tod_time}()},
\code{\link{print.solar_time}()},
\code{\link{print.tod_time}()},
\code{\link{solar_time}()},
\code{\link{sun_angles}()}
}
\concept{astronomy related functions}
