% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{is_mole_based}
\alias{is_mole_based}
\alias{is_mass_based}
\title{Query if a spectrum contains mole or mass based data}
\usage{
is_mole_based(x)

is_mass_based(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{is_mole_based} returns TRUE if its argument is a
  \code{solute_spct} object that contains spectral \code{K.mole} data and
  \code{FALSE} if it contains \code{K.mass} data, but returns NA for any
  other R object, including those belonging other \code{generic_spct}-derived
  classes. \code{is_mass_based} returns the complement of
  \code{is_mole_based}.
}
\description{
Functions to check if an solute attenuation spectrum contains coefficients
on expressed on mole of mass base.
}
\examples{
print("missing example")

}
\seealso{
Other query units functions: 
\code{\link{is_absorbance_based}()},
\code{\link{is_photon_based}()}
}
\concept{query units functions}
