% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.normalize.r
\name{is_normalized}
\alias{is_normalized}
\title{Query whether a generic spectrum has been normalized.}
\usage{
is_normalized(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A \code{logical} value. If \code{x} is not normalized or \code{x} is
  not a \code{generic_spct} object the value returned is \code{FALSE}.
}
\description{
This function tests a \code{generic_spct} object for an attribute that
signals whether the spectral data has been normalized or not after the object
was created.
}
\seealso{
Other rescaling functions: \code{\link{fscale}},
  \code{\link{fshift}}, \code{\link{getNormalized}},
  \code{\link{is_scaled}}, \code{\link{normalize}},
  \code{\link{setNormalized}}, \code{\link{setScaled}}
}

