% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{isValidInstrDesc}
\alias{isValidInstrDesc}
\title{Check the "instr.desc" attribute}
\usage{
isValidInstrDesc(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
logical TRUE if at least instrument name and serial number is found.
}
\description{
Function to validate the "instr.settings" attribute of an existing generic_spct
object.
}
\seealso{
Other measurement metadata functions: \code{\link{getInstrDesc}},
  \code{\link{getInstrSettings}},
  \code{\link{getWhatMeasured}},
  \code{\link{getWhenMeasured}},
  \code{\link{getWhereMeasured}},
  \code{\link{isValidInstrSettings}},
  \code{\link{setInstrDesc}},
  \code{\link{setInstrSettings}},
  \code{\link{setWhatMeasured}},
  \code{\link{setWhenMeasured}},
  \code{\link{setWhereMeasured}},
  \code{\link{trimInstrDesc}},
  \code{\link{trimInstrSettings}}
}
