% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{isValidInstrSettings}
\alias{isValidInstrSettings}
\title{Check the "instr.settings" attribute}
\usage{
isValidInstrSettings(x)
}
\arguments{
\item{x}{a generic_spct object}
}
\value{
logical TRUE if at least integration time data is found.
}
\description{
Function to validate the "instr.settings" attribute of an existing generic_spct
object.
}
\seealso{
Other measurement metadata functions: \code{\link{getInstrDesc}},
  \code{\link{getInstrSettings}},
  \code{\link{getWhatMeasured}},
  \code{\link{getWhenMeasured}},
  \code{\link{getWhereMeasured}},
  \code{\link{get_attributes}},
  \code{\link{isValidInstrDesc}},
  \code{\link{setInstrDesc}},
  \code{\link{setInstrSettings}},
  \code{\link{setWhatMeasured}},
  \code{\link{setWhenMeasured}},
  \code{\link{setWhereMeasured}},
  \code{\link{trimInstrDesc}},
  \code{\link{trimInstrSettings}}
}
\concept{measurement metadata functions}
