% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.hinges.R
\name{l_insert_hinges}
\alias{l_insert_hinges}
\title{Insert wavelength values into spectral data.}
\usage{
l_insert_hinges(x, y, h)
}
\arguments{
\item{x}{numeric array (sorted in increasing order)}

\item{y}{numeric array}

\item{h}{a numeric array giving the wavelengths at which the y values
should be inserted by interpolation, no interpolation is indicated by an
empty array (numeric(0))}
}
\value{
a list with variables \code{x} and \code{y}
}
\description{
Inserting wavelengths values immediately before and after a discontinuity in
the SWF, greatly reduces the errors caused by interpolating the weighted
irradiance during integration of the effective spectral irradiance. This is
specially true when data have a large wavelength step size.
}
\keyword{internal}
