% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{labels}
\alias{labels}
\alias{labels.waveband}
\alias{labels.generic_spct}
\title{Find labels from "waveband" object}
\usage{
\method{labels}{waveband}(object, ...)

\method{labels}{generic_spct}(object, ...)
}
\arguments{
\item{object}{an object of class "waveband"}

\item{...}{not used in current version}
}
\description{
A function to obtain the name and label of objects of class "waveband".
}
\section{Methods (by class)}{
\itemize{
\item \code{generic_spct}: 
}}

\examples{
labels(sun.spct)

}
\seealso{
Other waveband attributes: 
\code{\link{is_effective}()},
\code{\link{normalization}()}
}
\concept{waveband attributes}
