% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{wl_max}
\alias{wl_max}
\alias{max}
\alias{max.waveband}
\alias{max.generic_spct}
\alias{max.generic_mspct}
\title{Wavelength maximum}
\usage{
wl_max(x, na.rm = FALSE)

\method{max}{waveband}(..., na.rm = FALSE)

\method{max}{generic_spct}(..., na.rm = FALSE)

\method{max}{generic_mspct}(..., na.rm = FALSE, idx = NULL)
}
\arguments{
\item{x}{generic_spct, generic_mspct or waveband object.}

\item{na.rm}{ignored}

\item{...}{not used in current version}

\item{idx}{logical whether to add a column with the names of the elements of
spct}
}
\description{
A method specialization that returns the wavelength maximum from objects of
classes "waveband" or of class "generic_spct" or derived.
}
\section{Methods (by class)}{
\itemize{
\item \code{generic_spct}: 

\item \code{generic_mspct}: 
}}

\examples{
max(sun.spct)
wl_max(sun.spct)

}
