% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{max}
\alias{max}
\alias{max.waveband}
\alias{max.generic_spct}
\alias{max.generic_mspct}
\title{Wavelength maximum}
\usage{
\method{max}{waveband}(..., na.rm = FALSE)

\method{max}{generic_spct}(..., na.rm = FALSE)

\method{max}{generic_mspct}(..., na.rm = FALSE, idx = NULL)
}
\arguments{
\item{...}{not used in current version}

\item{na.rm}{ignored}

\item{idx}{logical whether to add a column with the names of the elements of
spct}
}
\description{
A function that returns the wavelength maximum from objects of class "waveband".
}
\section{Methods (by class)}{
\itemize{
\item \code{generic_spct}: 

\item \code{generic_mspct}: 
}}

\examples{
max(sun.spct)

}
