% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.attr.R
\name{merge_attributes}
\alias{merge_attributes}
\alias{merge_attributes.default}
\alias{merge_attributes.generic_spct}
\title{Merge and copy attributes}
\usage{
merge_attributes(x, y, z, which, ...)

\method{merge_attributes}{default}(x, y, z, which = NULL, ...)

\method{merge_attributes}{generic_spct}(x, y, z, which = NULL,
  copy.class = FALSE, ...)
}
\arguments{
\item{x, y, z}{R objects}

\item{which}{character}

\item{...}{not used}

\item{copy.class}{logical If TRUE class attributes are also copied.}
}
\value{
A copy of \code{z} with additional attributes set.
}
\description{
Merge attributes from \code{x} and \code{y} and copy them to \code{z}.
Methods defined for spectral objects of classes from package 'photobiology'.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default for generic function

\item \code{generic_spct}: 
}}

