% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{min}
\alias{min}
\alias{min.waveband}
\alias{min.generic_spct}
\alias{min.generic_mspct}
\title{Wavelength minimum}
\usage{
\method{min}{waveband}(..., na.rm = FALSE)

\method{min}{generic_spct}(..., na.rm = FALSE)

\method{min}{generic_mspct}(..., na.rm = FALSE, idx = NULL)
}
\arguments{
\item{...}{not used in current version}

\item{na.rm}{ignored}

\item{idx}{logical whether to add a column with the names of the elements of
spct}
}
\description{
A function that returns the wavelength minimum.
}
\section{Methods (by class)}{
\itemize{
\item \code{generic_spct}: 

\item \code{generic_mspct}: 
}}

\examples{
min(sun.spct)

}
\seealso{
Other wavelength summaries: \code{\link{midpoint}},
  \code{\link{range}}, \code{\link{stepsize}}
}
