% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.et.r
\name{net_lw_radiation}
\alias{net_lw_radiation}
\title{Net long wave radiation}
\usage{
net_lw_radiation(
  temperature,
  lw.down.irradiance = NULL,
  water.vp = 0,
  emissivity = 0.98
)
}
\arguments{
\item{temperature}{numeric vector of air temperatures (C) at 2 m height.}

\item{lw.down.irradiance}{numeric Long wave downwelling radiation (W/m2)}

\item{water.vp}{numeric vector of water vapour pressure in air (Pa),
ignored if \code{lw.down.irradiance} is available.}

\item{emissivity}{numeric Emissivity of the surface (ground or vegetation)
for long wave radiation.}
}
\value{
A numeric vector of $R_{nl}$ estimates expressed as W / m-2.
}
\description{
Helper function returning the computed long-wave radiation balance if
\code{lw.down.irradiance} is passed a value in W / m2 and if not
it returns an estimate of the net lw radiation uncorrected for cloudiness.
This approach is used in the FAO56 approach.
}
\seealso{
Other Evapotranspiration and energy balance related functions.: 
\code{\link{ET_ref}()},
\code{\link{ET_zero}()}
}
\concept{Evapotranspiration and energy balance related functions.}
