% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{normalization}
\alias{normalization}
\alias{normalization.default}
\alias{normalization.waveband}
\title{Normalization of an R object}
\usage{
normalization(x)

\method{normalization}{default}(x)

\method{normalization}{waveband}(x)
}
\arguments{
\item{x}{an R object}
}
\value{
A single numeric value of wavelength [\eqn{nm}].
}
\description{
Normalization wavelength [\eqn{nm}] of an R object, retrieved from the object's
attributes.
}
\section{Methods (by class)}{
\itemize{
\item \code{normalization(default)}: Default methods.

\item \code{normalization(waveband)}: Normalization of a \code{\link{waveband}} object.

}}
\seealso{
Other waveband attributes: 
\code{\link{is_effective}()},
\code{\link{labels}()}
}
\concept{waveband attributes}
