% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photobiology.r
\docType{package}
\name{photobiology-package}
\alias{photobiology}
\alias{photobiology-package}
\title{photobiology: Photobiological Calculations}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Definitions of classes, methods, operators and functions for use 
    in photobiology and radiation meteorology and climatology. Calculation of
    effective (weighted) and not-weighted irradiances/doses, fluence rates,
    transmittance, reflectance, absorptance, absorbance and diverse ratios and 
    other derived quantities from spectral data. Local maxima and minima: peaks,
    valleys and spikes. Conversion between energy-and photon-based units. 
    Wavelength interpolation. Astronomical calculations related solar angles and 
    day length. Colours and vision. This package is part of the 'r4photobiology' 
    suite, Aphalo, P. J. (2015) <doi:10.19232/uv4pb.2015.1.14>.
}
\details{
Package 'photobiology' is at the core of a suite of packages for
  analysis and plotting of data relevant to photobiology (described at
  \url{https://www.r4photobiology.info/}). The accompanying packages (under
  development) provide data and definitions that are to a large extent
  application-area specific while the functions in the present package are
  widely useful in photobiology and radiation quantification in geophysics
  and meteorology. Package 'photobiology' has its main focus in the
  characterization of the light environment in a biologically relevant manner
  and in the manipulation of spectral data to simulate photo-physical,
  photo-chemical and photo-biological interactions and responses. The focus of
  package 'pavo' (Maia et al., 2003) is in colour perception by animals and
  assessment of animal coloration. In spite of the different focus, there is
  some degree of overlap.
}
\note{
Code for some of the astronomical calculations has been adapted from
  that in package 'pavo'.
}
\section{Acknowledgements}{
 This work was funded by the Academy of Finland
  (decision 252548). COST Action FA9604 'UV4Growth' facilitated discussions
  and exchanges of ideas that lead to the development of this package. The
  contributions of Andy McLeod, Lars Olof Björn, Nigel Paul, Lasse Ylianttila,
  T. Matthew Robson and Titta Kotilainen were specially significant.
  Tutorials by Hadley Wickham and comments on my presentation at UseR!2015
  allowed me to significantly improve the coding and functionality.
}

\examples{
# irradiance of the whole spectrum
irrad(sun.spct)
# photon irradiance 400 nm to 700 nm
q_irrad(sun.spct, waveband(c(400,700)))
# energy irradiance 400 nm to 700 nm
e_irrad(sun.spct, waveband(c(400,700)))
# simulating the effect of a filter on solar irradiance
e_irrad(sun.spct * yellow_gel.spct, waveband(c(400,500)))
e_irrad(sun.spct * yellow_gel.spct, waveband(c(500,700)))
# daylength
sunrise_time(lubridate::today(tzone = "EET"), tz = "EET",
             geocode = data.frame(lat = 60, lon = 25), unit.out = "hour")
day_length(lubridate::today(tzone = "EET"), tz = "EET",
           geocode = data.frame(lat = 60, lon = 25), unit.out = "hour")
# colour as seen by humans
color_of(sun.spct)
color_of(sun.spct * yellow_gel.spct)
# filter transmittance
transmittance(yellow_gel.spct)
transmittance(yellow_gel.spct, waveband(c(400,500)))
transmittance(yellow_gel.spct, waveband(c(500,700)))
}
\references{
Aphalo, P. J., Albert, A., Björn, L. O., McLeod, A. R., Robson, T. M.,
Rosenqvist, E. (Eds.). (2012). Beyond the Visible: A handbook of best
practice in plant UV photobiology (1st ed., p. xx + 174). Helsinki:
University of Helsinki, Department of Biosciences, Division of Plant Biology.
ISBN 978-952-10-8363-1 (PDF), 978-952-10-8362-4 (paperback). Open access PDF
download available at https://hdl.handle.net/10138/37558

Aphalo, Pedro J. (2015) The r4photobiology suite. UV4Plants Bulletin, 2015:1,
21-29. \doi{10.19232/uv4pb.2015.1.14}.

Maia, R., Eliason, C. M., Bitton, P. P., Doucet, S. M., Shawkey, M. D. (2013)
pavo: an R package for the analysis, visualization and organization of
spectral data. Methods in Ecology and Evolution, 4(10):906-913.
\doi{10.1111/2041-210X.12069}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.r4photobiology.info/photobiology/}
  \item \url{https://github.com/aphalo/photobiology}
  \item Report bugs at \url{https://github.com/aphalo/photobiology/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} (\href{https://orcid.org/0000-0003-3385-972X}{ORCID})

Other contributors:
\itemize{
  \item Titta K. Kotilainen (\href{https://orcid.org/0000-0002-2822-9734}{ORCID}) [contributor]
  \item Glenn Davis \email{gdavis@gluonics.com} [contributor]
  \item Agnese Fazio \email{agnese.fazio@uni-jena.de} [contributor]
}

}
