% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.data.r
\docType{data}
\name{photodiode.spct}
\alias{photodiode.spct}
\title{Spectral response of a GaAsP photodiode}
\format{A \code{response_spct} object with 311 rows and 2 variables}
\description{
A dataset containing wavelengths at a 1 nm interval and
  spectral response as $A W^{-1} nm^{-1}$ for GaAsP photodiode type
  G6262 from Hamamatsu. Data digitized from manufacturer's data sheet.
}
\details{
\itemize{ \item w.length (nm). \item s.e.response
(A/W)  }
}
\references{
Hamamatsu (2011) Datasheet: GaAsP Photodiodes G5645 G5842 G6262. Hamamatsu
Photonics KK, Hamamatsu, City.
http://www.hamamatsu.com/resources/pdf/ssd/g5645_etc_kgpd1004e.pdf.
Visited 2016-02-01.
}
\keyword{datasets}

