% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photon.irradiance.r
\name{photon_irradiance}
\alias{photon_irradiance}
\title{Photon irradiance}
\usage{
photon_irradiance(w.length, s.irrad, w.band = NULL, unit.in = "energy",
  check.spectrum = TRUE,
  use.cached.mult = getOption("photobiology.use.cached.mult", default =
  FALSE), use.hinges = getOption("photobiology.use.hinges", default = NULL))
}
\arguments{
\item{w.length}{numeric array of wavelength (nm)}

\item{s.irrad}{numeric array of spectral irradiances, by default as energy (W
m-2 nm-1)}

\item{w.band}{waveband}

\item{unit.in}{character Values recognized "photon" or "energy"}

\item{check.spectrum}{logical Flag telling whether to sanity check input data,
default is TRUE}

\item{use.cached.mult}{logical Flag telling whether multiplier values should be
cached between calls}

\item{use.hinges}{logical Flag telling whether to use hinges to reduce
interpolation errors}
}
\value{
A single numeric value with no change in scale factor: [W m-2 nm-1]
  -> [W m-2]
}
\description{
This function returns the photon irradiance for a given waveband of a
radiation spectrum, optionally applies a BSWF.
}
\examples{
with(sun.data, photon_irradiance(w.length, s.e.irrad))
with(sun.data, photon_irradiance(w.length, s.e.irrad, new_waveband(400,700)))

}
\seealso{
Other irradiance functions: \code{\link{e_fluence.source_mspct}},
  \code{\link{e_irrad.source_mspct}},
  \code{\link{energy_irradiance}},
  \code{\link{fluence.source_mspct}},
  \code{\link{irrad.source_mspct}},
  \code{\link{irradiance}},
  \code{\link{q_fluence.source_mspct}},
  \code{\link{q_irrad.source_mspct}}
}

