% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.data.R
\docType{data}
\name{polyester.spct}
\alias{polyester.spct}
\title{Transmittance spectrum of clear polyester film}
\format{A \code{filter_spct} object with 611 rows and 2 variables}
\description{
A dataset containing the wavelengths at a 1 nm interval and
  fractional total transmittance for polyester film.
}
\details{
\itemize{ \item w.length (nm). \item Tfr
(0..1)  }
}
\seealso{
Other Spectral data examples: \code{\link{A.illuminant.spct}},
  \code{\link{D65.illuminant.spct}},
  \code{\link{black_body.spct}}, \code{\link{ccd.spct}},
  \code{\link{clear.spct}}, \code{\link{clear_body.spct}},
  \code{\link{filter_cps.mspct}},
  \code{\link{green_leaf.spct}},
  \code{\link{photodiode.spct}},
  \code{\link{sun.daily.data}},
  \code{\link{sun.daily.spct}}, \code{\link{sun.data}},
  \code{\link{sun.spct}}, \code{\link{white_body.spct}},
  \code{\link{white_led.cps_spct}},
  \code{\link{white_led.raw_spct}},
  \code{\link{white_led.source_spct}},
  \code{\link{yellow_gel.spct}}
}
\keyword{datasets}
