% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.summaries.r
\name{print.summary_generic_spct}
\alias{print.summary_generic_spct}
\title{Print spectral summary}
\usage{
\method{print}{summary_generic_spct}(x, ...)
}
\arguments{
\item{x}{An object of one of the summary classes for spectra}

\item{...}{not used in current version}
}
\description{
A function to nicely print objects of classes "summary...spct".
}
\examples{
print(summary(sun.spct))

}
