% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{print.tod_time}
\alias{print.tod_time}
\title{Print time-of-day objects}
\usage{
\method{print}{tod_time}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{passed to \code{format} method}
}
\description{
Print time-of-day objects
}
\note{
Default is to print the underlying \code{numeric} vector as a solar time.
}
\seealso{
Other Time of day functions: 
\code{\link{as_tod}()},
\code{\link{format.tod_time}()}
}
\concept{Time of day functions}
