% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.irrad.R
\name{q_fluence}
\alias{q_fluence}
\alias{q_fluence.default}
\alias{q_fluence.source_spct}
\alias{q_fluence.source_mspct}
\title{Photon fluence}
\usage{
q_fluence(spct, w.band, exposure.time, wb.trim, use.cached.mult, use.hinges,
  allow.scaled, ...)

\method{q_fluence}{default}(spct, w.band, exposure.time, wb.trim,
  use.cached.mult, use.hinges, allow.scaled, ...)

\method{q_fluence}{source_spct}(spct, w.band = NULL, exposure.time,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.cached.mult = getOption("photobiology.use.cached.mult", default =
  FALSE), use.hinges = NULL, allow.scaled = FALSE, ...)

\method{q_fluence}{source_mspct}(spct, w.band = NULL, exposure.time,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.cached.mult = getOption("photobiology.use.cached.mult", default =
  FALSE), use.hinges = NULL, allow.scaled = FALSE, ...,
  idx = !is.null(names(spct)))
}
\arguments{
\item{spct}{an R object}

\item{w.band}{a list of \code{waveband} objects or a \code{waveband} object}

\item{exposure.time}{lubridate::duration}

\item{wb.trim}{logical if TRUE wavebands crossing spectral data boundaries
are trimmed, if FALSE, they are discarded}

\item{use.cached.mult}{logical indicating whether multiplier values should be
cached between calls}

\item{use.hinges}{logical indicating whether to use hinges to reduce
interpolation errors}

\item{allow.scaled}{logical indicating whether scaled or normalized spectra
as argument to spct are flagged as an error}

\item{...}{other arguments (possibly ignored)}

\item{idx}{logical whether to add a column with the names of the elements of
spct}
}
\value{
One numeric value for each waveband with no change in scale factor,
  with name attribute set to the name of each waveband unless a named list is
  supplied in which case the names of the list elements are used. The
  exposure.time is copied from the spectrum object to the output as an attibute.
  Units are as follows: moles of photons per exposure.
}
\description{
This function returns the photon irradiance (or quantum irradiance) for a
given waveband of a light source spectrum.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default for generic function

\item \code{source_spct}: Calculate photon fluence from a \code{source_spct}
object and the duration of the exposure

\item \code{source_mspct}: Calculates photon (quantum) fluence from a
\code{source_mspct} object.
}}

\note{
The last two parameters control speed optimizations. The defaults
  should be suitable in mosts cases. If you will use repeatedly the same SWFs
  on many spectra measured at exactly the same wavelengths you may obtain
  some speed up by setting \code{use.cached.mult=TRUE}. However, be aware
  that you are responsible for ensuring that the wavelengths are the same in
  each call, as the only test done is for the length of the \code{w.length}
  vector.
}
\examples{
library(lubridate)
q_fluence(sun.spct,
          w.band = waveband(c(400,700)),
          exposure.time = lubridate::duration(3, "minutes") )

}
\seealso{
Other irradiance functions: \code{\link{e_fluence}},
  \code{\link{e_irrad}}, \code{\link{energy_irradiance}},
  \code{\link{fluence}}, \code{\link{irradiance}},
  \code{\link{irrad}}, \code{\link{photon_irradiance}},
  \code{\link{q_irrad}}
}
