% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.fractions.R
\name{q_fraction}
\alias{q_fraction}
\alias{q_fraction.default}
\alias{q_fraction.source_spct}
\title{Photon:photon fraction}
\usage{
q_fraction(
  spct,
  w.band.num,
  w.band.denom,
  scale.factor,
  wb.trim,
  use.cached.mult,
  use.hinges,
  ...
)

\method{q_fraction}{default}(
  spct,
  w.band.num,
  w.band.denom,
  scale.factor,
  wb.trim,
  use.cached.mult,
  use.hinges,
  ...
)

\method{q_fraction}{source_spct}(
  spct,
  w.band.num = NULL,
  w.band.denom = NULL,
  scale.factor = 1,
  wb.trim = getOption("photobiology.waveband.trim", default = TRUE),
  use.cached.mult = FALSE,
  use.hinges = NULL,
  naming = "short",
  name.tag = ifelse(naming != "none", "[q:q]", ""),
  ...
)
}
\arguments{
\item{spct}{an object of class "source_spct".}

\item{w.band.num}{waveband object or a list of waveband objects used to
compute the numerator(s) and denominator(s) of the fraction(s).}

\item{w.band.denom}{waveband object or a list of waveband objects used to
compute the denominator(s) of the fraction(s).}

\item{scale.factor}{numeric vector of length 1, or length equal to that of
\code{w.band}. Numeric multiplier applied to returned values.}

\item{wb.trim}{logical if TRUE wavebands crossing spectral data boundaries
are trimmed, if FALSE, they are discarded}

\item{use.cached.mult}{logical indicating whether multiplier values should be
cached between calls}

\item{use.hinges}{logical Flag indicating whether to insert "hinges" into the
spectral data before integration so as to reduce interpolation errors at
the boundaries of the wavebands.}

\item{...}{other arguments (possibly ignored)}

\item{naming}{character one of "long", "default", "short" or "none". Used to
select the type of names to assign to returned value.}

\item{name.tag}{character Used to tag the name of the returned values.}
}
\value{
In the case of methods for individual spectra, a \code{numeric}
  vector of adimensional values giving a photon fraction between integrated
  photon irradiances for pairs of wavebands, with name attribute set to
  the name of the wavebands unless a named list of wavebands is supplied in
  which case the names of the list elements are used, with "(q:q)" appended.
  A \code{data.frame} in the case of collections of spectra, containing one
  column for each fraction definition, an index column with the names of the
  spectra, and optionally additional columns with metadata values retrieved
  from the attributes of the member spectra.

  Fraction definitions are "assembled" from the arguments passed to
  \code{w.band.num} and \code{w.band.denom}. If both arguments are of equal
  length, then the wavebands are paired to obtain as many fractions as the
  number of wavebands in each list. Recycling for wavebands takes place when
  the number of denominator and numerator wavebands differ.
}
\description{
This function returns the photon fraction for a given pair of wavebands of a
light source spectrum.
}
\section{Methods (by class)}{
\itemize{
\item \code{q_fraction(default)}: Default for generic function

\item \code{q_fraction(source_spct)}: Method for \code{source_spct} objects

}}
\note{
The last two parameters control speed
  optimizations. The defaults should be suitable in most cases. If you will
  use repeatedly the same SWFs on many spectra measured at exactly the same
  wavelengths you may obtain some speed up by setting
  \code{use.cached.mult=TRUE}. However, be aware that you are responsible for
  ensuring that the wavelengths are the same in each call, as the only test
  done is for the length of the \code{w.length} vector.
}
\examples{
q_fraction(sun.spct, new_waveband(400,500), new_waveband(400,700))

}
\seealso{
Other photon and energy ratio functions: 
\code{\link{e_ratio}()},
\code{\link{eq_ratio}()},
\code{\link{q_ratio}()},
\code{\link{qe_ratio}()}
}
\concept{photon and energy ratio functions}
