% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveband.class.r
\name{range}
\alias{range}
\alias{range.generic_mspct}
\alias{range.generic_spct}
\alias{range.waveband}
\title{Wavelength range}
\usage{
\method{range}{waveband}(..., na.rm = FALSE)

\method{range}{generic_spct}(..., na.rm = FALSE)

\method{range}{generic_mspct}(..., na.rm = FALSE, idx = NULL)
}
\arguments{
\item{...}{not used in current version}

\item{na.rm}{ignored}

\item{idx}{logical whether to add a column with the names of the elements of
spct}
}
\description{
A function that returns the wavelength range.
}
\section{Methods (by class)}{
\itemize{
\item \code{generic_spct}: 

\item \code{generic_mspct}: 
}}
\examples{
range(sun.spct)

}
\seealso{
Other wavelength summaries: \code{\link{midpoint}},
  \code{\link{min}}, \code{\link{stepsize}}
}

