% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.reflectance.r
\name{reflectance_spct}
\alias{reflectance_spct}
\title{Calculate reflectance from spectral reflectance}
\usage{
reflectance_spct(spct, w.band, quantity, wb.trim, use.hinges)
}
\arguments{
\item{spct}{an object of class generic_spct"}

\item{w.band}{list of waveband definitions created with new_waveband()}

\item{quantity}{character string}

\item{wb.trim}{logical if TRUE wavebands crossing spectral data boundaries
are trimmed, if FALSE, they are discarded}

\item{use.hinges}{logical indicating whether to use hinges to reduce
interpolation errors}
}
\value{
A single numeric value expressed as a fraction of one
}
\description{
This function returns the mean reflectance for a given waveband and a
reflectance spectrum.
}
\keyword{internal}

