% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.normalize.r
\name{restore_normalization}
\alias{restore_normalization}
\title{Restore normalization}
\usage{
restore_normalization(x, old.normalization.ls)
}
\arguments{
\item{x}{generic_spct A spectrum with no normalization applied.}
}
\description{
After altering the columns present in a spectrum restore normalization
with the previously used criteria.
}
\details{
The normalization criteria are extracted from
\code{old.normalization.ls} and applied. In the case of spectra in
long form, the normalization can be stored as a named list with, with
values stored for each spectrum. If consistent, they are simplified and
applied. Otherwise, they re-normalization is skipped with a warning.
}
\keyword{internal}
