% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspct.prod.r
\name{s_prod}
\alias{s_prod}
\alias{s_prod.default}
\alias{s_prod.source_mspct}
\alias{s_prod.response_mspct}
\alias{s_prod.filter_mspct}
\alias{s_prod.reflector_mspct}
\alias{s_prod.calibration_mspct}
\alias{s_prod.cps_mspct}
\alias{s_prod.raw_mspct}
\title{Product from collection of spectra}
\usage{
s_prod(x, na.rm, ...)

\method{s_prod}{default}(x, na.rm = FALSE, ...)

\method{s_prod}{source_mspct}(x, na.rm = FALSE, ...)

\method{s_prod}{response_mspct}(x, na.rm = FALSE, ...)

\method{s_prod}{filter_mspct}(x, na.rm = FALSE, ...)

\method{s_prod}{reflector_mspct}(x, na.rm = FALSE, ...)

\method{s_prod}{calibration_mspct}(x, na.rm = FALSE, ...)

\method{s_prod}{cps_mspct}(x, na.rm = FALSE, ...)

\method{s_prod}{raw_mspct}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An R object. Currently this package defines methods for collections of
spectral objects.}

\item{na.rm}{logical. A value indicating whether NA values should be stripped
before the computation proceeds.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
If \code{x} is a collection spectral of objects, such as a
  "filter_mspct" object, the returned object is of same class as the
  members of the collection, such as "filter_spct", containing the product
  of the spectra.
}
\description{
A method to compute the product of values across members of a collections of
spectra. Computes the product at each wavelength across all the spectra in the
collection returning a spectral object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{source_mspct}: 

\item \code{response_mspct}: 

\item \code{filter_mspct}: 

\item \code{reflector_mspct}: 

\item \code{calibration_mspct}: 

\item \code{cps_mspct}: 

\item \code{raw_mspct}: 
}}

\note{
Omission of NAs is done separately at each wavelength. Interpolation is
  not applied, so all spectra in \code{x} must share the same set of
  wavelengths.

  A product of spectral irradiance or spectral response is no longer a well
  defined physical quanttiy, and these product operations return an object of
  class generic_spct.

  Objects of classes raw_spct and cps_spct can contain data from multiple
  scans. This functions are implemented for these classes only for the case
  when all member spectra contain data for a single scan, or spliced into a
  single column in the case of cps_spct members.
}
\seealso{
See \code{\link[base]{prod}} for the \code{prod()} method used for
  the computations.
}
