% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setIdFactor}
\alias{setIdFactor}
\title{Set the "idfactor" attribute}
\usage{
setIdFactor(x, idfactor)
}
\arguments{
\item{x}{a generic_spct object}

\item{idfactor}{character The name of a factor identifying multiple
spectra stored longitudinally.}
}
\value{
x
}
\description{
Function to set by reference the "idfactor" attribute  of an existing
generic_spct or an object of a class derived from generic_spct.
}
\note{
This function alters x itself by reference and in addition
  returns x invisibly. If x is not a generic_spct or an object of a class derived from
  generic_spct, x is not modified.
}
\seealso{
Other idfactor attribute functions: 
\code{\link{getIdFactor}()}
}
\concept{idfactor attribute functions}
