% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setInstrDesc}
\alias{setInstrDesc}
\title{Set the "instr.desc" attribute}
\usage{
setInstrDesc(x, instr.desc)
}
\arguments{
\item{x}{a generic_spct object}

\item{instr.desc}{a list}
}
\value{
x
}
\description{
Function to set by reference the "instr.desc" attribute  of an existing
generic_spct or derived-class object.
}
\note{
if x is not a generic_spct x is not modified.
}
\seealso{
Other measurement metadata functions: \code{\link{getInstrDesc}},
  \code{\link{getInstrSettings}},
  \code{\link{getWhenMeasured.generic_mspct}},
  \code{\link{getWhereMeasured.generic_mspct}},
  \code{\link{setInstrSettings}},
  \code{\link{setWhenMeasured.generic_mspct}},
  \code{\link{setWhereMeasured.generic_mspct}}
}

