% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setInstrDesc}
\alias{setInstrDesc}
\title{Set the "instr.desc" attribute}
\usage{
setInstrDesc(x, instr.desc)
}
\arguments{
\item{x}{a generic_spct object}

\item{instr.desc}{a list}
}
\value{
x
}
\description{
Function to set by reference the "instr.desc" attribute  of an existing
generic_spct or derived-class object.
}
\note{
This function alters x itself by reference and in addition
  returns x invisibly. If x is not a generic_spct object, x is not
  modified.


}
\seealso{
Other measurement metadata functions: \code{\link{getInstrDesc}},
  \code{\link{getInstrSettings}},
  \code{\link{getWhatMeasured}},
  \code{\link{getWhenMeasured}},
  \code{\link{getWhereMeasured}},
  \code{\link{get_attributes}},
  \code{\link{isValidInstrDesc}},
  \code{\link{isValidInstrSettings}},
  \code{\link{setInstrSettings}},
  \code{\link{setWhatMeasured}},
  \code{\link{setWhenMeasured}},
  \code{\link{setWhereMeasured}},
  \code{\link{trimInstrDesc}},
  \code{\link{trimInstrSettings}}
}
\concept{measurement metadata functions}
