% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setInstrSettings}
\alias{setInstrSettings}
\title{Set the "instr.settings" attribute}
\usage{
setInstrSettings(x, instr.settings)
}
\arguments{
\item{x}{a generic_spct object}

\item{instr.settings}{a list}
}
\value{
x
}
\description{
Function to set by reference the "what.measured" attribute  of an existing
generic_spct or derived-class object.
}
\note{
This function alters x itself by reference and in addition
  returns x invisibly. If x is not a generic_spct object, x is not
  modified.
}
\seealso{
Other measurement metadata functions: \code{\link{getHowMeasured}},
  \code{\link{getInstrDesc}},
  \code{\link{getInstrSettings}},
  \code{\link{getWhatMeasured}},
  \code{\link{getWhenMeasured}},
  \code{\link{getWhereMeasured}},
  \code{\link{get_attributes}},
  \code{\link{isValidInstrDesc}},
  \code{\link{isValidInstrSettings}},
  \code{\link{setHowMeasured}}, \code{\link{setInstrDesc}},
  \code{\link{setWhatMeasured}},
  \code{\link{setWhenMeasured}},
  \code{\link{setWhereMeasured}},
  \code{\link{trimInstrDesc}},
  \code{\link{trimInstrSettings}}
}
\concept{measurement metadata functions}
