% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setInstrSettings}
\alias{setInstrSettings}
\title{Set the "instr.settings" attribute}
\usage{
setInstrSettings(x, instr.settings)
}
\arguments{
\item{x}{a generic_spct object}

\item{instr.settings}{a list}
}
\value{
x
}
\description{
Function to set by reference the "what.measured" attribute  of an existing
generic_spct or derived-class object.
}
\note{
if x is not a generic_spct x is not modified.
}
\seealso{
Other measurement metadata functions: \code{\link{getInstrDesc}},
  \code{\link{getInstrSettings}},
  \code{\link{getWhatMeasured}},
  \code{\link{getWhenMeasured}},
  \code{\link{getWhereMeasured}},
  \code{\link{setInstrDesc}},
  \code{\link{setWhatMeasured}},
  \code{\link{setWhenMeasured}},
  \code{\link{setWhereMeasured}}
}

