% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setMultipleWl}
\alias{setMultipleWl}
\title{Set the "multiple.wl" attribute}
\usage{
setMultipleWl(x, multiple.wl = NULL)
}
\arguments{
\item{x}{a generic_spct object}

\item{multiple.wl}{numeric >= 1 If \code{multiple.wl} is \code{NULL}, the
default, the attribute is not modified if it is already present and valid,
and set to 1 otherwise.}
}
\value{
x
}
\description{
Function to set by reference the "multiple.wl" attribute  of an existing
generic_spct or an object of a class derived from generic_spct.
}
\note{
This function alters x itself by reference and in addition
  returns x invisibly. If x is not a generic_spct or an object of a class derived from
  generic_spct, x is not modified. If \code{multiple.wl}
}
\seealso{
Other multiple.wl attribute functions: 
\code{\link{getMultipleWl}()}
}
\concept{multiple.wl attribute functions}
