% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.normalize.r
\name{setNormalized}
\alias{setNormalized}
\title{Set the "normalized" attribute}
\usage{
setNormalized(x, norm = FALSE)
}
\arguments{
\item{x}{a generic_spct object}

\item{norm}{numeric or logical}
}
\description{
Funtion to write the "normalized" attribute of an existing generic_spct
object.
}
\note{
if x is not a \code{generic_spct} object, x is not modified.
}
\seealso{
Other rescaling functions: \code{\link{fscale}},
  \code{\link{fshift}}, \code{\link{getNormalized}},
  \code{\link{is_normalized}}, \code{\link{is_scaled}},
  \code{\link{normalize}}, \code{\link{setScaled}}
}
