% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setRfrType}
\alias{setRfrType}
\title{Set the "Rfr.type" attribute}
\usage{
setRfrType(x, Rfr.type = c("total", "specular"))
}
\arguments{
\item{x}{a reflector_spct or an object_spct object}

\item{Rfr.type}{a character string, either "total" or "specular"}
}
\value{
x
}
\description{
Function to set by reference the "Rfr.type" attribute  of an existing
reflector_spct or object_spct object.
}
\note{
if x is not a reflector_spct or object_spct object, x is not modified.
  The behaviour of this function is 'unusual' in that the default for
  parameter Rfr.type is used only if \code{x} does not already have this
  attribute set.
}

