% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.fscale.r
\name{setScaled}
\alias{setScaled}
\title{Set the "scaled" attribute}
\usage{
setScaled(x, scaled = FALSE)
}
\arguments{
\item{x}{a generic_spct object}

\item{scaled}{logical}
}
\description{
Function to write the "scaled" attribute of an existing generic_spct
object.
}
\note{
if x is not a \code{generic_spct} object, x is not modified.
  attribute set.
}
\seealso{
Other rescaling functions: \code{\link{fscale}},
  \code{\link{fshift}}, \code{\link{getNormalized}},
  \code{\link{is_normalized}}, \code{\link{is_scaled}},
  \code{\link{normalize}}, \code{\link{setNormalized}}
}
