% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setSensorProperties}
\alias{setSensorProperties}
\alias{sensor_properties<-}
\title{Set the "sensor.properties" attribute}
\usage{
setSensorProperties(x, sensor.properties = NULL, verbose = FALSE)

sensor_properties(x) <- value
}
\arguments{
\item{x}{a response_spct object}

\item{sensor.properties, value}{a list with fields named as described in
\strong{Details}.}

\item{verbose}{logical Flag to enable warning when applied to object of
unsupported class.}
}
\value{
\code{x}
}
\description{
Function to set by reference the "sensor.properties" attribute  of an
existing response_spct object.
}
\details{
Storing sensor properties in an attribute with a well defined
format and minimum of information is intended to help with reproducibility.
Unlike the \code{filter.properties} metadata, the \code{sensor.properties}
metadata are currently not used in any computations. The attribute is
intended to be used with photoelectric sensors, broadband and image, i.e.,
not for biological photoreceptors or photobiological responses.

\describe{
\item{model}{\code{character}, \emph{identification code or name used by the supplier}, required}
\item{type}{\code{character}, \code{"broadband"}, \code{"spectral"}, \code{"image"} required}
\item{supplier}{\code{character}, required}
\item{entrance.optics}{\code{character}, \code{"cosine"}, \code{"dome"}, \code{"sphere"}, \code{"narrow"} required}
\item{channels}{\code{data.frame}, description of the channels, one channel per row, optional}
\item{signal.interface}{\code{character}, \code{"analog"}, \code{"digital"}, and possibly others.}
\item{module.name}{\code{character}, optional}
\item{module.supplier}{\code{character}, optional}
\item{module.interface}{\code{character}, optional}
\item{note}{\code{character}, optional}
}
}
\note{
This function alters \code{x} itself by reference and in addition
  returns \code{x} invisibly. If \code{x} is not a \code{source_spct} object,
  \code{x} is not modified.
}
\examples{

my.spct <- ccd.spct
sensor_properties(my.spct)
sensor_properties(my.spct) <- NULL
sensor_properties(my.spct)
sensor_properties(my.spct) <- list(model = "ccd",
                                   type = "ABC",
                                   supplier = "unknown")
sensor_properties(my.spct)

}
\seealso{
Other measurement metadata functions: 
\code{\link{add_attr2tb}()},
\code{\link{getFilterProperties}()},
\code{\link{getHowMeasured}()},
\code{\link{getInstrDesc}()},
\code{\link{getInstrSettings}()},
\code{\link{getSensorProperties}()},
\code{\link{getSoluteProperties}()},
\code{\link{getWhatMeasured}()},
\code{\link{getWhenMeasured}()},
\code{\link{getWhereMeasured}()},
\code{\link{get_attributes}()},
\code{\link{isValidInstrDesc}()},
\code{\link{isValidInstrSettings}()},
\code{\link{select_spct_attributes}()},
\code{\link{setFilterProperties}()},
\code{\link{setHowMeasured}()},
\code{\link{setInstrDesc}()},
\code{\link{setInstrSettings}()},
\code{\link{setSoluteProperties}()},
\code{\link{setWhatMeasured}()},
\code{\link{setWhenMeasured}()},
\code{\link{setWhereMeasured}()},
\code{\link{spct_attr2tb}()},
\code{\link{spct_metadata}()},
\code{\link{subset_attributes}()},
\code{\link{trimInstrDesc}()},
\code{\link{trimInstrSettings}()}
}
\concept{measurement metadata functions}
