% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setWhatMeasured}
\alias{setWhatMeasured}
\title{Set the "what.measured" attribute}
\usage{
setWhatMeasured(x, what.measured)
}
\arguments{
\item{x}{a generic_spct object}

\item{what.measured}{a list}
}
\value{
x
}
\description{
Function to set by reference the "what.measured" attribute  of an existing
generic_spct or derived-class object.
}
\note{
if x is not a generic_spct x is not modified.
}
\seealso{
Other measurement metadata functions: \code{\link{getInstrDesc}},
  \code{\link{getInstrSettings}},
  \code{\link{getWhatMeasured}},
  \code{\link{getWhenMeasured}},
  \code{\link{getWhereMeasured}},
  \code{\link{setInstrDesc}},
  \code{\link{setInstrSettings}},
  \code{\link{setWhenMeasured}},
  \code{\link{setWhereMeasured}}
}

