% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.classes.r
\name{setWhereMeasured}
\alias{setWhereMeasured}
\alias{setWhereMeasured.default}
\alias{setWhereMeasured.generic_spct}
\alias{setWhereMeasured.summary_generic_spct}
\alias{setWhereMeasured.generic_mspct}
\title{Set the "where.measured" attribute}
\usage{
setWhereMeasured(x, where.measured, lat, lon, ...)

\method{setWhereMeasured}{default}(x, where.measured, lat, lon, ...)

\method{setWhereMeasured}{generic_spct}(x, where.measured = NA, lat = NA,
  lon = NA, ...)

\method{setWhereMeasured}{summary_generic_spct}(x, where.measured = NA,
  lat = NA, lon = NA, ...)

\method{setWhereMeasured}{generic_mspct}(x, where.measured = NA, lat = NA,
  lon = NA, ...)
}
\arguments{
\item{x}{a generic_spct object}

\item{where.measured}{A one row data.frame such as returned by
\code{\link[ggmap]{geocode}} for a location search.}

\item{lat}{numeric Latitude in decimal degrees North}

\item{lon}{numeric Longitude in decimal degrees West}

\item{...}{Allows use of additional arguments in methods for other classes.}
}
\value{
x
}
\description{
Function to set by reference the "where.measured" attribute  of an existing
generic_spct or an object of a class derived from generic_spct.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default

\item \code{generic_spct}: generic_spct

\item \code{summary_generic_spct}: summary_generic_spct

\item \code{generic_mspct}: generic_mspct
}}

\note{
This method alters x itself by reference and in addition
  returns x invisibly. If x is not a generic_spct or an object of a class derived from
  generic_spct, x is not modified. If \code{where} is not a POSIXct object
  or \code{NULL} an error is triggered. A \code{POSIXct} describes an
  instant in time (date plus time-of-day plus time zone).

Method for collections of spectra recycles the location information
  only if it is of length one.
}
\seealso{
Other measurement metadata functions: \code{\link{getInstrDesc}},
  \code{\link{getInstrSettings}},
  \code{\link{getWhatMeasured}},
  \code{\link{getWhenMeasured}},
  \code{\link{getWhereMeasured}},
  \code{\link{isValidInstrDesc}},
  \code{\link{isValidInstrSettings}},
  \code{\link{setInstrDesc}},
  \code{\link{setInstrSettings}},
  \code{\link{setWhatMeasured}},
  \code{\link{setWhenMeasured}},
  \code{\link{trimInstrDesc}},
  \code{\link{trimInstrSettings}}
}
