% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmspct.classes.R
\name{shared_member_class}
\alias{shared_member_class}
\title{Classes common to all collection members.}
\usage{
shared_member_class(l, target.set = spct_classes())
}
\arguments{
\item{l}{a list or a generic_mspct object or of a derived class.}

\item{target.set}{character The target set of classes within which to search
for classes common to all members.}
}
\value{
A character vector containing the class attribute values.
}
\description{
Finds the set intersection among the class attributes of all collection
member as a target set of class names.
}
\seealso{
Other set and unset 'multi spectral' class functions: 
\code{\link{rmDerivedMspct}()}
}
\concept{set and unset 'multi spectral' class functions}
