% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.operators.r
\name{sign}
\alias{sign}
\alias{sign.generic_spct}
\title{Sign}
\usage{
\method{sign}{generic_spct}(x)
}
\arguments{
\item{x}{an object of class "generic_spct"}
}
\description{
\code{sign} returns a vector with the signs of the corresponding elements of
x (the sign of a real number is 1, 0, or -1 if the number is positive, zero,
or negative, respectively).
}
\seealso{
Other math operators and functions: 
\code{\link{MathFun}},
\code{\link{^.generic_spct}()},
\code{\link{convolve_each}()},
\code{\link{div-.generic_spct}},
\code{\link{log}()},
\code{\link{minus-.generic_spct}},
\code{\link{mod-.generic_spct}},
\code{\link{plus-.generic_spct}},
\code{\link{round}()},
\code{\link{slash-.generic_spct}},
\code{\link{times-.generic_spct}}
}
\concept{math operators and functions}
