% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spct.spikes.R
\name{spikes}
\alias{spikes}
\alias{spikes.default}
\alias{spikes.numeric}
\alias{spikes.data.frame}
\alias{spikes.generic_spct}
\alias{spikes.source_spct}
\alias{spikes.response_spct}
\alias{spikes.filter_spct}
\alias{spikes.reflector_spct}
\alias{spikes.solute_spct}
\alias{spikes.cps_spct}
\alias{spikes.raw_spct}
\alias{spikes.generic_mspct}
\alias{spikes.source_mspct}
\alias{spikes.response_mspct}
\alias{spikes.filter_mspct}
\alias{spikes.reflector_mspct}
\alias{spikes.solute_mspct}
\alias{spikes.cps_mspct}
\alias{spikes.raw_mspct}
\title{Spikes}
\usage{
spikes(x, height.threshold, z.threshold, k, spike.direction, na.rm, ...)

\method{spikes}{default}(
  x,
  height.threshold = NA,
  z.threshold = NA,
  k = NA,
  spike.direction = NA,
  na.rm = FALSE,
  ...
)

\method{spikes}{numeric}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  ...
)

\method{spikes}{data.frame}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  ...,
  y.var.name = NULL,
  var.name = y.var.name
)

\method{spikes}{generic_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  var.name = NULL,
  ...
)

\method{spikes}{source_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  ...
)

\method{spikes}{response_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  ...
)

\method{spikes}{filter_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  filter.qty = getOption("photobiology.filter.qty", default = "transmittance"),
  ...
)

\method{spikes}{reflector_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  ...
)

\method{spikes}{solute_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  ...
)

\method{spikes}{cps_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  var.name = "cps",
  ...
)

\method{spikes}{raw_spct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  var.name = "counts",
  ...
)

\method{spikes}{generic_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  ...,
  var.name = NULL,
  .parallel = FALSE,
  .paropts = NULL
)

\method{spikes}{source_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  ...,
  .parallel = FALSE,
  .paropts = NULL
)

\method{spikes}{response_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  ...,
  .parallel = FALSE,
  .paropts = NULL
)

\method{spikes}{filter_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  filter.qty = getOption("photobiology.filter.qty", default = "transmittance"),
  ...,
  .parallel = FALSE,
  .paropts = NULL
)

\method{spikes}{reflector_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  ...,
  .parallel = FALSE,
  .paropts = NULL
)

\method{spikes}{solute_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  ...,
  .parallel = FALSE,
  .paropts = NULL
)

\method{spikes}{cps_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  ...,
  var.name = "cps",
  .parallel = FALSE,
  .paropts = NULL
)

\method{spikes}{raw_mspct}(
  x,
  height.threshold = 10,
  z.threshold = 5,
  k = 20,
  spike.direction = "both",
  na.rm = FALSE,
  ...,
  var.name = "counts",
  .parallel = FALSE,
  .paropts = NULL
)
}
\arguments{
\item{x}{numeric vector containing the data.}

\item{height.threshold}{numeric The minimum height of spikes expressed
relative to the median amplitude of the baseline local variation of
\code{x}.}

\item{z.threshold}{numeric Modified local \eqn{Z} values larger than
\code{z.threshold} are detected as boundaries of spikes.}

\item{k}{integer width of median window used for smoothing; must be odd}

\item{spike.direction}{character Controls the direction of spikes to be
detected. Accepted arguments are \code{"up"}, \code{"down"},
\code{"both"}.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped
before searching for spikes.}

\item{...}{ignored}

\item{var.name, y.var.name}{character Name of column where to look
for spikes.}

\item{unit.out}{character One of "energy" or "photon"}

\item{filter.qty}{character One of "transmittance" or "absorbance"}

\item{.parallel}{if TRUE, apply function in parallel, using parallel backend
provided by foreach}

\item{.paropts}{a list of additional options passed into the foreach function
when parallel computation is enabled. This is important if (for example)
your code relies on external data or packages: use the .export and
.packages arguments to supply them so that all cluster nodes have the
correct environment set up for computing.}
}
\value{
A subset of the object passed as argument to \code{x} with rows
  corresponding to spikes.
}
\description{
Function that returns a subset of an R object with observations corresponding
to spikes. Spikes are values in spectra that are unusually high compared to
neighbors. They are usually individual values or very short runs of similar
"unusual" values.
}
\section{Spike detection}{

  Spikes are detected based on a modified \eqn{Z} score calculated
  from the differenced spectrum. The \eqn{Z} threshold used should be
  adjusted to the characteristics of the input and desired sensitivity. The
  lower the threshold the more stringent the test becomes, with shorter
  spikes being detected.

  \strong{The algorithms assume a consistent step size for the underlying
  independent variable, e.g., wavelength or time, and should not be applied
  if the data do not fulfil this assumption, at least approximately. As
  \code{find_spkikes()} operates on a single vector, checking this remains
  the responsibility of calling functions or methods such as
  \code{\link{spikes}()} and \code{\link{despike}()}.}

  The algorithm uses running differences to detect abrupt changes in value,
  compared to an estimate of the baseline variation of the differences,
  approximating a baseline \eqn{Z} from MAD and a baseline value from the
  median differences. Currently, a single estimate of MAD is used but running
  medians, when possible, as baseline. This comparison detects running
  differences that are unusually large, in most cases signalling a transition
  between values near the baseline and far from it, in both directions.

  Transitions into- and out of spikes are distinguished based on the median
  of the non-differenced values, as a descriptor of the data baseline. As for
  the median of the differences, a running median is used when possible.

  This function thus detects the start and end of each spike, and
  distinguishes upward and downward spikes.

  \code{k} is the width in number of observations of the window used for
  running median smoothing to extract the baseline. A value several times the
  width of the broader spike but narrow enough to track broader peaks needs
  to be manually set in most cases.

  With \code{na.rm = TRUE}, \code{NA} values are omitted before searching for
  spikes and set to \code{0L} in the returned vector.

  If all spikes are guaranteed to be one observation-wide and either going up
  or down from the baseline, it is possible to detect them based purely on
  the \code{z.threshold} by passing \code{height.threshold = NA} and either
  \code{spike.direction = "up"} or \code{spike.direction = "down"}, which
  ensures very fast computation.

  Parameters of the algorithm need to be adjusted depending on the data, so
  inspection of returned values is needed together with adjustment by trial
  and error of suitable values for \code{z.threshold},
  \code{height.threshold}, and \code{k}.
}

\examples{
spikes(sun.spct)

}
\seealso{
See \code{\link{find_spikes}()} for locating spikes in a vector,
  \code{\link{despike}()} for replacement of spikes by interpolation in
  spectra and and \code{\link{replace_bad_pixs}()} for replacing by
  interpolation missing or bad values in a vector.

Other peaks and valleys functions:
\code{\link[=find_peaks]{find_peaks()}},
\code{\link[=find_spikes]{find_spikes()}},
\code{\link[=get_peaks]{get_peaks()}},
\code{\link[=peaks]{peaks()}},
\code{\link[=replace_bad_pixs]{replace_bad_pixs()}},
\code{\link[=valleys]{valleys()}},
\code{\link[=wls_at_target]{wls_at_target()}}
}
\concept{peaks and valleys functions}
