% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.data.r
\docType{data}
\name{sun.spct}
\alias{sun.spct}
\title{Solar spectral irradiance (simulated)}
\format{A \code{source_spct} object with 508 rows and 3 variables}
\description{
A dataset containing the wavelengths at a 1 nm interval and the
  corresponding spectral (energy) irradiance and spectral photon irradiance.
  Values simulated for 22 June 2010, near midday, at Helsinki, under partly
  cloudy conditions. The variables are as follows:
}
\details{
\itemize{ \item w.length (nm), range 293 to 800 nm. \item s.e.irrad
(W m-2 nm-1) \item s.q.irrad (mol m-2 nm-1) }
}
\author{
Anders K. Lindfors (data)
}
\references{
Lindfors, A.; Heikkilä, A.; Kaurola, J.; Koskela, T. & Lakkala,
K. (2009) Reconstruction of Solar Spectral Surface UV Irradiances Using
Radiative Transfer Simulations. Photochemistry and Photobiology, 85:
1233–1239
}
\keyword{datasets}

