% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.calc.r
\name{tz_time_diff}
\alias{tz_time_diff}
\title{Time difference between two time zones}
\usage{
tz_time_diff(
  when = lubridate::now(),
  tz.target = lubridate::tz(when),
  tz.reference = "UTC"
)
}
\arguments{
\item{when}{datetime A time instant}

\item{tz.target, tz.reference}{character Two time zones using names
recognized by functions from package 'lubridate'}
}
\value{
A \code{numeric} value.
}
\description{
Returns the time difference in hours between two time zones at a given
instant in time.
}
