% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/led.data.R
\docType{data}
\name{white_led.source_spct}
\alias{white_led.source_spct}
\title{White led bulb spectrum}
\format{
A \code{source_spct} object with 1421 rows and 2 variables
}
\usage{
white_led.source_spct
}
\description{
A dataset containing wavelengths and the
  corresponding spectral irradiance data for an Osram warm white led lamp:
}
\details{
\itemize{ \item w.length (nm), range 250 to 900 nm. \item s.e.irrad
(W m-2 nm-1)}
}
\examples{
white_led.source_spct

}
\seealso{
Other Spectral data examples: 
\code{\link{A.illuminant.spct}},
\code{\link{D65.illuminant.spct}},
\code{\link{Ler_leaf.spct}},
\code{\link{Ler_leaf_rflt.spct}},
\code{\link{Ler_leaf_trns.spct}},
\code{\link{Ler_leaf_trns_i.spct}},
\code{\link{black_body.spct}},
\code{\link{ccd.spct}},
\code{\link{clear.spct}},
\code{\link{clear_body.spct}},
\code{\link{filter_cps.mspct}},
\code{\link{green_leaf.spct}},
\code{\link{opaque.spct}},
\code{\link{phenylalanine.spct}},
\code{\link{photodiode.spct}},
\code{\link{polyester.spct}},
\code{\link{sun.daily.data}},
\code{\link{sun.daily.spct}},
\code{\link{sun.data}},
\code{\link{sun.spct}},
\code{\link{water.spct}},
\code{\link{white_body.spct}},
\code{\link{white_led.cps_spct}},
\code{\link{white_led.raw_spct}},
\code{\link{yellow_gel.spct}}
}
\concept{Spectral data examples}
\keyword{datasets}
