% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind-wlbind.R
\name{wlbind}
\alias{wlbind}
\title{Row bind spectra by wavelength}
\usage{
wlbind(x, y, ids = c(x = "x", y = "y"), strict.wls = FALSE)
}
\value{
An object of the same class as \code{x} and \code{y} with data for
columns shared by \code{x} and \code{y} based on names.
}
\description{
Row bind spectra maintaining wavelength values sorted by interspersing the
rows as needed.
}
\details{
Two objects belonging to the same class, and containing each data for a
single spectrum and no common wavelengths are row bound keeping the combined
wavelengths sorted.
}
\examples{
wlbind(sun.spct[-(1:3), ], sun.spct[1:3, ])
wlbindspct(peaks(white_led.source_spct, span = NULL),
                 wls_at_target(white_led.source_spct),
                 ids = c(x = "peak", y = "fwhm"))

}
\seealso{
Other Methods for row-binding spectra:
\code{\link[=rbindspct]{rbindspct()}}
}
\concept{Methods for row-binding spectra}
